Monte Carlo Confidence Intervals

Calculates Monte Carlo confidence intervals for free and defined parameters

Usage

MC(object, R = 20000L, alpha = c(0.001, 0.01, 0.05))

Arguments

object

object of class lavaan.

R

Positive integer. Number of Monte Carlo replications.

alpha

Numeric vector. Significance level. Default value is alpha = c(0.001, 0.01, 0.05).

Value

Returns an object of class semmcci which is a list with the following elements:

R

Number of Monte Carlo replications.

alpha

Significance level specified.

lavaan

lavaan object.

mvn

Method used to generate multivariate normal random variates.

thetahat

Parameter estimates.

thetahatstar

Sampling distribution of parameter estimates.

ci

Confidence intervals.

The list element ci is a matrix with the following columns:

est

Parameter estimates.

se

Standard errors or the square root of the diagonals of the Monte Carlo sampling distribution of parameter estimates.

R

Number of valid Monte Carlo replications.

...

Percentiles that correspond to the confidence intervals defined by alpha.

Note that the rows in ci correspond to the model parameters.

Details

A sampling distribution of parameter estimates is generated from the multivariate normal distribution using the parameter estimates and the sampling variance-covariance matrix. Confidence intervals for free and defined parameters are generated using the simulated sampling distribution. Parameters can be defined using the := operator in the lavaan model syntax.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Generate Data ------------------------------------------------------------
n <- 1000
x <- rnorm(n = n)
m <- 0.50 * x + rnorm(n = n)
y <- 0.25 * x + 0.50 * m + rnorm(n = n)
data <- data.frame(x, m, y)

# Fit Model in lavaan ------------------------------------------------------
model <- "
  y ~ cp * x + b * m
  m ~ a * x
  ab := a * b
"
fit <- sem(data = data, model = model)

# Monte Carlo --------------------------------------------------------------
MC(
  fit,
  R = 100L, # use a large value e.g., 20000L for actual research
  alpha = c(0.001, 0.01, 0.05)
)
#> Monte Carlo Confidence Intervals
#>         est     se   R  0.05%   0.5%   2.5%  97.5%  99.5% 99.95%
#> cp   0.2235 0.0338 100 0.1548 0.1566 0.1609 0.2819 0.3069 0.3167
#> b    0.5380 0.0317 100 0.4641 0.4653 0.4845 0.6010 0.6202 0.6234
#> a    0.5024 0.0337 100 0.4046 0.4175 0.4492 0.5595 0.5671 0.5703
#> y~~y 1.0582 0.0437 100 0.9435 0.9528 0.9694 1.1544 1.1714 1.1809
#> m~~m 1.0338 0.0492 100 0.8990 0.9147 0.9377 1.1192 1.1303 1.1346
#> ab   0.2703 0.0226 100 0.2136 0.2151 0.2278 0.3126 0.3176 0.3182