Standardized Monte Carlo Confidence Intervals

Calculates standardized Monte Carlo confidence intervals for free and defined parameters.

Usage

MCStd(object, alpha = c(0.001, 0.01, 0.05))

Arguments

object

object of class semmcci. Output of the MC() function.

alpha

Numeric vector. Significance level. Default value is alpha = c(0.001, 0.01, 0.05).

Value

Returns an object of class semmcci_std

which is a list with the following elements:

R

Number of Monte Carlo replications.

alpha

Significance level specified.

lavaan

lavaan object.

mvn

Method used to generate multivariate normal random variates.

thetahat

Parameter estimates.

thetahatstar

Sampling distribution of parameter estimates.

ci

Confidence intervals.

thetahat_std

Standardized parameter estimates.

thetahatstar_std

Standardized sampling distribution of parameter estimates.

ci_std

Standardized confidence intervals.

The list element ci_std is a matrix with the following columns:

est

Standardized parameter estimates.

se

Standard errors or the square root of the diagonals of the standardized Monte Carlo sampling distribution of parameter estimates.

R

Number of valid Monte Carlo replications.

...

Percentiles that correspond to the confidence intervals defined by alpha.

Note that the rows in ci_std correspond to the standardized model parameters.

Details

The empirical sampling distribution of parameter estimates from the argument object is standardized, that is, each randomly generated vector of parameters is standardized. Defined parameters are computed from the standardized component parameters. Confidence intervals are generated using the standardized empirical sampling distribution.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Generate Data ------------------------------------------------------------
n <- 1000
x <- rnorm(n = n)
m <- 0.50 * x + rnorm(n = n)
y <- 0.25 * x + 0.50 * m + rnorm(n = n)
data <- data.frame(x, m, y)

# Fit Model in lavaan ------------------------------------------------------
model <- "
  y ~ cp * x + b * m
  m ~ a * x
  ab := a * b
"
fit <- sem(data = data, model = model, fixed.x = FALSE)

# Monte Carlo --------------------------------------------------------------
output <- MC(
  fit,
  R = 100L, # use a large value e.g., 20000L for actual research
  alpha = c(0.001, 0.01, 0.05)
)

# Standardized Monte Carlo -------------------------------------------------
MCStd(output)
#> Standardized Monte Carlo Confidence Intervals
#>         est     se   R  0.05%   0.5%   2.5%  97.5%  99.5% 99.95%
#> cp   0.1991 0.0231 100 0.1178 0.1257 0.1493 0.2424 0.2536 0.2625
#> b    0.4810 0.0236 100 0.4202 0.4210 0.4320 0.5309 0.5362 0.5380
#> a    0.4320 0.0227 100 0.3710 0.3735 0.3805 0.4672 0.4721 0.4733
#> y~~y 0.6462 0.0228 100 0.5907 0.5907 0.5932 0.6903 0.7004 0.7033
#> m~~m 0.8134 0.0194 100 0.7759 0.7772 0.7817 0.8552 0.8605 0.8623
#> x~~x 1.0000 0.0000 100 1.0000 1.0000 1.0000 1.0000 1.0000 1.0000
#> ab   0.2078 0.0152 100 0.1714 0.1735 0.1773 0.2322 0.2364 0.2369