Monte Carlo Confidence Intervals

Calculates Monte Carlo confidence intervals.

mc(
  object,
  R = 20000L,
  alpha = c(0.001, 0.01, 0.05),
  par = FALSE,
  ncores = NULL
)

Arguments

object
object of class lavaan.

R
Integer. Number of Monte Carlo replications.

alpha
Numeric vector. Significance level \(\alpha\). Default value is alpha = c(0.001, 0.01, 0.05).

par
Logical. If par = TRUE, use multiple cores.

ncores
Integer. Number of cores to use if par = TRUE.

Details

An empirical sampling distribution of parameter estimates is generated using the fitted model given by

$$ \hat{\boldsymbol{\theta}}^{\ast} \sim \mathcal{N}_{p} \left( \boldsymbol{\mu} = \hat{\boldsymbol{\theta}}, \boldsymbol{\Sigma} = \mathrm{Cov} \left( \hat{\boldsymbol{\theta}} \right) \right) . $$

Confidence intervals are generated using the empirical sampling distribution.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

#> This is lavaan 0.6-8
#> lavaan is FREE software! Please report any bugs.

# Generate Data ------------------------------------------------------------
n <- 1000
x <- rnorm(n = n)
m <- 0.50 * x + rnorm(n = n)
y <- 0.25 * x + 0.50 * m + rnorm(n = n)
data <- data.frame(x, m, y)

# Fit Model in lavaan ------------------------------------------------------
model <- "
  y ~ cp * x + b * m
  m ~ a * x
  ab := a * b
"
fit <- sem(
  data = data,
  model = model
)

# Monte Carlo --------------------------------------------------------------
mc <- mc(
  fit,
  R = 100L, # use a large value e.g., 20000L for actual research
  alpha = c(0.001, 0.01, 0.05)
)
print(mc)

#> Monte Carlo Confidence Intervals
#>         est     se   R ci_0.05 ci_0.5 ci_2.5 ci_97.5 ci_99.5 ci_99.95
#> cp   0.3190 0.0346 100  0.2462 0.2504 0.2579  0.3828  0.4018   0.4129
#> b    0.4822 0.0291 100  0.4180 0.4208 0.4263  0.5432  0.5522   0.5556
#> a    0.5421 0.0314 100  0.4333 0.4407 0.4768  0.5896  0.6064   0.6096
#> y~~y 0.9948 0.0444 100  0.8946 0.8948 0.8977  1.0746  1.1005   1.1077
#> m~~m 0.9331 0.0373 100  0.8327 0.8373 0.8584  1.0046  1.0211   1.0264
#> x~~x 0.9990     NA  NA      NA     NA     NA      NA      NA       NA
#> ab   0.2614 0.0223 100  0.2093 0.2100 0.2143  0.3009  0.3086   0.3089