Standardized Monte Carlo Confidence Intervals

Calculates standardized Monte Carlo confidence intervals.

std(
  object,
  type = "std.all",
  alpha = c(0.001, 0.01, 0.05),
  par = FALSE,
  ncores = NULL
)

Arguments

object
object of class semmcci. The output of semmcci::mc.

type
If "std.lv", the standardized estimates are on the variances of the (continuous) latent variables only. If "std.all", the standardized estimates are based on both the variances of both (continuous) observed and latent variables. If "std.nox", the standardized estimates are based on both the variances of both (continuous) observed and latent variables, but not the variances of exogenous covariates.

alpha
Numeric vector. Significance level \(\alpha\). Default value is alpha = c(0.001, 0.01, 0.05).

par
Logical. If par = TRUE, use multiple cores.

ncores
Integer. Number of cores to use if par = TRUE. If unspecified, uses the output of parallel::detectCores().

Details

The empirical sampling distribution of parameter estimates from the argument object are standardized. Confidence intervals are generated using the standardized empirical sampling distribution.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Generate Data ------------------------------------------------------------
n <- 1000
x <- rnorm(n = n)
m <- 0.50 * x + rnorm(n = n)
y <- 0.25 * x + 0.50 * m + rnorm(n = n)
data <- data.frame(x, m, y)

# Fit Model in lavaan ------------------------------------------------------
model <- "
  y ~ cp * x + b * m
  m ~ a * x
  ab := a * b
"
fit <- sem(
  data = data,
  model = model,
  fixed.x = FALSE
)

# Monte Carlo --------------------------------------------------------------
mc <- mc(
  fit,
  R = 100L, # use a large value e.g., 20000L for actual research
  alpha = c(0.001, 0.01, 0.05)
)

# Standardized Monte Carlo -------------------------------------------------
std <- std(mc)
print(std)

#> Standardized Monte Carlo Confidence Intervals
#>         est     se   R ci_0.05 ci_0.5 ci_2.5 ci_97.5 ci_99.5 ci_99.95
#> cp   0.2312 0.0250 100  0.1654 0.1658 0.1784  0.2719  0.2773   0.2779
#> b    0.4306 0.0284 100  0.3735 0.3778 0.3869  0.4899  0.5014   0.5074
#> a    0.4460 0.0281 100  0.3791 0.3828 0.3886  0.4900  0.4956   0.4966
#> y~~y 0.6724 0.0251 100  0.6176 0.6214 0.6265  0.7184  0.7443   0.7540
#> m~~m 0.8010 0.0370 100  0.7102 0.7141 0.7230  0.8687  0.8788   0.8824
#> x~~x 1.0000 0.0455 100  0.8766 0.8833 0.9142  1.1023  1.1140   1.1171
#> ab   0.1920 0.0178 100  0.1538 0.1549 0.1577  0.2266  0.2337   0.2391