Standardized Monte Carlo Confidence Intervals

Calculates standardized Monte Carlo confidence intervals for free and defined parameters.

Usage

MCStd(mc, alpha = c(0.001, 0.01, 0.05))

Arguments

mc

Output of the MC() or MCMI() function.

alpha

Numeric vector. Significance level \(\alpha\).

Value

Returns an object of class semmcci which is a list with the following elements:

call

Function call.

args

List of function arguments.

thetahat

Parameter estimates \(\hat{\theta}\).

thetahatstar

Sampling distribution of parameter estimates \(\hat{\theta}^{\ast}\).

fun

Function used ("MCStd").

Details

The empirical sampling distribution of parameter estimates from the argument mc is standardized, that is, each randomly generated vector of parameters is standardized. Defined parameters are computed from the standardized component parameters. Confidence intervals are generated using the standardized empirical sampling distribution.

References

Pesigan, I. J. A., & Cheung, S. F. (2023). Monte Carlo confidence intervals for the indirect effect with missing data. Behavior Research Methods. doi:10.3758/s13428-023-02114-4

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 100L, # use a large value e.g., 20000L for actual research
  alpha = 0.05
)

## Standardized Monte Carlo ------------------------------------------------
MCStd(unstd, alpha = 0.05)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R    2.5%  97.5%
#> cp                 0.0084 0.0905 100 -0.1797 0.1890
#> b                  0.4640 0.0847 100  0.3026 0.6169
#> a                  0.1661 0.0930 100  0.0101 0.3631
#> cond~~cond         1.0000 0.0000 100  1.0000 1.0000
#> reaction~~reaction 0.7833 0.0777 100  0.6085 0.9100
#> pmi~~pmi           0.9724 0.0364 100  0.8681 0.9993
#> indirect           0.3264 0.0477 100  0.0046 0.1911
#> direct             3.9740 0.0905 100 -0.1797 0.1890
#> total              0.9214 0.0917 100 -0.0881 0.2969

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 100L, # use a large value e.g., 20000L for actual research
  alpha = 0.05
)

## Standardized Monte Carlo ------------------------------------------------
MCStd(unstd, alpha = 0.05)
#> Standardized Monte Carlo Confidence Intervals
#>                        est     se   R    2.5%  97.5%
#> cp                 -0.0382 0.1161 100 -0.2286 0.1995
#> b                   0.4897 0.0751 100  0.3565 0.6206
#> a                   0.1293 0.1110 100 -0.1245 0.3216
#> cond~~cond          1.0000 0.0000 100  1.0000 1.0000
#> reaction~~reaction  0.7635 0.0736 100  0.5990 0.8716
#> pmi~~pmi            0.9833 0.0282 100  0.8965 1.0000
#> indirect            0.0633 0.0543 100 -0.0678 0.1508
#> direct             -0.0382 0.1161 100 -0.2286 0.1995
#> total               0.0251 0.1173 100 -0.1499 0.2666