Standardized Monte Carlo Confidence Intervals

Calculates standardized Monte Carlo confidence intervals for free and defined parameters.

Usage

MCStd(mc, alpha = c(0.001, 0.01, 0.05))

Arguments

mc

Output of the MC() or MCMI() function.

alpha

Numeric vector. Significance level \(\alpha\).

Value

Returns an object of class semmcci which is a list with the following elements:

call

Function call.

args

List of function arguments.

thetahat

Parameter estimates \(\hat{\theta}\).

thetahatstar

Sampling distribution of parameter estimates \(\hat{\theta}^{\ast}\).

fun

Function used ("MCStd").

Details

The empirical sampling distribution of parameter estimates from the argument mc is standardized, that is, each randomly generated vector of parameters is standardized. Defined parameters are computed from the standardized component parameters. Confidence intervals are generated using the standardized empirical sampling distribution.

References

Pesigan, I. J. A., & Cheung, S. F. (2023). Monte Carlo confidence intervals for the indirect effect with missing data. Behavior Research Methods. doi:10.3758/s13428-023-02114-4

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 100L, # use a large value e.g., 20000L for actual research
  alpha = 0.05
)

## Standardized Monte Carlo ------------------------------------------------
MCStd(unstd, alpha = 0.05)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R    2.5%  97.5%
#> cp                 0.1656 0.0928 100  0.0011 0.3480
#> b                  0.4155 0.0776 100  0.2789 0.5846
#> a                  0.1579 0.0929 100 -0.0370 0.3200
#> cond~~cond         1.0000 0.0000 100  1.0000 1.0000
#> reaction~~reaction 0.7782 0.0692 100  0.6123 0.8834
#> pmi~~pmi           0.9751 0.0294 100  0.8976 0.9994
#> indirect           0.4091 0.0411 100 -0.0139 0.1410
#> direct             4.0147 0.0928 100  0.0011 0.3480
#> total              0.9057 0.0962 100  0.0300 0.4011

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 100L, # use a large value e.g., 20000L for actual research
  alpha = 0.05
)

## Standardized Monte Carlo ------------------------------------------------
MCStd(unstd, alpha = 0.05)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R    2.5%  97.5%
#> cp                 0.1414 0.0839 100  0.0313 0.3206
#> b                  0.4326 0.0776 100  0.2625 0.5601
#> a                  0.1034 0.0936 100 -0.0463 0.3445
#> cond~~cond         1.0000 0.0000 100  1.0000 1.0000
#> reaction~~reaction 0.7802 0.0679 100  0.6395 0.9024
#> pmi~~pmi           0.9893 0.0300 100  0.8812 0.9994
#> indirect           0.0447 0.0411 100 -0.0180 0.1505
#> direct             0.1414 0.0839 100  0.0313 0.3206
#> total              0.1861 0.0890 100  0.0658 0.3845