Monte Carlo Confidence Intervals for the Parameter Estimates

Monte Carlo Confidence Intervals for the Parameter Estimates

Usage

# S3 method for semmcci
confint(object, parm = NULL, level = 0.95, ...)

Arguments

object

Object of class semmcci.

parm

a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.

level

the confidence level required.

...

additional arguments.

Value

Returns a matrix of confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
confint(unstd)
#>                          2.5 %    97.5 %
#> cp                 -0.39780543 0.5485197
#> b                   0.31566807 0.5682637
#> a                   0.03210880 0.7453891
#> cond~~cond          0.19797622 0.3294590
#> reaction~~reaction  1.47438210 2.1818080
#> pmi~~pmi            1.04686118 2.0658581
#> reaction~1          0.21199508 1.6880193
#> pmi~1               5.09041567 5.7492333
#> cond~1              0.38131517 0.5231021
#> indirect            0.01523185 0.3910535
#> direct             -0.39780543 0.5485197
#> total              -0.27581460 0.7843364
confint(std)
#>                          2.5 %    97.5 %
#> cp                 -0.14510655 0.1689560
#> b                   0.26422057 0.5336205
#> a                   0.01221177 0.3264820
#> cond~~cond          1.00000000 1.0000000
#> reaction~~reaction  0.69944030 0.9287294
#> pmi~~pmi            0.89033846 0.9997187
#> indirect            0.00537243 0.1321498
#> direct             -0.14510655 0.1689560
#> total              -0.10021996 0.2415254

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
confint(unstd)
#>                          2.5 %    97.5 %
#> cp                 -0.77055963 0.2847204
#> b                   0.27714427 0.6265213
#> a                   0.17333283 0.7672689
#> cond~~cond          0.18300412 0.3107192
#> reaction~~reaction  1.34390417 2.3340930
#> pmi~~pmi            1.24654533 2.0689588
#> indirect            0.07899933 0.4195946
#> direct             -0.77055963 0.2847204
#> total              -0.59229367 0.4496156
confint(std)
#>                          2.5 %     97.5 %
#> cp                 -0.26415931 0.09748831
#> b                   0.25180617 0.52825277
#> a                   0.06640212 0.30574725
#> cond~~cond          1.00000000 1.00000000
#> reaction~~reaction  0.73482221 0.92640148
#> pmi~~pmi            0.90642632 0.99559036
#> indirect            0.02360781 0.13955997
#> direct             -0.26415931 0.09748831
#> total              -0.20292851 0.14618342