Monte Carlo Confidence Intervals for the Parameter Estimates

Monte Carlo Confidence Intervals for the Parameter Estimates

Usage

# S3 method for semmcci
confint(object, parm = NULL, level = 0.95, ...)

Arguments

object

Object of class semmcci.

parm

a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.

level

the confidence level required.

...

additional arguments.

Value

Returns a matrix of confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
confint(unstd)
#>                         2.5 %    97.5 %
#> cp                 -0.4434206 0.6291347
#> b                   0.3199047 0.6700595
#> a                  -0.2817425 0.7870645
#> cond~~cond          0.1831785 0.2780075
#> reaction~~reaction  1.4686404 2.3460583
#> pmi~~pmi            1.4237934 2.0217474
#> reaction~1         -0.3501571 1.5560908
#> pmi~1               5.2018819 5.8318863
#> cond~1              0.3553902 0.5235130
#> indirect           -0.1757997 0.4404426
#> direct             -0.4434206 0.6291347
#> total              -0.4014424 0.9267266
confint(std)
#>                          2.5 %    97.5 %
#> cp                 -0.14082513 0.1829051
#> b                   0.30371604 0.5779087
#> a                  -0.10607291 0.3160679
#> cond~~cond          1.00000000 1.0000000
#> reaction~~reaction  0.65708796 0.8842155
#> pmi~~pmi            0.89757452 0.9998589
#> indirect           -0.05737928 0.1514923
#> direct             -0.14082513 0.1829051
#> total              -0.12534900 0.2714504

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
confint(unstd)
#>                         2.5 %    97.5 %
#> cp                 -0.5394799 0.4359613
#> b                   0.3393585 0.7535271
#> a                  -0.1289325 0.8216298
#> cond~~cond          0.2041871 0.2750116
#> reaction~~reaction  1.5104898 2.2222464
#> pmi~~pmi            1.4404475 2.1270709
#> indirect           -0.0538671 0.3783238
#> direct             -0.5394799 0.4359613
#> total              -0.3711740 0.6642205
confint(std)
#>                          2.5 %    97.5 %
#> cp                 -0.16677206 0.1362168
#> b                   0.30963879 0.5935253
#> a                  -0.04691493 0.3143315
#> cond~~cond          1.00000000 1.0000000
#> reaction~~reaction  0.64213691 0.9044351
#> pmi~~pmi            0.90119522 0.9995152
#> indirect           -0.01636247 0.1263070
#> direct             -0.16677206 0.1362168
#> total              -0.12087775 0.2029918