Monte Carlo Confidence Intervals for the Parameter Estimates

Monte Carlo Confidence Intervals for the Parameter Estimates

Usage

# S3 method for semmcci
confint(object, parm = NULL, level = 0.95, ...)

Arguments

object

Object of class semmcci.

parm

a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.

level

the confidence level required.

...

additional arguments.

Value

Returns a matrix of confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
confint(unstd)
#>                         2.5 %    97.5 %
#> cp                 -0.3315630 0.7839978
#> b                   0.4448484 0.7676445
#> a                  -0.1516987 0.8195741
#> cond~~cond          0.2087919 0.3229464
#> reaction~~reaction  1.4560281 2.2914830
#> pmi~~pmi            1.3505845 2.0212202
#> reaction~1         -1.2695065 0.7554947
#> pmi~1               5.2197662 5.7786952
#> cond~1              0.4192167 0.5692743
#> indirect           -0.1107567 0.4596358
#> direct             -0.3315630 0.7839978
#> total              -0.0707285 1.0138153
confint(std)
#>                          2.5 %    97.5 %
#> cp                 -0.12677550 0.2149055
#> b                   0.37589019 0.5933124
#> a                  -0.05786579 0.3086314
#> cond~~cond          1.00000000 1.0000000
#> reaction~~reaction  0.64293575 0.8347392
#> pmi~~pmi            0.90424900 0.9999776
#> indirect           -0.03372012 0.1380706
#> direct             -0.12677550 0.2149055
#> total              -0.03178053 0.3038903

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
confint(unstd)
#>                           2.5 %    97.5 %
#> cp                 -0.001177133 0.9619012
#> b                   0.345528816 0.6740571
#> a                  -0.170128912 0.8080297
#> cond~~cond          0.212903057 0.2921208
#> reaction~~reaction  1.432819209 2.0880148
#> pmi~~pmi            1.212165792 2.0786840
#> indirect           -0.104567003 0.4122194
#> direct             -0.001177133 0.9619012
#> total               0.090770126 0.9886338
confint(std)
#>                            2.5 %    97.5 %
#> cp                  5.805985e-05 0.2904700
#> b                   3.015360e-01 0.5604160
#> a                  -7.478358e-02 0.2991140
#> cond~~cond          1.000000e+00 1.0000000
#> reaction~~reaction  6.422171e-01 0.8942270
#> pmi~~pmi            9.100839e-01 0.9999282
#> indirect           -3.380142e-02 0.1406798
#> direct              5.805985e-05 0.2904700
#> total               3.569215e-02 0.2918734