Monte Carlo Confidence Intervals for the Parameter Estimates

Monte Carlo Confidence Intervals for the Parameter Estimates

Usage

# S3 method for semmcci
confint(object, parm = NULL, level = 0.95, ...)

Arguments

object

Object of class semmcci.

parm

a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.

level

the confidence level required.

...

additional arguments.

Value

Returns a matrix of confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
confint(unstd)
#>                           2.5 %    97.5 %
#> cp                 -0.214553324 0.7053201
#> b                   0.444231958 0.6642972
#> a                   0.004138904 0.6617782
#> cond~~cond          0.207649141 0.3027639
#> reaction~~reaction  1.457067551 2.1993242
#> pmi~~pmi            1.350673409 2.0306110
#> reaction~1         -0.364764444 1.1018464
#> pmi~1               5.099440240 5.7332149
#> cond~1              0.403374423 0.5456881
#> indirect            0.003623563 0.3601801
#> direct             -0.214553324 0.7053201
#> total              -0.187192347 1.0655002
confint(std)
#>                            2.5 %    97.5 %
#> cp                 -0.0656376887 0.2043794
#> b                   0.3697871591 0.5538773
#> a                   0.0010781029 0.2442717
#> cond~~cond          1.0000000000 1.0000000
#> reaction~~reaction  0.6845470676 0.8407877
#> pmi~~pmi            0.9403093480 0.9999441
#> indirect            0.0009576158 0.1114146
#> direct             -0.0656376887 0.2043794
#> total              -0.0556670180 0.3097529

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
confint(unstd)
#>                           2.5 %    97.5 %
#> cp                 -0.221728410 0.9367991
#> b                   0.345357830 0.6376603
#> a                   0.004977875 0.6038626
#> cond~~cond          0.183238400 0.2958147
#> reaction~~reaction  1.367220123 2.3478640
#> pmi~~pmi            1.386277762 2.1073964
#> indirect            0.002712093 0.2786031
#> direct             -0.221728410 0.9367991
#> total              -0.168206995 1.1283020
confint(std)
#>                            2.5 %     97.5 %
#> cp                 -0.0769530422 0.28249524
#> b                   0.2767826999 0.58936949
#> a                   0.0018938675 0.23343675
#> cond~~cond          1.0000000000 1.00000000
#> reaction~~reaction  0.6246689325 0.90211506
#> pmi~~pmi            0.9452356460 0.99999632
#> indirect            0.0009380358 0.09292105
#> direct             -0.0769530422 0.28249524
#> total              -0.0585466757 0.34415511