Monte Carlo Confidence Intervals for the Parameter Estimates

Monte Carlo Confidence Intervals for the Parameter Estimates

Usage

# S3 method for semmcci
confint(object, parm = NULL, level = 0.95, ...)

Arguments

object

Object of class semmcci.

parm

a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.

level

the confidence level required.

...

additional arguments.

Value

Returns a matrix of confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
confint(unstd)
#>                          2.5 %    97.5 %
#> cp                 -0.33973885 0.7489992
#> b                   0.23980400 0.6434866
#> a                  -0.04906937 1.0414413
#> cond~~cond          0.17668480 0.3151498
#> reaction~~reaction  1.27342334 2.4463271
#> pmi~~pmi            1.37330474 2.1944089
#> reaction~1         -0.17620912 2.1059346
#> pmi~1               4.93873463 5.7018802
#> cond~1              0.38828457 0.5664771
#> indirect           -0.02300270 0.5288534
#> direct             -0.33973885 0.7489992
#> total              -0.10704119 0.9935174
confint(std)
#>                           2.5 %    97.5 %
#> cp                 -0.117553528 0.2432426
#> b                   0.216207362 0.5541482
#> a                  -0.018322896 0.3849620
#> cond~~cond          1.000000000 1.0000000
#> reaction~~reaction  0.661458985 0.9461697
#> pmi~~pmi            0.851803609 0.9996724
#> indirect           -0.007880333 0.1649807
#> direct             -0.117553528 0.2432426
#> total              -0.040552572 0.3416846

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
confint(unstd)
#>                            2.5 %    97.5 %
#> cp                 -0.2326306616 0.8656293
#> b                   0.2500186781 0.6227070
#> a                   0.0009871247 0.8985578
#> cond~~cond          0.1733132231 0.3234662
#> reaction~~reaction  1.4900550782 2.3768064
#> pmi~~pmi            1.2864216520 2.1126492
#> indirect           -0.0001556533 0.4402331
#> direct             -0.2326306616 0.8656293
#> total              -0.0078304520 1.0375572
confint(std)
#>                            2.5 %    97.5 %
#> cp                 -7.550508e-02 0.2867835
#> b                   2.141753e-01 0.5128836
#> a                   5.668209e-04 0.3251722
#> cond~~cond          1.000000e+00 1.0000000
#> reaction~~reaction  7.130692e-01 0.9264291
#> pmi~~pmi            8.942336e-01 0.9996669
#> indirect           -2.535189e-05 0.1397379
#> direct             -7.550508e-02 0.2867835
#> total              -2.904763e-03 0.3309403