Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.4471 0.2944 20 -0.0114  0.0058  0.0825 1.1391 1.2127
#> b                  0.5029 0.0772 20  0.3811  0.3819  0.3858 0.6356 0.6395
#> a                  0.3964 0.2608 20 -0.4058 -0.3739 -0.2317 0.6883 0.7129
#> cond~~cond         0.2489 0.0286 20  0.1932  0.1937  0.1961 0.2838 0.2841
#> reaction~~reaction 1.8328 0.2016 20  1.4875  1.4993  1.5522 2.2740 2.2758
#> pmi~~pmi           1.8205 0.2461 20  1.2760  1.2992  1.4025 2.1590 2.1701
#> reaction~1         0.4643 0.4515 20 -0.2456 -0.2413 -0.2223 1.1383 1.1877
#> pmi~1              5.3809 0.1754 20  5.0358  5.0421  5.0700 5.6886 5.7831
#> cond~1             0.4728 0.0432 20  0.3951  0.3977  0.4090 0.5619 0.5676
#> indirect           0.1993 0.1475 20 -0.2290 -0.2111 -0.1318 0.3828 0.3863
#> direct             0.4471 0.2944 20 -0.0114  0.0058  0.0825 1.1391 1.2127
#> total              0.6465 0.3258 20  0.0767  0.0863  0.1289 1.3433 1.3910
#>                    99.95%
#> cp                 1.2292
#> b                  0.6403
#> a                  0.7184
#> cond~~cond         0.2842
#> reaction~~reaction 2.2762
#> pmi~~pmi           2.1726
#> reaction~1         1.1988
#> pmi~1              5.8043
#> cond~1             0.5689
#> indirect           0.3871
#> direct             1.2292
#> total              1.4018
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1441 0.0929 20 -0.0042  0.0019  0.0290 0.3718 0.3887
#> b                  0.4429 0.0650 20  0.3522  0.3528  0.3552 0.5483 0.5529
#> a                  0.1450 0.0979 20 -0.1493 -0.1374 -0.0847 0.2654 0.2663
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7645 0.0679 20  0.6237  0.6255  0.6337 0.8459 0.8610
#> pmi~~pmi           0.9790 0.0204 20  0.9290  0.9291  0.9296 0.9995 0.9998
#> indirect           0.2999 0.0471 20 -0.0791 -0.0728 -0.0449 0.1175 0.1181
#> direct             3.9458 0.0929 20 -0.0042  0.0019  0.0290 0.3718 0.3887
#> total              0.9477 0.1025 20  0.0265  0.0301  0.0463 0.4388 0.4460
#>                    99.95%
#> cp                 0.3925
#> b                  0.5540
#> a                  0.2665
#> cond~~cond         1.0000
#> reaction~~reaction 0.8644
#> pmi~~pmi           0.9999
#> indirect           0.1182
#> direct             0.3925
#> total              0.4476

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se  R   0.05%    0.5%   2.5%  97.5%  99.5% 99.95%
#> cp                 0.3847 0.3193 20 -0.0117 -0.0094 0.0009 0.9210 0.9522 0.9592
#> b                  0.5046 0.0939 20  0.3085  0.3120 0.3273 0.6566 0.6610 0.6620
#> a                  0.4172 0.1938 20 -0.0475 -0.0373 0.0083 0.7251 0.7452 0.7497
#> cond~~cond         0.2485 0.0320 20  0.1910  0.1918 0.1956 0.2945 0.2964 0.2968
#> reaction~~reaction 1.8470 0.2437 20  1.4029  1.4129 1.4575 2.2816 2.2934 2.2960
#> pmi~~pmi           1.7810 0.2567 20  1.3902  1.3931 1.4060 2.2073 2.2097 2.2102
#> indirect           0.2103 0.0941 20 -0.0144 -0.0092 0.0142 0.3312 0.3313 0.3313
#> direct             0.3847 0.3193 20 -0.0117 -0.0094 0.0009 0.9210 0.9522 0.9592
#> total              0.5950 0.3417 20  0.0627  0.0738 0.1232 1.2006 1.2520 1.2636
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%   2.5%  97.5%  99.5% 99.95%
#> cp                 0.2025 0.1001 20 -0.0038 -0.0031 0.0000 0.2864 0.2906 0.2916
#> b                  0.4654 0.0880 20  0.2699  0.2710 0.2761 0.5718 0.5877 0.5913
#> a                  0.2123 0.0739 20 -0.0176 -0.0142 0.0012 0.2797 0.2972 0.3012
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000 1.0000 1.0000 1.0000 1.0000
#> reaction~~reaction 0.7023 0.0866 20  0.6313  0.6314 0.6319 0.9144 0.9202 0.9215
#> pmi~~pmi           0.9549 0.0220 20  0.9093  0.9115 0.9212 0.9996 0.9997 0.9997
#> indirect           0.0988 0.0313 20 -0.0049 -0.0033 0.0036 0.1141 0.1150 0.1153
#> direct             0.2025 0.1001 20 -0.0038 -0.0031 0.0000 0.2864 0.2906 0.2916
#> total              0.3014 0.1057 20  0.0180  0.0225 0.0424 0.3556 0.3786 0.3838