Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0177 0.3270 20 -0.5030 -0.4920 -0.4434 0.6291 0.7592
#> b                  0.5231 0.1172 20  0.3075  0.3098  0.3199 0.6701 0.6803
#> a                  0.3979 0.3291 20 -0.4448 -0.4148 -0.2817 0.7871 0.8730
#> cond~~cond         0.2464 0.0273 20  0.1735  0.1753  0.1832 0.2780 0.2875
#> reaction~~reaction 1.8600 0.2548 20  1.4581  1.4601  1.4686 2.3461 2.4836
#> pmi~~pmi           1.7112 0.1929 20  1.4121  1.4143  1.4238 2.0217 2.0372
#> reaction~1         0.5229 0.6248 20 -0.4526 -0.4338 -0.3502 1.5561 1.6383
#> pmi~1              5.4286 0.2044 20  5.1965  5.1975  5.2019 5.8319 5.8451
#> cond~1             0.4439 0.0509 20  0.3459  0.3476  0.3554 0.5235 0.5320
#> indirect           0.2081 0.1792 20 -0.2757 -0.2574 -0.1758 0.4404 0.4555
#> direct             0.0177 0.3270 20 -0.5030 -0.4920 -0.4434 0.6291 0.7592
#> total              0.2259 0.3999 20 -0.4112 -0.4094 -0.4014 0.9267 0.9905
#>                    99.95%
#> cp                 0.7885
#> b                  0.6826
#> a                  0.8924
#> cond~~cond         0.2896
#> reaction~~reaction 2.5146
#> pmi~~pmi           2.0406
#> reaction~1         1.6568
#> pmi~1              5.8481
#> cond~1             0.5339
#> indirect           0.4589
#> direct             0.7885
#> total              1.0048
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0058 0.1007 20 -0.1691 -0.1639 -0.1408 0.1829 0.2155
#> b                  0.4523 0.0957 20  0.2920  0.2941  0.3037 0.5779 0.5850
#> a                  0.1493 0.1246 20 -0.1724 -0.1602 -0.1061 0.3161 0.3543
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7946 0.0801 20  0.6490  0.6505  0.6571 0.8842 0.8940
#> pmi~~pmi           0.9777 0.0326 20  0.8682  0.8736  0.8976 0.9999 0.9999
#> indirect           0.3418 0.0591 20 -0.0925 -0.0861 -0.0574 0.1515 0.1705
#> direct             4.1034 0.1007 20 -0.1691 -0.1639 -0.1408 0.1829 0.2155
#> total              0.8943 0.1202 20 -0.1382 -0.1358 -0.1253 0.2715 0.2818
#>                    99.95%
#> cp                 0.2229
#> b                  0.5866
#> a                  0.3629
#> cond~~cond         1.0000
#> reaction~~reaction 0.8961
#> pmi~~pmi           0.9999
#> indirect           0.1747
#> direct             0.2229
#> total              0.2841

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0394 0.3213 20 -0.5610 -0.5571 -0.5395 0.4360 0.4422
#> b                  0.5074 0.1259 20  0.3250  0.3276  0.3394 0.7535 0.8546
#> a                  0.3742 0.2439 20 -0.2691 -0.2433 -0.1289 0.8216 0.8447
#> cond~~cond         0.2442 0.0249 20  0.2039  0.2039  0.2042 0.2750 0.2754
#> reaction~~reaction 1.8845 0.2018 20  1.5027  1.5041  1.5105 2.2222 2.2629
#> pmi~~pmi           1.7159 0.1860 20  1.4303  1.4322  1.4404 2.1271 2.2056
#> indirect           0.1905 0.1225 20 -0.1141 -0.1031 -0.0539 0.3783 0.4523
#> direct             0.0394 0.3213 20 -0.5610 -0.5571 -0.5395 0.4360 0.4422
#> total              0.2299 0.3358 20 -0.3994 -0.3942 -0.3712 0.6642 0.6802
#>                    99.95%
#> cp                 0.4436
#> b                  0.8774
#> a                  0.8499
#> cond~~cond         0.2755
#> reaction~~reaction 2.2721
#> pmi~~pmi           2.2232
#> indirect           0.4689
#> direct             0.4436
#> total              0.6838
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0289 0.1025 20 -0.1810 -0.1784 -0.1668 0.1362 0.1388
#> b                  0.5229 0.0853 20  0.3085  0.3087  0.3096 0.5935 0.6494
#> a                  0.1188 0.0927 20 -0.0984 -0.0889 -0.0469 0.3143 0.3148
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7221 0.0794 20  0.5570  0.5726  0.6421 0.9044 0.9066
#> pmi~~pmi           0.9859 0.0279 20  0.9008  0.9009  0.9012 0.9995 0.9998
#> indirect           0.0621 0.0389 20 -0.0347 -0.0313 -0.0164 0.1263 0.1474
#> direct             0.0289 0.1025 20 -0.1810 -0.1784 -0.1668 0.1362 0.1388
#> total              0.0910 0.1069 20 -0.1228 -0.1225 -0.1209 0.2030 0.2126
#>                    99.95%
#> cp                 0.1394
#> b                  0.6619
#> a                  0.3150
#> cond~~cond         1.0000
#> reaction~~reaction 0.9071
#> pmi~~pmi           0.9999
#> indirect           0.1521
#> direct             0.1394
#> total              0.2148