Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.2311 0.2485 100 -0.5807 -0.4971 -0.1419 0.7388 0.7903
#> b                  0.4743 0.0967 100  0.2544  0.2616  0.2870 0.6577 0.7120
#> a                  0.4903 0.2465 100 -0.1924 -0.1299  0.0452 1.0540 1.1821
#> cond~~cond         0.2491 0.0369 100  0.1304  0.1546  0.1865 0.3202 0.3346
#> reaction~~reaction 1.8017 0.2325 100  1.2025  1.2370  1.3909 2.2159 2.4051
#> pmi~~pmi           1.7440 0.2432 100  1.1259  1.1701  1.2453 2.2169 2.4041
#> reaction~1         0.6536 0.5459 100 -0.7898 -0.7481 -0.3894 1.6867 1.8960
#> pmi~1              5.3875 0.1908 100  4.7515  4.8824  5.0727 5.7357 5.8505
#> cond~1             0.4718 0.0404 100  0.3798  0.3800  0.3910 0.5406 0.5666
#> indirect           0.2325 0.1379 100 -0.1229 -0.0785  0.0224 0.5604 0.6662
#> direct             0.2311 0.2485 100 -0.5807 -0.4971 -0.1419 0.7388 0.7903
#> total              0.4636 0.2949 100 -0.5344 -0.4249 -0.0565 1.0727 1.3058
#>                    99.95%
#> cp                 0.7992
#> b                  0.7175
#> a                  1.1851
#> cond~~cond         0.3347
#> reaction~~reaction 2.5000
#> pmi~~pmi           2.5085
#> reaction~1         1.9438
#> pmi~1              5.9082
#> cond~1             0.5762
#> indirect           0.7098
#> direct             0.7992
#> total              1.4939
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0769 0.0826 100 -0.1980 -0.1758 -0.0483 0.2246 0.2500
#> b                  0.4249 0.0803 100  0.2287  0.2334  0.2592 0.5675 0.5930
#> a                  0.1822 0.0901 100 -0.0678 -0.0468  0.0158 0.3607 0.4384
#> cond~~cond         1.0000 0.0000 100  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.8016 0.0696 100  0.5868  0.6040  0.6641 0.9017 0.9273
#> pmi~~pmi           0.9668 0.0366 100  0.7644  0.8051  0.8698 0.9989 0.9997
#> indirect           0.4360 0.0423 100 -0.0370 -0.0242  0.0075 0.1738 0.1998
#> direct             4.0113 0.0826 100 -0.1980 -0.1758 -0.0483 0.2246 0.2500
#> total              0.9454 0.0951 100 -0.1821 -0.1495 -0.0195 0.3565 0.3875
#>                    99.95%
#> cp                 0.2657
#> b                  0.6095
#> a                  0.4849
#> cond~~cond         1.0000
#> reaction~~reaction 0.9403
#> pmi~~pmi           1.0000
#> indirect           0.2078
#> direct             0.2657
#> total              0.4028

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.3195 0.2515 100 -0.3592 -0.2917 -0.1886 0.7394 0.9670
#> b                  0.4968 0.1023 100  0.2843  0.2877  0.3141 0.6734 0.7260
#> a                  0.4631 0.2438 100 -0.0677 -0.0260  0.0411 0.9338 1.0214
#> cond~~cond         0.2492 0.0367 100  0.1633  0.1671  0.1787 0.3136 0.3313
#> reaction~~reaction 1.7871 0.2702 100  1.0615  1.0823  1.2028 2.2847 2.3367
#> pmi~~pmi           1.6823 0.2024 100  1.1767  1.2013  1.2714 2.0005 2.1663
#> indirect           0.2300 0.1312 100 -0.0395 -0.0150  0.0233 0.5301 0.5901
#> direct             0.3195 0.2515 100 -0.3592 -0.2917 -0.1886 0.7394 0.9670
#> total              0.5494 0.2755 100 -0.1581 -0.1539  0.0080 1.0691 1.1980
#>                    99.95%
#> cp                 0.9967
#> b                  0.7334
#> a                  1.0315
#> cond~~cond         0.3403
#> reaction~~reaction 2.3484
#> pmi~~pmi           2.2074
#> indirect           0.6036
#> direct             0.9967
#> total              1.1980
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1351 0.0817 100 -0.1310 -0.1015 -0.0578 0.2471 0.3009
#> b                  0.4246 0.0774 100  0.2482  0.2587  0.2843 0.5438 0.5954
#> a                  0.2082 0.0901 100 -0.0248 -0.0092  0.0171 0.3426 0.3988
#> cond~~cond         1.0000 0.0000 100  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7776 0.0706 100  0.6305  0.6390  0.6569 0.9027 0.9204
#> pmi~~pmi           0.9566 0.0354 100  0.8199  0.8401  0.8826 0.9995 0.9998
#> indirect           0.0884 0.0423 100 -0.0133 -0.0049  0.0074 0.1651 0.1778
#> direct             0.1351 0.0817 100 -0.1310 -0.1015 -0.0578 0.2471 0.3009
#> total              0.2235 0.0898 100 -0.0545 -0.0541  0.0031 0.3350 0.3799
#>                    99.95%
#> cp                 0.3328
#> b                  0.6141
#> a                  0.4242
#> cond~~cond         1.0000
#> reaction~~reaction 0.9255
#> pmi~~pmi           0.9999
#> indirect           0.1789
#> direct             0.3328
#> total              0.4007