Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.3886 0.3053 100 -0.3348 -0.3328 -0.2723 1.0084 1.0850
#> b                  0.5207 0.1016 100  0.2859  0.2863  0.3367 0.7456 0.7839
#> a                  0.4344 0.2800 100 -0.1780 -0.1421 -0.0463 1.0103 1.0779
#> cond~~cond         0.2478 0.0312 100  0.1641  0.1679  0.1973 0.3023 0.3161
#> reaction~~reaction 1.9325 0.2727 100  1.1182  1.1984  1.3463 2.4793 2.6422
#> pmi~~pmi           1.7281 0.2320 100  1.2370  1.2449  1.3236 2.1497 2.3351
#> reaction~1         0.4436 0.5751 100 -1.3307 -1.0925 -0.7279 1.4072 1.7638
#> pmi~1              5.3359 0.1781 100  4.9252  4.9312  5.0016 5.6658 5.7362
#> cond~1             0.4587 0.0467 100  0.3182  0.3184  0.3726 0.5273 0.5367
#> indirect           0.2262 0.1537 100 -0.0845 -0.0645 -0.0248 0.5822 0.6686
#> direct             0.3886 0.3053 100 -0.3348 -0.3328 -0.2723 1.0084 1.0850
#> total              0.6148 0.3292 100 -0.1560 -0.1430 -0.0843 1.1790 1.2376
#>                    99.95%
#> cp                 1.1183
#> b                  0.7968
#> a                  1.0802
#> cond~~cond         0.3180
#> reaction~~reaction 2.6789
#> pmi~~pmi           2.4529
#> reaction~1         1.8583
#> pmi~1              5.7482
#> cond~1             0.5438
#> indirect           0.7089
#> direct             1.1183
#> total              1.2768
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1225 0.0955 100 -0.1103 -0.1092 -0.0768 0.3010 0.3427
#> b                  0.4392 0.0737 100  0.2593  0.2756  0.3148 0.6019 0.6162
#> a                  0.1623 0.1013 100 -0.0645 -0.0524 -0.0140 0.3465 0.3713
#> cond~~cond         1.0000 0.0000 100  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7746 0.0722 100  0.5109  0.5304  0.5882 0.8827 0.8949
#> pmi~~pmi           0.9737 0.0349 100  0.8555  0.8621  0.8797 0.9998 1.0000
#> indirect           0.2809 0.0460 100 -0.0265 -0.0204 -0.0069 0.1664 0.1830
#> direct             4.0052 0.0955 100 -0.1103 -0.1092 -0.0768 0.3010 0.3427
#> total              0.9215 0.1016 100 -0.0447 -0.0441 -0.0252 0.3855 0.3928
#>                    99.95%
#> cp                 0.3657
#> b                  0.6249
#> a                  0.3801
#> cond~~cond         1.0000
#> reaction~~reaction 0.9025
#> pmi~~pmi           1.0000
#> indirect           0.1858
#> direct             0.3657
#> total              0.3938

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.3787 0.2829 100 -0.3703 -0.2819 -0.1121 0.9541 0.9811
#> b                  0.4860 0.1063 100  0.1765  0.2187  0.2875 0.7005 0.7790
#> a                  0.3449 0.2266 100 -0.2360 -0.1842 -0.0955 0.7212 0.7645
#> cond~~cond         0.2483 0.0313 100  0.1662  0.1819  0.2056 0.3137 0.3341
#> reaction~~reaction 2.0435 0.2937 100  1.1632  1.2742  1.4036 2.4640 2.6606
#> pmi~~pmi           1.7375 0.2335 100  1.2014  1.2243  1.3311 2.2676 2.3094
#> indirect           0.1666 0.1066 100 -0.1277 -0.0913 -0.0379 0.3318 0.3949
#> direct             0.3787 0.2829 100 -0.3703 -0.2819 -0.1121 0.9541 0.9811
#> total              0.5453 0.2970 100 -0.1879 -0.1643  0.0082 1.1138 1.2424
#>                    99.95%
#> cp                 0.9975
#> b                  0.8055
#> a                  0.7668
#> cond~~cond         0.3444
#> reaction~~reaction 2.8250
#> pmi~~pmi           2.3297
#> indirect           0.4385
#> direct             0.9975
#> total              1.2786
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1015 0.0852 100 -0.1091 -0.0864 -0.0368 0.2776 0.2992
#> b                  0.4471 0.0910 100  0.1497  0.1778  0.2372 0.5840 0.6631
#> a                  0.1216 0.0870 100 -0.0786 -0.0645 -0.0381 0.2773 0.3150
#> cond~~cond         1.0000 0.0000 100  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7788 0.0756 100  0.4768  0.5388  0.6522 0.9160 0.9303
#> pmi~~pmi           0.9852 0.0231 100  0.8991  0.9008  0.9230 0.9999 1.0000
#> indirect           0.0544 0.0342 100 -0.0363 -0.0279 -0.0117 0.1111 0.1266
#> direct             0.1015 0.0852 100 -0.1091 -0.0864 -0.0368 0.2776 0.2992
#> total              0.1559 0.0894 100 -0.0554 -0.0501  0.0030 0.3305 0.3869
#>                    99.95%
#> cp                 0.3028
#> b                  0.7214
#> a                  0.3177
#> cond~~cond         1.0000
#> reaction~~reaction 0.9424
#> pmi~~pmi           1.0000
#> indirect           0.1333
#> direct             0.3028
#> total              0.3928