Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.3562 0.3233 20 -0.4482 -0.4179 -0.2834 0.8379 0.8802
#> b                  0.4715 0.1406 20  0.1705  0.1796  0.2200 0.6723 0.6889
#> a                  0.4071 0.2853 20 -0.0090  0.0091  0.0894 1.0698 1.0817
#> cond~~cond         0.2481 0.0256 20  0.1977  0.2007  0.2142 0.3094 0.3237
#> reaction~~reaction 1.9496 0.2970 20  1.2455  1.2731  1.3957 2.3798 2.4146
#> pmi~~pmi           1.7868 0.2690 20  1.2512  1.2577  1.2868 2.1467 2.1494
#> reaction~1         0.6929 0.7190 20 -0.4168 -0.3994 -0.3217 1.9827 2.1486
#> pmi~1              5.3715 0.1412 20  5.0249  5.0251  5.0262 5.4705 5.4822
#> cond~1             0.4570 0.0410 20  0.3796  0.3820  0.3927 0.5377 0.5411
#> indirect           0.1919 0.1907 20 -0.0009  0.0081  0.0479 0.6732 0.6818
#> direct             0.3562 0.3233 20 -0.4482 -0.4179 -0.2834 0.8379 0.8802
#> total              0.5481 0.3306 20 -0.1354 -0.1141 -0.0199 1.0613 1.0715
#>                    99.95%
#> cp                 0.8898
#> b                  0.6927
#> a                  1.0843
#> cond~~cond         0.3269
#> reaction~~reaction 2.4224
#> pmi~~pmi           2.1500
#> reaction~1         2.1859
#> pmi~1              5.4848
#> cond~1             0.5418
#> indirect           0.6837
#> direct             0.8898
#> total              1.0738
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1140 0.1086 20 -0.1337 -0.1252 -0.0875 0.3025 0.3193
#> b                  0.4097 0.1127 20  0.1823  0.1900  0.2240 0.5969 0.6095
#> a                  0.1500 0.1000 20 -0.0029  0.0047  0.0384 0.3826 0.3832
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.8052 0.0920 20  0.5839  0.5884  0.6086 0.9031 0.9139
#> pmi~~pmi           0.9775 0.0424 20  0.8531  0.8532  0.8536 0.9966 0.9993
#> indirect           0.4453 0.0603 20 -0.0004  0.0023  0.0141 0.2105 0.2188
#> direct             3.9730 0.1086 20 -0.1337 -0.1252 -0.0875 0.3025 0.3193
#> total              0.9175 0.1064 20 -0.0403 -0.0335 -0.0031 0.3292 0.3338
#>                    99.95%
#> cp                 0.3231
#> b                  0.6124
#> a                  0.3833
#> cond~~cond         1.0000
#> reaction~~reaction 0.9163
#> pmi~~pmi           0.9999
#> indirect           0.2207
#> direct             0.3231
#> total              0.3348

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.3624 0.2420 20 -0.0397 -0.0317  0.0038 0.7871 0.8531
#> b                  0.4530 0.1230 20  0.1325  0.1437  0.1938 0.6263 0.6541
#> a                  0.3494 0.3553 20 -0.7020 -0.6549 -0.4457 0.8888 0.9631
#> cond~~cond         0.2481 0.0290 20  0.1949  0.1957  0.1996 0.3036 0.3044
#> reaction~~reaction 1.9579 0.2906 20  1.6464  1.6489  1.6600 2.5496 2.5626
#> pmi~~pmi           1.7744 0.3276 20  1.2108  1.2134  1.2254 2.3905 2.4195
#> indirect           0.1572 0.1486 20 -0.2650 -0.2510 -0.1891 0.3761 0.3920
#> direct             0.3624 0.2420 20 -0.0397 -0.0317  0.0038 0.7871 0.8531
#> total              0.5196 0.2510 20  0.0276  0.0289  0.0346 0.8806 0.9090
#>                    99.95%
#> cp                 0.8680
#> b                  0.6604
#> a                  0.9799
#> cond~~cond         0.3046
#> reaction~~reaction 2.5655
#> pmi~~pmi           2.4261
#> indirect           0.3955
#> direct             0.8680
#> total              0.9154
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1290 0.0840 20 -0.0130 -0.0102  0.0019 0.2798 0.3092
#> b                  0.3793 0.1050 20  0.1273  0.1360  0.1747 0.5514 0.6084
#> a                  0.1339 0.1276 20 -0.2380 -0.2221 -0.1513 0.3185 0.3368
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.8264 0.0761 20  0.6014  0.6133  0.6661 0.9221 0.9223
#> pmi~~pmi           0.9821 0.0311 20  0.8837  0.8863  0.8980 0.9999 0.9999
#> indirect           0.0508 0.0474 20 -0.0888 -0.0840 -0.0628 0.1190 0.1244
#> direct             0.1290 0.0840 20 -0.0130 -0.0102  0.0019 0.2798 0.3092
#> total              0.1798 0.0848 20  0.0090  0.0095  0.0115 0.3017 0.3271
#>                    99.95%
#> cp                 0.3158
#> b                  0.6213
#> a                  0.3410
#> cond~~cond         1.0000
#> reaction~~reaction 0.9223
#> pmi~~pmi           0.9999
#> indirect           0.1256
#> direct             0.3158
#> total              0.3328