Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.4738 0.3898 20 -0.2612 -0.2452 -0.1741 1.1016 1.1833
#> b                  0.5043 0.0722 20  0.3440  0.3464  0.3571 0.6186 0.6192
#> a                  0.4113 0.2988 20 -0.1750 -0.1678 -0.1358 0.8855 1.0011
#> cond~~cond         0.2465 0.0261 20  0.2015  0.2025  0.2069 0.2923 0.3012
#> reaction~~reaction 1.8012 0.2250 20  1.1214  1.1448  1.2486 2.0230 2.0590
#> pmi~~pmi           1.6987 0.1745 20  1.2922  1.3038  1.3554 1.9434 1.9480
#> reaction~1         0.5628 0.4176 20 -0.1249 -0.1197 -0.0964 1.2443 1.3095
#> pmi~1              5.3900 0.1758 20  5.1099  5.1152  5.1389 5.7426 5.7516
#> cond~1             0.4478 0.0523 20  0.3472  0.3487  0.3553 0.5308 0.5433
#> indirect           0.2074 0.1421 20 -0.0999 -0.0956 -0.0763 0.3812 0.3825
#> direct             0.4738 0.3898 20 -0.2612 -0.2452 -0.1741 1.1016 1.1833
#> total              0.6812 0.3884 20 -0.1495 -0.1188  0.0175 1.2301 1.2637
#>                    99.95%
#> cp                 1.2017
#> b                  0.6193
#> a                  1.0271
#> cond~~cond         0.3032
#> reaction~~reaction 2.0671
#> pmi~~pmi           1.9491
#> reaction~1         1.3241
#> pmi~1              5.7536
#> cond~1             0.5461
#> indirect           0.3829
#> direct             1.2017
#> total              1.2713
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1535 0.1248 20 -0.0815 -0.0763 -0.0535 0.3462 0.3549
#> b                  0.4342 0.0557 20  0.2837  0.2926  0.3321 0.5337 0.5464
#> a                  0.1548 0.1084 20 -0.0653 -0.0630 -0.0524 0.3098 0.3470
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7673 0.0597 20  0.6294  0.6323  0.6454 0.8356 0.8422
#> pmi~~pmi           0.9760 0.0312 20  0.8736  0.8788  0.9016 0.9992 0.9993
#> indirect           0.3673 0.0477 20 -0.0297 -0.0286 -0.0242 0.1345 0.1420
#> direct             4.0857 0.1248 20 -0.0815 -0.0763 -0.0535 0.3462 0.3549
#> total              0.9019 0.1280 20 -0.0466 -0.0369  0.0060 0.4365 0.4698
#>                    99.95%
#> cp                 0.3568
#> b                  0.5493
#> a                  0.3554
#> cond~~cond         1.0000
#> reaction~~reaction 0.8437
#> pmi~~pmi           0.9993
#> indirect           0.1437
#> direct             0.3568
#> total              0.4773

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.4506 0.2123 20  0.0117  0.0223  0.0694 0.7803 0.8064
#> b                  0.5086 0.0817 20  0.4161  0.4166  0.4190 0.6859 0.7174
#> a                  0.3448 0.2351 20 -0.0754 -0.0683 -0.0369 0.8475 0.8553
#> cond~~cond         0.2457 0.0315 20  0.1700  0.1719  0.1806 0.2936 0.3004
#> reaction~~reaction 1.8086 0.3512 20  1.0155  1.0440  1.1707 2.4218 2.5035
#> pmi~~pmi           1.6953 0.2361 20  1.2304  1.2379  1.2717 2.0373 2.0661
#> indirect           0.1749 0.1123 20 -0.0484 -0.0439 -0.0239 0.3714 0.3761
#> direct             0.4506 0.2123 20  0.0117  0.0223  0.0694 0.7803 0.8064
#> total              0.6255 0.2313 20  0.2497  0.2585  0.2975 1.1127 1.1747
#>                    99.95%
#> cp                 0.8123
#> b                  0.7245
#> a                  0.8571
#> cond~~cond         0.3019
#> reaction~~reaction 2.5219
#> pmi~~pmi           2.0726
#> indirect           0.3772
#> direct             0.8123
#> total              1.1887
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1566 0.0729 20  0.0037  0.0069  0.0212 0.2735 0.3019
#> b                  0.4471 0.0757 20  0.3122  0.3164  0.3350 0.5844 0.5884
#> a                  0.1192 0.0965 20 -0.0250 -0.0226 -0.0120 0.3336 0.3486
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7589 0.0845 20  0.5962  0.5967  0.5991 0.8678 0.8788
#> pmi~~pmi           0.9858 0.0350 20  0.8761  0.8783  0.8883 0.9997 0.9999
#> indirect           0.0533 0.0393 20 -0.0147 -0.0134 -0.0072 0.1363 0.1420
#> direct             0.1566 0.0729 20  0.0037  0.0069  0.0212 0.2735 0.3019
#> total              0.2099 0.0870 20  0.0796  0.0820  0.0925 0.4093 0.4439
#>                    99.95%
#> cp                 0.3084
#> b                  0.5893
#> a                  0.3520
#> cond~~cond         1.0000
#> reaction~~reaction 0.8813
#> pmi~~pmi           1.0000
#> indirect           0.1433
#> direct             0.3084
#> total              0.4517