Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.2912 0.3113 100 -0.6507 -0.5687 -0.3798 0.7844 0.8796
#> b                  0.5285 0.0990 100  0.3256  0.3368  0.3582 0.7119 0.7763
#> a                  0.4357 0.2566 100 -0.2491 -0.2030 -0.0943 0.9626 0.9824
#> cond~~cond         0.2490 0.0338 100  0.1587  0.1718  0.1919 0.3144 0.3288
#> reaction~~reaction 1.8454 0.2394 100  1.4905  1.5103  1.5435 2.4779 2.5691
#> pmi~~pmi           1.7262 0.2383 100  1.1722  1.1779  1.2188 2.1426 2.3034
#> reaction~1         0.4414 0.5530 100 -1.3289 -1.0007 -0.4955 1.5042 1.6704
#> pmi~1              5.3472 0.1558 100  4.9926  5.0253  5.0814 5.6778 5.7112
#> cond~1             0.4752 0.0502 100  0.3521  0.3572  0.3831 0.5654 0.6145
#> indirect           0.2303 0.1504 100 -0.1552 -0.1142 -0.0416 0.5847 0.6331
#> direct             0.2912 0.3113 100 -0.6507 -0.5687 -0.3798 0.7844 0.8796
#> total              0.5214 0.3221 100 -0.2806 -0.2572 -0.2177 1.0593 1.2147
#>                    99.95%
#> cp                 0.8891
#> b                  0.8027
#> a                  0.9876
#> cond~~cond         0.3411
#> reaction~~reaction 2.5852
#> pmi~~pmi           2.3097
#> reaction~1         1.7382
#> pmi~1              5.7133
#> cond~1             0.6173
#> indirect           0.6379
#> direct             0.8891
#> total              1.2228
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0939 0.0985 100 -0.2081 -0.1887 -0.1327 0.2421 0.2548
#> b                  0.4548 0.0724 100  0.2609  0.2871  0.3214 0.5755 0.5971
#> a                  0.1632 0.0944 100 -0.0937 -0.0722 -0.0308 0.3358 0.3602
#> cond~~cond         1.0000 0.0000 100  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7704 0.0644 100  0.5802  0.6041  0.6539 0.8797 0.8842
#> pmi~~pmi           0.9734 0.0320 100  0.8689  0.8703  0.8872 0.9998 1.0000
#> indirect           0.2852 0.0455 100 -0.0482 -0.0345 -0.0120 0.1782 0.1927
#> direct             4.0153 0.0985 100 -0.2081 -0.1887 -0.1327 0.2421 0.2548
#> total              0.9524 0.0994 100 -0.0980 -0.0865 -0.0685 0.3248 0.3486
#>                    99.95%
#> cp                 0.2567
#> b                  0.6048
#> a                  0.3621
#> cond~~cond         1.0000
#> reaction~~reaction 0.8864
#> pmi~~pmi           1.0000
#> indirect           0.1934
#> direct             0.2567
#> total              0.3552

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.3035 0.2610 100 -0.4462 -0.4252 -0.3057 0.7598 0.9100
#> b                  0.5207 0.0953 100  0.2287  0.2478  0.3654 0.7112 0.7400
#> a                  0.4351 0.2854 100 -0.1142 -0.1030 -0.0305 0.9805 1.0832
#> cond~~cond         0.2484 0.0320 100  0.1786  0.1829  0.1896 0.3127 0.3251
#> reaction~~reaction 1.8359 0.2125 100  1.1411  1.2106  1.3734 2.1683 2.2603
#> pmi~~pmi           1.7102 0.2222 100  1.1476  1.1708  1.2561 2.0878 2.2165
#> indirect           0.2257 0.1526 100 -0.0587 -0.0587 -0.0182 0.4992 0.6002
#> direct             0.3035 0.2610 100 -0.4462 -0.4252 -0.3057 0.7598 0.9100
#> total              0.5292 0.3009 100 -0.3454 -0.2608 -0.1030 1.1050 1.2458
#>                    99.95%
#> cp                 0.9428
#> b                  0.7535
#> a                  1.0884
#> cond~~cond         0.3324
#> reaction~~reaction 2.3266
#> pmi~~pmi           2.3246
#> indirect           0.6674
#> direct             0.9428
#> total              1.2520
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1268 0.0845 100 -0.1515 -0.1377 -0.0996 0.2423 0.2704
#> b                  0.5034 0.0789 100  0.2068  0.2196  0.3008 0.5980 0.6272
#> a                  0.1803 0.1041 100 -0.0480 -0.0417 -0.0107 0.3611 0.3747
#> cond~~cond         1.0000 0.0000 100  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7074 0.0735 100  0.5856  0.5867  0.6102 0.8835 0.8910
#> pmi~~pmi           0.9675 0.0362 100  0.8579  0.8596  0.8695 1.0000 1.0000
#> indirect           0.0908 0.0478 100 -0.0213 -0.0204 -0.0053 0.1603 0.1733
#> direct             0.1268 0.0845 100 -0.1515 -0.1377 -0.0996 0.2423 0.2704
#> total              0.2176 0.0953 100 -0.1178 -0.0883 -0.0321 0.3480 0.3681
#>                    99.95%
#> cp                 0.2776
#> b                  0.6350
#> a                  0.3769
#> cond~~cond         1.0000
#> reaction~~reaction 0.8921
#> pmi~~pmi           1.0000
#> indirect           0.1796
#> direct             0.2776
#> total              0.3685