Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.3688 0.3046 20 -0.3941 -0.3595 -0.2057 0.9389 1.0066
#> b                  0.4464 0.1220 20  0.1912  0.1940  0.2061 0.6254 0.6297
#> a                  0.3277 0.2270 20 -0.1278 -0.1100 -0.0310 0.6983 0.7032
#> cond~~cond         0.2470 0.0281 20  0.1891  0.1899  0.1935 0.2819 0.2827
#> reaction~~reaction 2.0459 0.2625 20  1.4725  1.4888  1.5609 2.4941 2.5233
#> pmi~~pmi           1.4702 0.2477 20  1.0153  1.0288  1.0888 1.9373 2.0274
#> reaction~1         0.7701 0.8068 20 -0.7438 -0.7264 -0.6489 2.2365 2.3268
#> pmi~1              5.5349 0.1765 20  5.1686  5.1784  5.2221 5.8087 5.8237
#> cond~1             0.4490 0.0486 20  0.3573  0.3575  0.3583 0.5246 0.5265
#> indirect           0.1463 0.1269 20 -0.0615 -0.0549 -0.0256 0.4290 0.4294
#> direct             0.3688 0.3046 20 -0.3941 -0.3595 -0.2057 0.9389 1.0066
#> total              0.5151 0.3266 20 -0.1580 -0.1294 -0.0019 1.2223 1.4060
#>                    99.95%
#> cp                 1.0219
#> b                  0.6307
#> a                  0.7043
#> cond~~cond         0.2829
#> reaction~~reaction 2.5298
#> pmi~~pmi           2.0477
#> reaction~1         2.3471
#> pmi~1              5.8271
#> cond~1             0.5270
#> indirect           0.4294
#> direct             1.0219
#> total              1.4474
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1182 0.0960 20 -0.1267 -0.1156 -0.0661 0.3008 0.3072
#> b                  0.3522 0.1018 20  0.1185  0.1239  0.1481 0.5043 0.5354
#> a                  0.1331 0.0866 20 -0.0478 -0.0403 -0.0070 0.2603 0.2642
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.8509 0.0878 20  0.6316  0.6343  0.6465 0.9584 0.9644
#> pmi~~pmi           0.9823 0.0239 20  0.9298  0.9302  0.9322 0.9981 0.9984
#> indirect           0.4967 0.0384 20 -0.0193 -0.0173 -0.0081 0.1283 0.1328
#> direct             4.5242 0.0960 20 -0.1267 -0.1156 -0.0661 0.3008 0.3072
#> total              0.9035 0.0998 20 -0.0508 -0.0418 -0.0015 0.3854 0.4083
#>                    99.95%
#> cp                 0.3087
#> b                  0.5424
#> a                  0.2650
#> cond~~cond         1.0000
#> reaction~~reaction 0.9657
#> pmi~~pmi           0.9985
#> indirect           0.1338
#> direct             0.3087
#> total              0.4134

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.4666 0.3442 20 -0.3634 -0.3328 -0.1967 1.0025 1.0026
#> b                  0.4308 0.1002 20  0.2548  0.2606  0.2866 0.6075 0.6406
#> a                  0.2667 0.2041 20 -0.0555 -0.0454 -0.0004 0.7294 0.8052
#> cond~~cond         0.2472 0.0191 20  0.2109  0.2119  0.2162 0.2736 0.2738
#> reaction~~reaction 2.0534 0.2863 20  1.6499  1.6525  1.6636 2.5271 2.6477
#> pmi~~pmi           1.4938 0.1886 20  1.2209  1.2213  1.2229 1.8297 1.8536
#> indirect           0.1189 0.1242 20 -0.0192 -0.0152  0.0029 0.4254 0.5129
#> direct             0.4666 0.3442 20 -0.3634 -0.3328 -0.1967 1.0025 1.0026
#> total              0.5854 0.3133 20  0.0993  0.1002  0.1042 1.1708 1.1786
#>                    99.95%
#> cp                 1.0026
#> b                  0.6481
#> a                  0.8223
#> cond~~cond         0.2739
#> reaction~~reaction 2.6748
#> pmi~~pmi           1.8590
#> indirect           0.5326
#> direct             1.0026
#> total              1.1804
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1052 0.1052 20 -0.1165 -0.1067 -0.0631 0.2841 0.2894
#> b                  0.3802 0.0809 20  0.2069  0.2087  0.2168 0.4902 0.5163
#> a                  0.1011 0.0839 20 -0.0223 -0.0184 -0.0009 0.2988 0.3220
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.8363 0.0627 20  0.7455  0.7462  0.7491 0.9415 0.9418
#> pmi~~pmi           0.9898 0.0272 20  0.8929  0.8960  0.9098 0.9995 0.9995
#> indirect           0.0385 0.0406 20 -0.0064 -0.0051  0.0007 0.1403 0.1652
#> direct             0.1052 0.1052 20 -0.1165 -0.1067 -0.0631 0.2841 0.2894
#> total              0.1437 0.0942 20  0.0304  0.0307  0.0322 0.3303 0.3351
#>                    99.95%
#> cp                 0.2906
#> b                  0.5222
#> a                  0.3272
#> cond~~cond         1.0000
#> reaction~~reaction 0.9419
#> pmi~~pmi           0.9995
#> indirect           0.1708
#> direct             0.2906
#> total              0.3363