Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1385 0.2744 100 -0.4258 -0.4180 -0.3787 0.6582 0.9123
#> b                  0.4923 0.0915 100  0.2615  0.2672  0.3209 0.6546 0.6982
#> a                  0.5307 0.2409 100  0.0521  0.0631  0.0845 1.0432 1.1676
#> cond~~cond         0.2469 0.0355 100  0.1488  0.1538  0.1729 0.2970 0.3261
#> reaction~~reaction 1.8260 0.2787 100  1.2366  1.2516  1.3237 2.3821 2.5939
#> pmi~~pmi           1.6592 0.2158 100  1.1996  1.2085  1.2689 2.1160 2.2420
#> reaction~1         0.6123 0.5448 100 -0.6097 -0.5276 -0.3976 1.7177 1.9910
#> pmi~1              5.4107 0.1476 100  5.0073  5.0542  5.1641 5.7363 5.8153
#> cond~1             0.4519 0.0516 100  0.3136  0.3321  0.3642 0.5482 0.5869
#> indirect           0.2613 0.1298 100  0.0328  0.0345  0.0396 0.5125 0.6193
#> direct             0.1385 0.2744 100 -0.4258 -0.4180 -0.3787 0.6582 0.9123
#> total              0.3998 0.3055 100 -0.2923 -0.2776 -0.2185 1.0004 1.2959
#>                    99.95%
#> cp                 0.9410
#> b                  0.7020
#> a                  1.2041
#> cond~~cond         0.3387
#> reaction~~reaction 2.6626
#> pmi~~pmi           2.2717
#> reaction~1         2.0107
#> pmi~1              5.8411
#> cond~1             0.5896
#> indirect           0.6207
#> direct             0.9410
#> total              1.4262
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0457 0.0864 100 -0.1611 -0.1522 -0.1220 0.1980 0.2529
#> b                  0.4298 0.0787 100  0.2339  0.2370  0.2591 0.5577 0.5835
#> a                  0.2005 0.0849 100  0.0215  0.0271  0.0339 0.3799 0.4203
#> cond~~cond         1.0000 0.0000 100  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.8053 0.0680 100  0.6456  0.6458  0.6666 0.9247 0.9422
#> pmi~~pmi           0.9598 0.0356 100  0.8186  0.8233  0.8554 0.9989 0.9992
#> indirect           0.4066 0.0417 100  0.0107  0.0115  0.0128 0.1648 0.2024
#> direct             4.1152 0.0864 100 -0.1611 -0.1522 -0.1220 0.1980 0.2529
#> total              0.9095 0.0934 100 -0.1105 -0.1000 -0.0667 0.3104 0.3548
#>                    99.95%
#> cp                 0.2756
#> b                  0.5915
#> a                  0.4259
#> cond~~cond         1.0000
#> reaction~~reaction 0.9468
#> pmi~~pmi           0.9995
#> indirect           0.2105
#> direct             0.2756
#> total              0.3622

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1344 0.2782 100 -0.5369 -0.5267 -0.4571 0.6521 0.6996
#> b                  0.4827 0.1026 100  0.2402  0.2451  0.2778 0.6840 0.7224
#> a                  0.5398 0.2079 100  0.0720  0.1036  0.1798 0.9240 1.0281
#> cond~~cond         0.2477 0.0324 100  0.1710  0.1724  0.1893 0.3041 0.3243
#> reaction~~reaction 1.8482 0.2716 100  1.0695  1.1088  1.2635 2.3555 2.4240
#> pmi~~pmi           1.6078 0.1762 100  1.2346  1.2500  1.2829 1.9587 2.0354
#> indirect           0.2609 0.1180 100  0.0354  0.0492  0.0785 0.5179 0.5669
#> direct             0.1344 0.2782 100 -0.5369 -0.5267 -0.4571 0.6521 0.6996
#> total              0.3953 0.2894 100 -0.3216 -0.2925 -0.1787 0.8592 0.9869
#>                    99.95%
#> cp                 0.7097
#> b                  0.7234
#> a                  1.0581
#> cond~~cond         0.3243
#> reaction~~reaction 2.4438
#> pmi~~pmi           2.0500
#> indirect           0.5846
#> direct             0.7097
#> total              1.0075
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                        est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 -0.0093 0.0920 100 -0.2116 -0.1964 -0.1383 0.2062 0.2266
#> b                   0.4506 0.0856 100  0.2178  0.2232  0.2367 0.5805 0.5978
#> a                   0.1634 0.0746 100  0.0291  0.0392  0.0713 0.3570 0.3674
#> cond~~cond          1.0000 0.0000 100  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction  0.7982 0.0731 100  0.6162  0.6191  0.6402 0.9340 0.9422
#> pmi~~pmi            0.9733 0.0321 100  0.8594  0.8650  0.8726 0.9949 0.9983
#> indirect            0.0736 0.0381 100  0.0113  0.0161  0.0252 0.1724 0.1923
#> direct             -0.0093 0.0920 100 -0.2116 -0.1964 -0.1383 0.2062 0.2266
#> total               0.0644 0.0947 100 -0.1125 -0.1080 -0.0570 0.2764 0.3189
#>                    99.95%
#> cp                 0.2303
#> b                  0.6027
#> a                  0.3750
#> cond~~cond         1.0000
#> reaction~~reaction 0.9460
#> pmi~~pmi           0.9991
#> indirect           0.1991
#> direct             0.2303
#> total              0.3424