Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.3213 0.2765 100 -0.4889 -0.3492 -0.0544 1.0203 1.1169
#> b                  0.5163 0.0935 100  0.2819  0.2914  0.3192 0.6713 0.7041
#> a                  0.3897 0.2645 100 -0.2823 -0.2588 -0.1368 0.9179 1.0712
#> cond~~cond         0.2473 0.0302 100  0.1672  0.1672  0.1833 0.3058 0.3162
#> reaction~~reaction 1.8309 0.2559 100  1.0850  1.1716  1.3200 2.3284 2.5353
#> pmi~~pmi           1.7248 0.2162 100  1.0191  1.1363  1.3203 2.0870 2.2186
#> reaction~1         0.4368 0.5450 100 -0.6480 -0.6041 -0.4594 1.6049 1.7474
#> pmi~1              5.3712 0.1790 100  4.9908  4.9911  5.0075 5.6827 5.7302
#> cond~1             0.4539 0.0468 100  0.3529  0.3546  0.3611 0.5411 0.5677
#> indirect           0.2012 0.1333 100 -0.1426 -0.1250 -0.0737 0.4187 0.5370
#> direct             0.3213 0.2765 100 -0.4889 -0.3492 -0.0544 1.0203 1.1169
#> total              0.5224 0.3128 100 -0.1992 -0.1504 -0.0013 1.2635 1.5158
#>                    99.95%
#> cp                 1.1910
#> b                  0.7278
#> a                  1.0845
#> cond~~cond         0.3215
#> reaction~~reaction 2.5511
#> pmi~~pmi           2.2220
#> reaction~1         1.7524
#> pmi~1              5.7418
#> cond~1             0.5722
#> indirect           0.6103
#> direct             1.1910
#> total              1.5419
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1040 0.0876 100 -0.1656 -0.1171 -0.0177 0.3264 0.3554
#> b                  0.4463 0.0750 100  0.2647  0.2680  0.2746 0.5682 0.6041
#> a                  0.1460 0.0957 100 -0.1072 -0.1000 -0.0516 0.3282 0.3671
#> cond~~cond         1.0000 0.0000 100  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7764 0.0658 100  0.5982  0.6069  0.6202 0.8934 0.9081
#> pmi~~pmi           0.9787 0.0292 100  0.8599  0.8652  0.8923 0.9999 1.0000
#> indirect           0.2844 0.0417 100 -0.0471 -0.0417 -0.0244 0.1405 0.1548
#> direct             4.0460 0.0876 100 -0.1656 -0.1171 -0.0177 0.3264 0.3554
#> total              0.9129 0.0962 100 -0.0658 -0.0501  0.0000 0.3972 0.4167
#>                    99.95%
#> cp                 0.3678
#> b                  0.6316
#> a                  0.3742
#> cond~~cond         1.0000
#> reaction~~reaction 0.9120
#> pmi~~pmi           1.0000
#> indirect           0.1615
#> direct             0.3678
#> total              0.4223

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.4295 0.2744 100 -0.1896 -0.1773 -0.1387 0.9190 1.0215
#> b                  0.4593 0.1041 100  0.2286  0.2301  0.2507 0.6625 0.6879
#> a                  0.4171 0.2439 100 -0.1780 -0.1432 -0.0775 0.8378 1.0096
#> cond~~cond         0.2473 0.0334 100  0.1717  0.1735  0.1801 0.3182 0.3247
#> reaction~~reaction 1.9132 0.2486 100  1.4408  1.4441  1.4627 2.3876 2.5086
#> pmi~~pmi           1.7071 0.2304 100  1.1005  1.1686  1.2662 2.1833 2.3766
#> indirect           0.1909 0.1238 100 -0.0727 -0.0676 -0.0381 0.4642 0.5123
#> direct             0.4295 0.2744 100 -0.1896 -0.1773 -0.1387 0.9190 1.0215
#> total              0.6205 0.3010 100 -0.0381 -0.0314  0.0452 1.1637 1.3889
#>                    99.95%
#> cp                 1.0924
#> b                  0.6953
#> a                  1.0352
#> cond~~cond         0.3257
#> reaction~~reaction 2.5647
#> pmi~~pmi           2.4421
#> indirect           0.5313
#> direct             1.0924
#> total              1.3925
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1063 0.0867 100 -0.0590 -0.0562 -0.0432 0.2996 0.3291
#> b                  0.4572 0.0833 100  0.1994  0.2016  0.2086 0.5431 0.5619
#> a                  0.1191 0.0899 100 -0.0690 -0.0564 -0.0308 0.3321 0.3476
#> cond~~cond         1.0000 0.0000 100  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7680 0.0672 100  0.5755  0.6133  0.6688 0.9075 0.9232
#> pmi~~pmi           0.9858 0.0291 100  0.8763  0.8791  0.8897 0.9997 0.9999
#> indirect           0.0545 0.0382 100 -0.0250 -0.0221 -0.0127 0.1329 0.1658
#> direct             0.1063 0.0867 100 -0.0590 -0.0562 -0.0432 0.2996 0.3291
#> total              0.1608 0.0939 100 -0.0128 -0.0103  0.0146 0.3561 0.4399
#>                    99.95%
#> cp                 0.3516
#> b                  0.5654
#> a                  0.3517
#> cond~~cond         1.0000
#> reaction~~reaction 0.9325
#> pmi~~pmi           0.9999
#> indirect           0.1661
#> direct             0.3516
#> total              0.4475