Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.2825 0.3620 20 -0.4695 -0.4486 -0.3557 0.9156 1.0697
#> b                  0.4754 0.1082 20  0.2559  0.2587  0.2709 0.6045 0.6055
#> a                  0.4495 0.3159 20 -0.0708 -0.0625 -0.0258 1.0368 1.2086
#> cond~~cond         0.2478 0.0294 20  0.2038  0.2038  0.2038 0.3036 0.3183
#> reaction~~reaction 1.9670 0.2607 20  1.5444  1.5454  1.5495 2.4583 2.5668
#> pmi~~pmi           1.7339 0.1951 20  1.4410  1.4437  1.4558 2.0406 2.0769
#> reaction~1         0.6477 0.5737 20 -0.3307 -0.3063 -0.1976 1.6758 1.7645
#> pmi~1              5.4077 0.1832 20  4.9213  4.9467  5.0598 5.7066 5.7249
#> cond~1             0.4590 0.0505 20  0.3628  0.3644  0.3719 0.5410 0.5502
#> indirect           0.2137 0.1816 20 -0.0336 -0.0297 -0.0124 0.5888 0.7056
#> direct             0.2825 0.3620 20 -0.4695 -0.4486 -0.3557 0.9156 1.0697
#> total              0.4962 0.4057 20 -0.3199 -0.2989 -0.2058 1.2129 1.2488
#>                    99.95%
#> cp                 1.1043
#> b                  0.6057
#> a                  1.2472
#> cond~~cond         0.3215
#> reaction~~reaction 2.5912
#> pmi~~pmi           2.0851
#> reaction~1         1.7845
#> pmi~1              5.7290
#> cond~1             0.5523
#> indirect           0.7319
#> direct             1.1043
#> total              1.2568
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0904 0.1126 20 -0.1344 -0.1293 -0.1062 0.2802 0.3099
#> b                  0.4082 0.0904 20  0.2414  0.2428  0.2491 0.5386 0.5442
#> a                  0.1675 0.1112 20 -0.0231 -0.0203 -0.0081 0.3712 0.4137
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.8129 0.0712 20  0.6630  0.6637  0.6667 0.8953 0.9023
#> pmi~~pmi           0.9719 0.0440 20  0.8207  0.8278  0.8591 0.9999 0.9999
#> indirect           0.4163 0.0543 20 -0.0099 -0.0087 -0.0035 0.1766 0.2051
#> direct             4.0487 0.1126 20 -0.1344 -0.1293 -0.1062 0.2802 0.3099
#> total              0.9222 0.1239 20 -0.0916 -0.0856 -0.0592 0.3489 0.3604
#>                    99.95%
#> cp                 0.3165
#> b                  0.5455
#> a                  0.4232
#> cond~~cond         1.0000
#> reaction~~reaction 0.9038
#> pmi~~pmi           0.9999
#> indirect           0.2115
#> direct             0.3165
#> total              0.3629

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.2451 0.2509 20 -0.2693 -0.2617 -0.2278 0.6696 0.7150
#> b                  0.4872 0.1073 20  0.3013  0.3037  0.3141 0.6635 0.7011
#> a                  0.4517 0.2002 20  0.1760  0.1798  0.1969 0.7641 0.7810
#> cond~~cond         0.2479 0.0320 20  0.1977  0.1980  0.1996 0.3121 0.3217
#> reaction~~reaction 1.9574 0.2966 20  1.3922  1.4161  1.5223 2.5417 2.7537
#> pmi~~pmi           1.7069 0.2057 20  1.1507  1.1673  1.2410 1.9842 1.9979
#> indirect           0.2196 0.0929 20  0.0964  0.0966  0.0979 0.3797 0.3879
#> direct             0.2451 0.2509 20 -0.2693 -0.2617 -0.2278 0.6696 0.7150
#> total              0.4647 0.2318 20  0.0051  0.0131  0.0488 0.8504 0.8518
#>                    99.95%
#> cp                 0.7252
#> b                  0.7096
#> a                  0.7849
#> cond~~cond         0.3239
#> reaction~~reaction 2.8015
#> pmi~~pmi           2.0009
#> indirect           0.3898
#> direct             0.7252
#> total              0.8522
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0896 0.0839 20 -0.0898 -0.0872 -0.0759 0.2269 0.2486
#> b                  0.3999 0.0841 20  0.2632  0.2642  0.2691 0.5275 0.5311
#> a                  0.1786 0.0698 20  0.0762  0.0764  0.0768 0.2668 0.2688
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.8192 0.0607 20  0.7171  0.7189  0.7272 0.9237 0.9324
#> pmi~~pmi           0.9681 0.0244 20  0.9275  0.9278  0.9288 0.9941 0.9942
#> indirect           0.0714 0.0284 20  0.0336  0.0337  0.0344 0.1255 0.1274
#> direct             0.0896 0.0839 20 -0.0898 -0.0872 -0.0759 0.2269 0.2486
#> total              0.1610 0.0752 20  0.0017  0.0044  0.0163 0.2871 0.2948
#>                    99.95%
#> cp                 0.2535
#> b                  0.5320
#> a                  0.2692
#> cond~~cond         1.0000
#> reaction~~reaction 0.9344
#> pmi~~pmi           0.9942
#> indirect           0.1278
#> direct             0.2535
#> total              0.2965