Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.4518 0.3378 20 -0.0774 -0.0765 -0.0726 0.9661 0.9782
#> b                  0.5243 0.1044 20  0.2790  0.2801  0.2850 0.6714 0.6970
#> a                  0.4679 0.2861 20 -0.1084 -0.0998 -0.0617 0.8859 0.9045
#> cond~~cond         0.2500 0.0333 20  0.2072  0.2077  0.2096 0.3005 0.3020
#> reaction~~reaction 1.7846 0.2384 20  1.3493  1.3592  1.4034 2.2371 2.3613
#> pmi~~pmi           1.7411 0.3232 20  1.1517  1.1548  1.1685 2.2711 2.3302
#> reaction~1         0.2930 0.5311 20 -0.6489 -0.6445 -0.6250 1.3876 1.4800
#> pmi~1              5.3438 0.1862 20  4.9613  4.9743  5.0322 5.6493 5.6587
#> cond~1             0.4905 0.0482 20  0.3881  0.3904  0.4007 0.5740 0.5883
#> indirect           0.2453 0.1486 20 -0.0590 -0.0544 -0.0337 0.4796 0.5023
#> direct             0.4518 0.3378 20 -0.0774 -0.0765 -0.0726 0.9661 0.9782
#> total              0.6971 0.3400 20  0.0212  0.0279  0.0580 1.1642 1.1867
#>                    99.95%
#> cp                 0.9809
#> b                  0.7027
#> a                  0.9087
#> cond~~cond         0.3023
#> reaction~~reaction 2.3892
#> pmi~~pmi           2.3436
#> reaction~1         1.5007
#> pmi~1              5.6609
#> cond~1             0.5915
#> indirect           0.5073
#> direct             0.9809
#> total              1.1917
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1463 0.1160 20 -0.0250 -0.0246 -0.0232 0.3330 0.3393
#> b                  0.4550 0.0791 20  0.2407  0.2479  0.2801 0.5749 0.5935
#> a                  0.1746 0.1107 20 -0.0433 -0.0398 -0.0245 0.3378 0.3453
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7483 0.0511 20  0.6261  0.6267  0.6293 0.8017 0.8025
#> pmi~~pmi           0.9695 0.0372 20  0.8796  0.8808  0.8858 0.9999 1.0000
#> indirect           0.1897 0.0480 20 -0.0196 -0.0180 -0.0111 0.1533 0.1591
#> direct             3.9877 0.1160 20 -0.0250 -0.0246 -0.0232 0.3330 0.3393
#> total              0.9811 0.1187 20  0.0067  0.0089  0.0187 0.4085 0.4194
#>                    99.95%
#> cp                 0.3408
#> b                  0.5977
#> a                  0.3469
#> cond~~cond         1.0000
#> reaction~~reaction 0.8026
#> pmi~~pmi           1.0000
#> indirect           0.1604
#> direct             0.3408
#> total              0.4219

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se  R  0.05%   0.5%   2.5%  97.5%  99.5% 99.95%
#> cp                 0.4429 0.1486 20 0.0625 0.0776 0.1443 0.6962 0.7195 0.7247
#> b                  0.5361 0.1005 20 0.1716 0.1927 0.2863 0.6218 0.6256 0.6265
#> a                  0.4287 0.2082 20 0.0919 0.0942 0.1044 0.7687 0.8828 0.9085
#> cond~~cond         0.2495 0.0268 20 0.1843 0.1858 0.1925 0.2831 0.2876 0.2887
#> reaction~~reaction 1.7859 0.2290 20 1.3090 1.3211 1.3752 2.1141 2.1358 2.1407
#> pmi~~pmi           1.7413 0.2490 20 1.2419 1.2538 1.3065 2.1711 2.2276 2.2404
#> indirect           0.2293 0.1025 20 0.0545 0.0552 0.0584 0.3678 0.3967 0.4033
#> direct             0.4429 0.1486 20 0.0625 0.0776 0.1443 0.6962 0.7195 0.7247
#> total              0.6722 0.1699 20 0.3588 0.3627 0.3802 0.9500 0.9834 0.9909
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R  0.05%   0.5%   2.5%  97.5%  99.5% 99.95%
#> cp                 0.1809 0.0462 20 0.0192 0.0243 0.0468 0.2139 0.2171 0.2178
#> b                  0.4067 0.0917 20 0.1558 0.1710 0.2387 0.5475 0.5553 0.5570
#> a                  0.2720 0.0726 20 0.0350 0.0358 0.0396 0.2712 0.3258 0.3380
#> cond~~cond         1.0000 0.0000 20 1.0000 1.0000 1.0000 1.0000 1.0000 1.0000
#> reaction~~reaction 0.7619 0.0685 20 0.6676 0.6676 0.6678 0.8915 0.9135 0.9185
#> pmi~~pmi           0.9260 0.0253 20 0.8855 0.8921 0.9213 0.9984 0.9987 0.9988
#> indirect           0.1106 0.0319 20 0.0195 0.0195 0.0199 0.1161 0.1321 0.1357
#> direct             0.1809 0.0462 20 0.0192 0.0243 0.0468 0.2139 0.2171 0.2178
#> total              0.2915 0.0492 20 0.1097 0.1118 0.1209 0.2802 0.2836 0.2843