Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.4738 0.3783 20 -0.2369 -0.2225 -0.1582 1.0504 1.1183
#> b                  0.5043 0.0907 20  0.3277  0.3318  0.3500 0.6303 0.6343
#> a                  0.4113 0.2928 20 -0.1971 -0.1854 -0.1331 0.8306 0.9044
#> cond~~cond         0.2465 0.0261 20  0.2019  0.2028  0.2071 0.2923 0.3009
#> reaction~~reaction 1.8012 0.2262 20  1.1146  1.1386  1.2453 2.0254 2.0673
#> pmi~~pmi           1.6987 0.1747 20  1.2952  1.3069  1.3591 1.9475 1.9524
#> reaction~1         0.5628 0.4250 20 -0.1529 -0.1449 -0.1094 1.2211 1.2306
#> pmi~1              5.3900 0.1826 20  4.9856  5.0018  5.0735 5.7292 5.7642
#> cond~1             0.4478 0.0523 20  0.3465  0.3482  0.3555 0.5303 0.5424
#> indirect           0.2074 0.1425 20 -0.0775 -0.0734 -0.0551 0.4164 0.4476
#> direct             0.4738 0.3783 20 -0.2369 -0.2225 -0.1582 1.0504 1.1183
#> total              0.6812 0.3876 20 -0.1512 -0.1207  0.0146 1.2612 1.3622
#>                    99.95%
#> cp                 1.1336
#> b                  0.6352
#> a                  0.9210
#> cond~~cond         0.3028
#> reaction~~reaction 2.0767
#> pmi~~pmi           1.9535
#> reaction~1         1.2327
#> pmi~1              5.7721
#> cond~1             0.5452
#> indirect           0.4546
#> direct             1.1336
#> total              1.3849
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1535 0.1226 20 -0.0721 -0.0681 -0.0505 0.3431 0.3546
#> b                  0.4342 0.0695 20  0.3135  0.3136  0.3139 0.5139 0.5200
#> a                  0.1548 0.1094 20 -0.0735 -0.0693 -0.0507 0.3089 0.3208
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7673 0.0711 20  0.5461  0.5594  0.6183 0.8728 0.8790
#> pmi~~pmi           0.9760 0.0294 20  0.8953  0.8970  0.9043 0.9997 0.9999
#> indirect           0.3673 0.0479 20 -0.0244 -0.0233 -0.0182 0.1361 0.1551
#> direct             4.0857 0.1226 20 -0.0721 -0.0681 -0.0505 0.3431 0.3546
#> total              0.9019 0.1288 20 -0.0459 -0.0366  0.0047 0.4406 0.4777
#>                    99.95%
#> cp                 0.3572
#> b                  0.5213
#> a                  0.3235
#> cond~~cond         1.0000
#> reaction~~reaction 0.8804
#> pmi~~pmi           1.0000
#> indirect           0.1594
#> direct             0.3572
#> total              0.4861

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.4506 0.2452 20 -0.0226 -0.0178  0.0033 0.8026 0.8157
#> b                  0.5086 0.0778 20  0.4128  0.4154  0.4266 0.6832 0.7193
#> a                  0.3448 0.2925 20 -0.2443 -0.2369 -0.2041 0.7437 0.7568
#> cond~~cond         0.2457 0.0316 20  0.1701  0.1720  0.1807 0.2938 0.3008
#> reaction~~reaction 1.8086 0.2812 20  0.9287  0.9728  1.1689 2.1268 2.1400
#> pmi~~pmi           1.6953 0.2353 20  1.3285  1.3346  1.3618 2.1068 2.1388
#> indirect           0.1749 0.1681 20 -0.1157 -0.1118 -0.0943 0.4668 0.4784
#> direct             0.4506 0.2452 20 -0.0226 -0.0178  0.0033 0.8026 0.8157
#> total              0.6255 0.3633 20 -0.0370 -0.0301  0.0005 1.1881 1.2534
#>                    99.95%
#> cp                 0.8186
#> b                  0.7274
#> a                  0.7597
#> cond~~cond         0.3023
#> reaction~~reaction 2.1430
#> pmi~~pmi           2.1460
#> indirect           0.4810
#> direct             0.8186
#> total              1.2680
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1566 0.0808 20 -0.0075 -0.0059  0.0014 0.2721 0.3021
#> b                  0.4471 0.0556 20  0.3324  0.3368  0.3562 0.5633 0.5746
#> a                  0.1192 0.1112 20 -0.0974 -0.0937 -0.0770 0.2918 0.2947
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7589 0.0754 20  0.5775  0.5788  0.5843 0.8457 0.8528
#> pmi~~pmi           0.9858 0.0270 20  0.9128  0.9132  0.9148 0.9998 0.9998
#> indirect           0.0533 0.0539 20 -0.0374 -0.0363 -0.0313 0.1452 0.1585
#> direct             0.1566 0.0808 20 -0.0075 -0.0059  0.0014 0.2721 0.3021
#> total              0.2099 0.1187 20 -0.0129 -0.0106 -0.0004 0.3973 0.4261
#>                    99.95%
#> cp                 0.3089
#> b                  0.5771
#> a                  0.2953
#> cond~~cond         1.0000
#> reaction~~reaction 0.8544
#> pmi~~pmi           0.9999
#> indirect           0.1615
#> direct             0.3089
#> total              0.4326