Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.4724 0.2838 20  0.0450  0.0473  0.0575 0.9832 1.0182
#> b                  0.4660 0.1109 20  0.2054  0.2115  0.2389 0.6045 0.6202
#> a                  0.4421 0.2719 20  0.0404  0.0478  0.0810 0.9731 1.0353
#> cond~~cond         0.2476 0.0382 20  0.1678  0.1699  0.1791 0.3027 0.3074
#> reaction~~reaction 1.9508 0.2706 20  1.5825  1.5860  1.6018 2.4109 2.4111
#> pmi~~pmi           1.7279 0.2304 20  1.3161  1.3238  1.3580 2.1740 2.2879
#> reaction~1         0.6105 0.6250 20 -0.3122 -0.2828 -0.1522 1.8605 2.0973
#> pmi~1              5.3990 0.1875 20  4.9883  4.9970  5.0355 5.7311 5.7939
#> cond~1             0.4591 0.0500 20  0.3811  0.3820  0.3856 0.5425 0.5504
#> indirect           0.2060 0.1240 20  0.0175  0.0208  0.0357 0.4562 0.5283
#> direct             0.4724 0.2838 20  0.0450  0.0473  0.0575 0.9832 1.0182
#> total              0.6784 0.3044 20  0.2771  0.2797  0.2912 1.2654 1.3068
#>                    99.95%
#> cp                 1.0260
#> b                  0.6237
#> a                  1.0493
#> cond~~cond         0.3085
#> reaction~~reaction 2.4111
#> pmi~~pmi           2.3135
#> reaction~1         2.1506
#> pmi~1              5.8080
#> cond~1             0.5522
#> indirect           0.5445
#> direct             1.0260
#> total              1.3161
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R  0.05%   0.5%   2.5%  97.5%  99.5% 99.95%
#> cp                 0.1505 0.0951 20 0.0115 0.0123 0.0161 0.3297 0.3535 0.3589
#> b                  0.3976 0.0889 20 0.2021 0.2037 0.2104 0.5225 0.5249 0.5255
#> a                  0.1651 0.0987 20 0.0148 0.0174 0.0292 0.3624 0.4032 0.4124
#> cond~~cond         1.0000 0.0000 20 1.0000 1.0000 1.0000 1.0000 1.0000 1.0000
#> reaction~~reaction 0.7995 0.0656 20 0.6550 0.6588 0.6754 0.8851 0.8946 0.8967
#> pmi~~pmi           0.9728 0.0411 20 0.8298 0.8364 0.8658 0.9989 0.9996 0.9998
#> indirect           0.3908 0.0389 20 0.0056 0.0067 0.0115 0.1470 0.1734 0.1793
#> direct             4.0509 0.0951 20 0.0115 0.0123 0.0161 0.3297 0.3535 0.3589
#> total              0.9226 0.1025 20 0.0707 0.0725 0.0804 0.4182 0.4308 0.4336

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.4961 0.2489 20 -0.0574 -0.0402  0.0364 0.8371 0.8512
#> b                  0.4630 0.1175 20  0.2266  0.2319  0.2551 0.6921 0.7362
#> a                  0.4150 0.2723 20 -0.1062 -0.0976 -0.0598 0.8670 0.9102
#> cond~~cond         0.2485 0.0337 20  0.1732  0.1752  0.1838 0.2973 0.3064
#> reaction~~reaction 2.0031 0.2800 20  1.4926  1.5033  1.5509 2.3784 2.3994
#> pmi~~pmi           1.7070 0.1621 20  1.3628  1.3686  1.3944 1.9829 2.0293
#> indirect           0.1925 0.1237 20 -0.0623 -0.0572 -0.0344 0.3885 0.4176
#> direct             0.4961 0.2489 20 -0.0574 -0.0402  0.0364 0.8371 0.8512
#> total              0.6885 0.3150 20 -0.0607 -0.0455  0.0218 1.0054 1.0265
#>                    99.95%
#> cp                 0.8543
#> b                  0.7461
#> a                  0.9199
#> cond~~cond         0.3084
#> reaction~~reaction 2.4041
#> pmi~~pmi           2.0397
#> indirect           0.4242
#> direct             0.8543
#> total              1.0312
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1523 0.0776 20 -0.0187 -0.0130  0.0124 0.2545 0.2545
#> b                  0.4274 0.0808 20  0.2254  0.2328  0.2654 0.5327 0.5469
#> a                  0.1623 0.1002 20 -0.0354 -0.0326 -0.0202 0.2990 0.3043
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7730 0.0729 20  0.6377  0.6410  0.6556 0.8982 0.9025
#> pmi~~pmi           0.9737 0.0310 20  0.9067  0.9074  0.9105 0.9994 0.9999
#> indirect           0.0694 0.0387 20 -0.0166 -0.0153 -0.0093 0.1176 0.1252
#> direct             0.1523 0.0776 20 -0.0187 -0.0130  0.0124 0.2545 0.2545
#> total              0.2217 0.0992 20 -0.0198 -0.0155  0.0039 0.3190 0.3335
#>                    99.95%
#> cp                 0.2546
#> b                  0.5501
#> a                  0.3055
#> cond~~cond         1.0000
#> reaction~~reaction 0.9035
#> pmi~~pmi           1.0000
#> indirect           0.1269
#> direct             0.2546
#> total              0.3368