Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.3486 0.2145 20 -0.1686 -0.1515 -0.0755 0.6113 0.6160
#> b                  0.4736 0.0849 20  0.3171  0.3207  0.3369 0.6071 0.6095
#> a                  0.3202 0.2998 20 -0.1424 -0.1382 -0.1192 0.8255 0.8606
#> cond~~cond         0.2501 0.0429 20  0.1773  0.1803  0.1936 0.3485 0.3659
#> reaction~~reaction 1.8843 0.2750 20  1.3812  1.3909  1.4340 2.3291 2.3997
#> pmi~~pmi           1.5977 0.2702 20  1.1120  1.1227  1.1705 2.0216 2.0414
#> reaction~1         0.6292 0.4976 20 -0.1336 -0.1209 -0.0644 1.5683 1.6582
#> pmi~1              5.4269 0.1973 20  5.0656  5.0712  5.0961 5.7276 5.7585
#> cond~1             0.4941 0.0443 20  0.4241  0.4267  0.4382 0.5901 0.5912
#> indirect           0.1517 0.1427 20 -0.0859 -0.0825 -0.0670 0.3990 0.4334
#> direct             0.3486 0.2145 20 -0.1686 -0.1515 -0.0755 0.6113 0.6160
#> total              0.5003 0.2479 20 -0.0696 -0.0573 -0.0029 0.8334 0.8779
#>                    99.95%
#> cp                 0.6171
#> b                  0.6100
#> a                  0.8686
#> cond~~cond         0.3698
#> reaction~~reaction 2.4156
#> pmi~~pmi           2.0459
#> reaction~1         1.6784
#> pmi~1              5.7654
#> cond~1             0.5915
#> indirect           0.4412
#> direct             0.6171
#> total              0.8879
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1148 0.0773 20 -0.0628 -0.0566 -0.0290 0.2310 0.2396
#> b                  0.3974 0.0803 20  0.2729  0.2749  0.2838 0.5320 0.5401
#> a                  0.1257 0.1210 20 -0.0605 -0.0591 -0.0528 0.3525 0.3530
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.8174 0.0668 20  0.6768  0.6775  0.6806 0.9002 0.9008
#> pmi~~pmi           0.9842 0.0375 20  0.8753  0.8754  0.8758 0.9999 1.0000
#> indirect           0.4144 0.0490 20 -0.0294 -0.0283 -0.0232 0.1388 0.1538
#> direct             4.2594 0.0773 20 -0.0628 -0.0566 -0.0290 0.2310 0.2396
#> total              0.9880 0.0884 20 -0.0260 -0.0218 -0.0031 0.2989 0.3112
#>                    99.95%
#> cp                 0.2415
#> b                  0.5419
#> a                  0.3531
#> cond~~cond         1.0000
#> reaction~~reaction 0.9010
#> pmi~~pmi           1.0000
#> indirect           0.1572
#> direct             0.2415
#> total              0.3139

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.3961 0.2958 20  0.0682  0.0786  0.1248 1.0825 1.1958
#> b                  0.4884 0.1300 20  0.2138  0.2201  0.2485 0.7207 0.7408
#> a                  0.2867 0.1758 20 -0.0368 -0.0358 -0.0312 0.5778 0.6493
#> cond~~cond         0.2499 0.0322 20  0.1807  0.1831  0.1940 0.3056 0.3209
#> reaction~~reaction 1.9480 0.2930 20  1.3417  1.3459  1.3643 2.2968 2.3245
#> pmi~~pmi           1.5687 0.2028 20  1.2310  1.2390  1.2748 1.9590 1.9740
#> indirect           0.1380 0.0752 20 -0.0165 -0.0159 -0.0133 0.2424 0.2430
#> direct             0.3961 0.2958 20  0.0682  0.0786  0.1248 1.0825 1.1958
#> total              0.5341 0.3186 20  0.1265  0.1359  0.1774 1.1917 1.2377
#>                    99.95%
#> cp                 1.2213
#> b                  0.7453
#> a                  0.6653
#> cond~~cond         0.3243
#> reaction~~reaction 2.3308
#> pmi~~pmi           1.9774
#> indirect           0.2432
#> direct             1.2213
#> total              1.2481
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1434 0.0936 20  0.0231  0.0257  0.0373 0.3447 0.3807
#> b                  0.3794 0.1021 20  0.1626  0.1670  0.1865 0.5438 0.5734
#> a                  0.1215 0.0643 20 -0.0147 -0.0143 -0.0122 0.2086 0.2113
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.8222 0.0715 20  0.6576  0.6589  0.6646 0.9109 0.9240
#> pmi~~pmi           0.9852 0.0136 20  0.9551  0.9554  0.9565 0.9999 1.0000
#> indirect           0.0461 0.0230 20 -0.0054 -0.0052 -0.0043 0.0751 0.0758
#> direct             0.1434 0.0936 20  0.0231  0.0257  0.0373 0.3447 0.3807
#> total              0.1895 0.0999 20  0.0430  0.0456  0.0574 0.3795 0.3941
#>                    99.95%
#> cp                 0.3888
#> b                  0.5801
#> a                  0.2119
#> cond~~cond         1.0000
#> reaction~~reaction 0.9270
#> pmi~~pmi           1.0000
#> indirect           0.0759
#> direct             0.3888
#> total              0.3973