Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.2494 0.2724 100 -0.2753 -0.2606 -0.2285 0.7609 0.8115
#> b                  0.5193 0.1169 100  0.1751  0.1883  0.2873 0.7104 0.8137
#> a                  0.3981 0.2721 100 -0.3398 -0.3195 -0.1701 0.8548 0.9509
#> cond~~cond         0.2497 0.0332 100  0.1568  0.1569  0.1779 0.3019 0.3098
#> reaction~~reaction 1.8828 0.2586 100  1.1910  1.2295  1.4316 2.5279 2.6467
#> pmi~~pmi           1.6324 0.2241 100  1.1429  1.1660  1.2373 2.1359 2.2633
#> reaction~1         0.4640 0.6407 100 -1.1791 -1.1294 -0.6064 1.8182 2.1146
#> pmi~1              5.3983 0.1665 100  4.9369  4.9821  5.0454 5.7050 5.7691
#> cond~1             0.4712 0.0522 100  0.3469  0.3546  0.3696 0.5623 0.5813
#> indirect           0.2068 0.1352 100 -0.1979 -0.1840 -0.1017 0.4244 0.4677
#> direct             0.2494 0.2724 100 -0.2753 -0.2606 -0.2285 0.7609 0.8115
#> total              0.4562 0.2850 100 -0.1015 -0.0852 -0.0534 0.9895 1.1130
#>                    99.95%
#> cp                 0.8331
#> b                  0.8374
#> a                  1.0159
#> cond~~cond         0.3108
#> reaction~~reaction 2.6525
#> pmi~~pmi           2.3228
#> reaction~1         2.3028
#> pmi~1              5.7746
#> cond~1             0.5885
#> indirect           0.4892
#> direct             0.8331
#> total              1.1632
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0809 0.0873 100 -0.0808 -0.0782 -0.0666 0.2415 0.2706
#> b                  0.4357 0.0861 100  0.1733  0.1756  0.2600 0.5495 0.6490
#> a                  0.1539 0.1009 100 -0.1360 -0.1190 -0.0625 0.3121 0.3443
#> cond~~cond         1.0000 0.0000 100  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7928 0.0706 100  0.5258  0.5749  0.6567 0.9172 0.9289
#> pmi~~pmi           0.9763 0.0295 100  0.8690  0.8811  0.9026 0.9998 0.9999
#> indirect           0.3011 0.0422 100 -0.0613 -0.0555 -0.0285 0.1305 0.1403
#> direct             4.1748 0.0873 100 -0.0808 -0.0782 -0.0666 0.2415 0.2706
#> total              0.9429 0.0915 100 -0.0331 -0.0295 -0.0146 0.3184 0.3427
#>                    99.95%
#> cp                 0.2861
#> b                  0.6927
#> a                  0.3619
#> cond~~cond         1.0000
#> reaction~~reaction 0.9335
#> pmi~~pmi           1.0000
#> indirect           0.1464
#> direct             0.2861
#> total              0.3468

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.2222 0.2314 100 -0.4018 -0.3468 -0.2311 0.6282 0.7370
#> b                  0.5125 0.1157 100  0.1590  0.1811  0.2784 0.7058 0.7154
#> a                  0.5189 0.2707 100 -0.1225 -0.0956  0.0221 1.0450 1.1199
#> cond~~cond         0.2489 0.0303 100  0.1715  0.1731  0.1856 0.3138 0.3224
#> reaction~~reaction 1.8516 0.2461 100  1.0811  1.2115  1.4438 2.2740 2.4890
#> pmi~~pmi           1.5992 0.2347 100  1.0873  1.1098  1.1730 2.0331 2.1668
#> indirect           0.2658 0.1578 100 -0.0733 -0.0548  0.0106 0.5861 0.6184
#> direct             0.2222 0.2314 100 -0.4018 -0.3468 -0.2311 0.6282 0.7370
#> total              0.4880 0.2809 100 -0.2845 -0.2421 -0.1130 1.0368 1.1473
#>                    99.95%
#> cp                 0.7434
#> b                  0.7180
#> a                  1.1403
#> cond~~cond         0.3260
#> reaction~~reaction 2.6206
#> pmi~~pmi           2.2420
#> indirect           0.6409
#> direct             0.7434
#> total              1.1737
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0435 0.0760 100 -0.1189 -0.1075 -0.0741 0.2125 0.2382
#> b                  0.3827 0.0914 100  0.1451  0.1710  0.2481 0.5808 0.5945
#> a                  0.2442 0.1043 100 -0.0497 -0.0406  0.0076 0.4004 0.4518
#> cond~~cond         1.0000 0.0000 100  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.8435 0.0758 100  0.6354  0.6362  0.6447 0.9349 0.9574
#> pmi~~pmi           0.9404 0.0454 100  0.7788  0.7954  0.8397 0.9987 0.9998
#> indirect           0.0935 0.0499 100 -0.0253 -0.0190  0.0037 0.1904 0.2013
#> direct             0.0435 0.0760 100 -0.1189 -0.1075 -0.0741 0.2125 0.2382
#> total              0.1369 0.0915 100 -0.0837 -0.0742 -0.0357 0.3336 0.3717
#>                    99.95%
#> cp                 0.2390
#> b                  0.5948
#> a                  0.4703
#> cond~~cond         1.0000
#> reaction~~reaction 0.9657
#> pmi~~pmi           1.0000
#> indirect           0.2101
#> direct             0.2390
#> total              0.3842