Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.2954 0.2533 100 -0.2695 -0.2325 -0.1154 0.9242 1.0203
#> b                  0.4536 0.1027 100  0.2340  0.2394  0.2583 0.6243 0.6934
#> a                  0.4926 0.2591 100 -0.2409 -0.1396  0.0325 1.0266 1.1469
#> cond~~cond         0.2493 0.0347 100  0.1575  0.1634  0.1804 0.3109 0.3346
#> reaction~~reaction 1.9118 0.2543 100  1.3260  1.4067  1.5514 2.4578 2.5407
#> pmi~~pmi           1.7760 0.2736 100  0.8798  1.0470  1.2991 2.3642 2.3929
#> reaction~1         0.7602 0.6059 100 -0.7066 -0.6130 -0.1524 1.9279 2.0694
#> pmi~1              5.3544 0.1721 100  4.9669  4.9695  5.0083 5.6409 5.7428
#> cond~1             0.4796 0.0450 100  0.3781  0.3806  0.3915 0.5531 0.5699
#> indirect           0.2234 0.1281 100 -0.0989 -0.0592  0.0120 0.4907 0.6420
#> direct             0.2954 0.2533 100 -0.2695 -0.2325 -0.1154 0.9242 1.0203
#> total              0.5189 0.2832 100 -0.0287  0.0076  0.0686 1.0695 1.2955
#>                    99.95%
#> cp                 1.0341
#> b                  0.7044
#> a                  1.1643
#> cond~~cond         0.3386
#> reaction~~reaction 2.5977
#> pmi~~pmi           2.4122
#> reaction~1         2.1309
#> pmi~1              5.7995
#> cond~1             0.5720
#> indirect           0.7145
#> direct             1.0341
#> total              1.3522
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0963 0.0792 100 -0.0928 -0.0744 -0.0377 0.2867 0.2983
#> b                  0.4015 0.0776 100  0.2165  0.2275  0.2455 0.5289 0.5519
#> a                  0.1815 0.0913 100 -0.0938 -0.0548  0.0105 0.3573 0.4166
#> cond~~cond         1.0000 0.0000 100  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.8155 0.0714 100  0.5921  0.6242  0.6698 0.9313 0.9417
#> pmi~~pmi           0.9671 0.0358 100  0.8244  0.8265  0.8722 0.9994 0.9999
#> indirect           0.4965 0.0395 100 -0.0315 -0.0190  0.0037 0.1529 0.1889
#> direct             3.9511 0.0792 100 -0.0928 -0.0744 -0.0377 0.2867 0.2983
#> total              0.9607 0.0876 100 -0.0090  0.0017  0.0246 0.3421 0.4031
#>                    99.95%
#> cp                 0.3040
#> b                  0.5627
#> a                  0.4190
#> cond~~cond         1.0000
#> reaction~~reaction 0.9419
#> pmi~~pmi           1.0000
#> indirect           0.2090
#> direct             0.3040
#> total              0.4052

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.2091 0.2636 100 -0.4132 -0.3696 -0.2955 0.6772 0.7366
#> b                  0.4531 0.1028 100  0.2395  0.2478  0.2659 0.6419 0.6825
#> a                  0.4934 0.2702 100 -0.0921 -0.0836 -0.0071 1.0071 1.1507
#> cond~~cond         0.2498 0.0333 100  0.1722  0.1822  0.1979 0.3201 0.3316
#> reaction~~reaction 1.9587 0.2603 100  1.2830  1.3166  1.4297 2.4202 2.4715
#> pmi~~pmi           1.7683 0.2459 100  1.2114  1.2216  1.3177 2.2488 2.3512
#> indirect           0.2233 0.1479 100 -0.0522 -0.0471 -0.0026 0.5663 0.7039
#> direct             0.2091 0.2636 100 -0.4132 -0.3696 -0.2955 0.6772 0.7366
#> total              0.4325 0.2461 100 -0.2032 -0.1418 -0.0520 0.8289 0.9195
#>                    99.95%
#> cp                 0.7689
#> b                  0.6856
#> a                  1.2592
#> cond~~cond         0.3362
#> reaction~~reaction 2.4790
#> pmi~~pmi           2.3974
#> indirect           0.7669
#> direct             0.7689
#> total              0.9349
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0775 0.0854 100 -0.1353 -0.1201 -0.0907 0.2164 0.2410
#> b                  0.4181 0.0821 100  0.2242  0.2246  0.2321 0.5480 0.5809
#> a                  0.1674 0.0955 100 -0.0372 -0.0348 -0.0029 0.3559 0.3896
#> cond~~cond         1.0000 0.0000 100  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.8083 0.0629 100  0.6566  0.6638  0.6771 0.9214 0.9352
#> pmi~~pmi           0.9720 0.0366 100  0.8431  0.8482  0.8734 0.9995 1.0000
#> indirect           0.0700 0.0458 100 -0.0164 -0.0153 -0.0010 0.1864 0.2087
#> direct             0.0775 0.0854 100 -0.1353 -0.1201 -0.0907 0.2164 0.2410
#> total              0.1475 0.0793 100 -0.0666 -0.0461 -0.0164 0.2696 0.2853
#>                    99.95%
#> cp                 0.2411
#> b                  0.5882
#> a                  0.3961
#> cond~~cond         1.0000
#> reaction~~reaction 0.9434
#> pmi~~pmi           1.0000
#> indirect           0.2095
#> direct             0.2411
#> total              0.2933