Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.4105 0.3083 20  0.0251  0.0261  0.0306 0.9119 0.9227
#> b                  0.5051 0.1091 20  0.2780  0.2887  0.3365 0.7473 0.8359
#> a                  0.4401 0.2148 20  0.0323  0.0523  0.1411 0.8883 0.9984
#> cond~~cond         0.2487 0.0347 20  0.1731  0.1750  0.1830 0.2982 0.3012
#> reaction~~reaction 1.7978 0.2357 20  1.1725  1.1932  1.2851 2.0731 2.0837
#> pmi~~pmi           1.6390 0.2615 20  1.1038  1.1101  1.1383 2.0326 2.0449
#> reaction~1         0.4768 0.6264 20 -1.3838 -1.2887 -0.8661 1.4354 1.8816
#> pmi~1              5.4249 0.1800 20  5.0158  5.0238  5.0595 5.6454 5.6657
#> cond~1             0.4646 0.0599 20  0.3489  0.3523  0.3677 0.5711 0.5812
#> indirect           0.2223 0.1161 20  0.0157  0.0234  0.0577 0.4472 0.4507
#> direct             0.4105 0.3083 20  0.0251  0.0261  0.0306 0.9119 0.9227
#> total              0.6328 0.2856 20  0.1790  0.1898  0.2374 1.1000 1.1172
#>                    99.95%
#> cp                 0.9251
#> b                  0.8558
#> a                  1.0232
#> cond~~cond         0.3019
#> reaction~~reaction 2.0861
#> pmi~~pmi           2.0476
#> reaction~1         1.9820
#> pmi~1              5.6702
#> cond~1             0.5835
#> indirect           0.4514
#> direct             0.9251
#> total              1.1210
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R  0.05%   0.5%   2.5%  97.5%  99.5% 99.95%
#> cp                 0.1345 0.0981 20 0.0092 0.0094 0.0102 0.2921 0.2937 0.2941
#> b                  0.4312 0.0770 20 0.2771 0.2817 0.3021 0.5885 0.6355 0.6461
#> a                  0.1690 0.0782 20 0.0137 0.0197 0.0467 0.3287 0.3384 0.3406
#> cond~~cond         1.0000 0.0000 20 1.0000 1.0000 1.0000 1.0000 1.0000 1.0000
#> reaction~~reaction 0.7764 0.0641 20 0.5765 0.5860 0.6278 0.8646 0.9096 0.9198
#> pmi~~pmi           0.9714 0.0298 20 0.8840 0.8854 0.8918 0.9966 0.9992 0.9998
#> indirect           0.3133 0.0377 20 0.0055 0.0083 0.0205 0.1424 0.1441 0.1445
#> direct             4.1765 0.0981 20 0.0092 0.0094 0.0102 0.2921 0.2937 0.2941
#> total              0.9317 0.0882 20 0.0637 0.0679 0.0861 0.3382 0.3390 0.3392

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.4469 0.2679 20  0.0243  0.0364  0.0900 1.0760 1.1610
#> b                  0.5192 0.0879 20  0.3390  0.3424  0.3576 0.6521 0.6844
#> a                  0.4290 0.2703 20 -0.1020 -0.1008 -0.0956 0.7814 0.8822
#> cond~~cond         0.2484 0.0227 20  0.2059  0.2067  0.2105 0.2832 0.2836
#> reaction~~reaction 1.7981 0.2381 20  1.4161  1.4175  1.4237 2.1859 2.2368
#> pmi~~pmi           1.6181 0.2782 20  1.1686  1.1781  1.2205 2.1207 2.1291
#> indirect           0.2197 0.1247 20 -0.0486 -0.0480 -0.0451 0.3568 0.3948
#> direct             0.4469 0.2679 20  0.0243  0.0364  0.0900 1.0760 1.1610
#> total              0.6666 0.3218 20  0.1870  0.1991  0.2530 1.4037 1.5500
#>                    99.95%
#> cp                 1.1801
#> b                  0.6917
#> a                  0.9048
#> cond~~cond         0.2837
#> reaction~~reaction 2.2482
#> pmi~~pmi           2.1310
#> indirect           0.4034
#> direct             1.1801
#> total              1.5829
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1470 0.0841 20  0.0076  0.0114  0.0285 0.3310 0.3428
#> b                  0.4277 0.0674 20  0.2952  0.2974  0.3073 0.5201 0.5232
#> a                  0.1569 0.0988 20 -0.0414 -0.0408 -0.0381 0.2904 0.3041
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7757 0.0539 20  0.6513  0.6542  0.6670 0.8432 0.8434
#> pmi~~pmi           0.9754 0.0268 20  0.9056  0.9074  0.9154 0.9998 0.9998
#> indirect           0.0671 0.0391 20 -0.0176 -0.0173 -0.0160 0.1104 0.1167
#> direct             0.1470 0.0841 20  0.0076  0.0114  0.0285 0.3310 0.3428
#> total              0.2141 0.0972 20  0.0581  0.0621  0.0794 0.4309 0.4573
#>                    99.95%
#> cp                 0.3454
#> b                  0.5239
#> a                  0.3072
#> cond~~cond         1.0000
#> reaction~~reaction 0.8434
#> pmi~~pmi           0.9998
#> indirect           0.1181
#> direct             0.3454
#> total              0.4633