Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.2136 0.3392 20 -0.3601 -0.3545 -0.3296 0.7902 0.8267
#> b                  0.4969 0.1411 20  0.2388  0.2411  0.2512 0.6626 0.6756
#> a                  0.4280 0.2459 20  0.0308  0.0333  0.0443 0.9040 0.9381
#> cond~~cond         0.2485 0.0364 20  0.1801  0.1821  0.1911 0.3125 0.3204
#> reaction~~reaction 1.9317 0.3171 20  1.3511  1.3764  1.4886 2.5195 2.5810
#> pmi~~pmi           1.7006 0.1991 20  1.4080  1.4182  1.4635 2.1233 2.2302
#> reaction~1         0.5536 0.7302 20 -0.3840 -0.3710 -0.3129 1.7115 1.7186
#> pmi~1              5.4018 0.1627 20  5.0737  5.0846  5.1327 5.6866 5.6898
#> cond~1             0.4657 0.0446 20  0.3872  0.3874  0.3882 0.5501 0.5543
#> indirect           0.2127 0.1251 20  0.0100  0.0107  0.0140 0.4757 0.5034
#> direct             0.2136 0.3392 20 -0.3601 -0.3545 -0.3296 0.7902 0.8267
#> total              0.4262 0.3832 20 -0.2806 -0.2767 -0.2594 1.0374 1.0403
#>                    99.95%
#> cp                 0.8349
#> b                  0.6785
#> a                  0.9457
#> cond~~cond         0.3222
#> reaction~~reaction 2.5948
#> pmi~~pmi           2.2543
#> reaction~1         1.7202
#> pmi~1              5.6905
#> cond~1             0.5552
#> indirect           0.5096
#> direct             0.8349
#> total              1.0410
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0688 0.1086 20 -0.1151 -0.1142 -0.1097 0.2311 0.2336
#> b                  0.4241 0.1130 20  0.1827  0.1869  0.2056 0.5934 0.5958
#> a                  0.1615 0.0881 20  0.0115  0.0124  0.0165 0.3108 0.3235
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.8060 0.0923 20  0.5485  0.5576  0.5976 0.9142 0.9162
#> pmi~~pmi           0.9739 0.0294 20  0.8935  0.8953  0.9031 0.9997 0.9998
#> indirect           0.3576 0.0401 20  0.0036  0.0039  0.0049 0.1492 0.1698
#> direct             4.0879 0.1086 20 -0.1151 -0.1142 -0.1097 0.2311 0.2336
#> total              0.9343 0.1228 20 -0.0897 -0.0891 -0.0864 0.3383 0.3532
#>                    99.95%
#> cp                 0.2342
#> b                  0.5963
#> a                  0.3263
#> cond~~cond         1.0000
#> reaction~~reaction 0.9167
#> pmi~~pmi           0.9999
#> indirect           0.1744
#> direct             0.2342
#> total              0.3565

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.2532 0.3769 20 -0.5390 -0.5379 -0.5331 0.7889 0.8086
#> b                  0.5156 0.0803 20  0.2924  0.2943  0.3027 0.5943 0.6133
#> a                  0.3826 0.2108 20 -0.1480 -0.1225 -0.0094 0.6877 0.6883
#> cond~~cond         0.2490 0.0360 20  0.2075  0.2075  0.2078 0.3307 0.3362
#> reaction~~reaction 1.8851 0.1940 20  1.6957  1.6984  1.7103 2.3503 2.4163
#> pmi~~pmi           1.7075 0.2391 20  1.2148  1.2219  1.2533 2.0846 2.1394
#> indirect           0.1973 0.1046 20 -0.0614 -0.0509 -0.0045 0.3801 0.4164
#> direct             0.2532 0.3769 20 -0.5390 -0.5379 -0.5331 0.7889 0.8086
#> total              0.4505 0.4146 20 -0.4520 -0.4455 -0.4166 0.9443 0.9487
#>                    99.95%
#> cp                 0.8131
#> b                  0.6175
#> a                  0.6885
#> cond~~cond         0.3374
#> reaction~~reaction 2.4312
#> pmi~~pmi           2.1518
#> indirect           0.4246
#> direct             0.8131
#> total              0.9497
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0628 0.1233 20 -0.1856 -0.1828 -0.1706 0.2702 0.2897
#> b                  0.4345 0.0538 20  0.2774  0.2783  0.2825 0.4681 0.4790
#> a                  0.1198 0.0794 20 -0.0565 -0.0465 -0.0020 0.2685 0.2732
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.8007 0.0515 20  0.6803  0.6870  0.7169 0.8937 0.8997
#> pmi~~pmi           0.9856 0.0219 20  0.9248  0.9253  0.9278 0.9966 0.9967
#> indirect           0.0521 0.0324 20 -0.0194 -0.0160 -0.0008 0.1200 0.1210
#> direct             0.0628 0.1233 20 -0.1856 -0.1828 -0.1706 0.2702 0.2897
#> total              0.1149 0.1346 20 -0.1594 -0.1549 -0.1350 0.3224 0.3362
#>                    99.95%
#> cp                 0.2941
#> b                  0.4814
#> a                  0.2743
#> cond~~cond         1.0000
#> reaction~~reaction 0.9011
#> pmi~~pmi           0.9967
#> indirect           0.1212
#> direct             0.2941
#> total              0.3394