Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.2795 0.2904 100 -0.6545 -0.5150 -0.2312 0.7573 0.9415
#> b                  0.5786 0.0954 100  0.3784  0.3793  0.4226 0.7624 0.7989
#> a                  0.6784 0.2275 100 -0.0013  0.0953  0.2520 1.0523 1.0976
#> cond~~cond         0.2486 0.0320 100  0.1755  0.1775  0.1851 0.3099 0.3225
#> reaction~~reaction 1.7545 0.2666 100  1.0981  1.1456  1.2267 2.1873 2.2502
#> pmi~~pmi           1.5220 0.1956 100  0.9170  0.9846  1.1350 1.8633 1.9364
#> reaction~1         0.1397 0.5550 100 -1.0640 -1.0473 -0.8983 1.0086 1.1142
#> pmi~1              5.3126 0.1568 100  4.9057  4.9459  5.0110 5.5803 5.7171
#> cond~1             0.4871 0.0415 100  0.3610  0.3759  0.4132 0.5694 0.5891
#> indirect           0.3925 0.1493 100 -0.0030  0.0414  0.1170 0.6961 0.7693
#> direct             0.2795 0.2904 100 -0.6545 -0.5150 -0.2312 0.7573 0.9415
#> total              0.6721 0.3185 100 -0.3503 -0.2191  0.0146 1.2277 1.4191
#>                    99.95%
#> cp                 0.9991
#> b                  0.8063
#> a                  1.1212
#> cond~~cond         0.3273
#> reaction~~reaction 2.2891
#> pmi~~pmi           1.9657
#> reaction~1         1.1940
#> pmi~1              5.7343
#> cond~1             0.6038
#> indirect           0.7911
#> direct             0.9991
#> total              1.5258
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0904 0.0930 100 -0.2288 -0.1811 -0.0803 0.2514 0.2822
#> b                  0.4802 0.0713 100  0.2928  0.3040  0.3208 0.6070 0.6477
#> a                  0.2644 0.0895 100 -0.0021  0.0297  0.0973 0.4298 0.4496
#> cond~~cond         1.0000 0.0000 100  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7383 0.0697 100  0.5568  0.5761  0.6137 0.8494 0.8747
#> pmi~~pmi           0.9301 0.0474 100  0.7885  0.7977  0.8152 0.9905 0.9978
#> indirect           0.0906 0.0491 100 -0.0012  0.0139  0.0348 0.2308 0.2596
#> direct             4.1531 0.0930 100 -0.2288 -0.1811 -0.0803 0.2514 0.2822
#> total              0.9770 0.0991 100 -0.1247 -0.0778  0.0046 0.3968 0.4273
#>                    99.95%
#> cp                 0.2831
#> b                  0.6493
#> a                  0.4599
#> cond~~cond         1.0000
#> reaction~~reaction 0.8841
#> pmi~~pmi           0.9998
#> indirect           0.2661
#> direct             0.2831
#> total              0.4326

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.2357 0.3062 100 -0.7328 -0.6489 -0.4184 0.7068 0.7441
#> b                  0.5750 0.1120 100  0.2994  0.3144  0.3686 0.7822 0.8698
#> a                  0.6712 0.2885 100  0.0566  0.1099  0.1761 1.2610 1.4027
#> cond~~cond         0.2499 0.0319 100  0.1901  0.1903  0.1972 0.3078 0.3135
#> reaction~~reaction 1.7565 0.2509 100  1.2459  1.2532  1.2969 2.2065 2.3747
#> pmi~~pmi           1.4732 0.1734 100  1.0575  1.0801  1.1537 1.8110 1.9344
#> indirect           0.3851 0.1785 100  0.0283  0.0574  0.1066 0.7557 0.8146
#> direct             0.2357 0.3062 100 -0.7328 -0.6489 -0.4184 0.7068 0.7441
#> total              0.6208 0.3200 100 -0.2966 -0.2699 -0.1398 1.0900 1.1081
#>                    99.95%
#> cp                 0.7577
#> b                  0.8867
#> a                  1.4756
#> cond~~cond         0.3166
#> reaction~~reaction 2.4115
#> pmi~~pmi           1.9359
#> indirect           0.8284
#> direct             0.7577
#> total              1.1118
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0437 0.1023 100 -0.2314 -0.2169 -0.1376 0.2461 0.2615
#> b                  0.4821 0.0811 100  0.2510  0.2586  0.3084 0.6043 0.6490
#> a                  0.2239 0.1098 100  0.0247  0.0381  0.0727 0.4922 0.5266
#> cond~~cond         1.0000 0.0000 100  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7562 0.0671 100  0.5474  0.5653  0.6221 0.8851 0.9106
#> pmi~~pmi           0.9499 0.0635 100  0.7065  0.7224  0.7575 0.9947 0.9983
#> indirect           0.1079 0.0578 100  0.0115  0.0186  0.0379 0.2512 0.2803
#> direct             0.0437 0.1023 100 -0.2314 -0.2169 -0.1376 0.2461 0.2615
#> total              0.1517 0.1054 100 -0.1068 -0.0920 -0.0478 0.3610 0.3865
#>                    99.95%
#> cp                 0.2617
#> b                  0.6757
#> a                  0.5417
#> cond~~cond         1.0000
#> reaction~~reaction 0.9115
#> pmi~~pmi           0.9993
#> indirect           0.2815
#> direct             0.2617
#> total              0.3997