Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.4000 0.3070 20 -0.2500 -0.2368 -0.1783 0.9763 1.0380
#> b                  0.4942 0.0592 20  0.3980  0.3996  0.4068 0.6147 0.6379
#> a                  0.3932 0.2706 20 -0.2911 -0.2567 -0.1035 0.8302 0.9046
#> cond~~cond         0.2478 0.0274 20  0.1893  0.1904  0.1955 0.2898 0.2914
#> reaction~~reaction 1.9095 0.2064 20  1.4097  1.4151  1.4389 2.0807 2.1057
#> pmi~~pmi           1.7197 0.2503 20  1.1665  1.1764  1.2205 2.0129 2.0269
#> reaction~1         0.5527 0.4405 20 -0.4582 -0.4312 -0.3109 1.1865 1.2250
#> pmi~1              5.3685 0.1786 20  5.0621  5.0669  5.0884 5.7106 5.8062
#> cond~1             0.4628 0.0559 20  0.3752  0.3762  0.3808 0.5616 0.5717
#> indirect           0.1943 0.1277 20 -0.1464 -0.1295 -0.0541 0.4025 0.4351
#> direct             0.4000 0.3070 20 -0.2500 -0.2368 -0.1783 0.9763 1.0380
#> total              0.5943 0.3508 20 -0.0412 -0.0281  0.0298 1.3183 1.3922
#>                    99.95%
#> cp                 1.0519
#> b                  0.6431
#> a                  0.9214
#> cond~~cond         0.2918
#> reaction~~reaction 2.1113
#> pmi~~pmi           2.0300
#> reaction~1         1.2337
#> pmi~1              5.8277
#> cond~1             0.5739
#> indirect           0.4425
#> direct             1.0519
#> total              1.4088
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1281 0.1001 20 -0.0813 -0.0770 -0.0577 0.3073 0.3353
#> b                  0.4214 0.0419 20  0.3508  0.3517  0.3560 0.4990 0.5102
#> a                  0.1477 0.0967 20 -0.1124 -0.0994 -0.0420 0.2922 0.3084
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7900 0.0652 20  0.6022  0.6084  0.6357 0.8498 0.8504
#> pmi~~pmi           0.9782 0.0258 20  0.9026  0.9047  0.9142 0.9977 0.9984
#> indirect           0.3555 0.0412 20 -0.0484 -0.0428 -0.0178 0.1301 0.1402
#> direct             4.0489 0.1001 20 -0.0813 -0.0770 -0.0577 0.3073 0.3353
#> total              0.9297 0.1143 20 -0.0134 -0.0091  0.0100 0.4148 0.4497
#>                    99.95%
#> cp                 0.3416
#> b                  0.5127
#> a                  0.3120
#> cond~~cond         1.0000
#> reaction~~reaction 0.8506
#> pmi~~pmi           0.9986
#> indirect           0.1425
#> direct             0.3416
#> total              0.4575

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.4634 0.2531 20 -0.0714 -0.0646 -0.0341 0.9006 0.9087
#> b                  0.4865 0.1062 20  0.3296  0.3326  0.3460 0.7082 0.7488
#> a                  0.3323 0.2372 20 -0.1626 -0.1500 -0.0937 0.7400 0.7584
#> cond~~cond         0.2484 0.0249 20  0.2056  0.2074  0.2155 0.3019 0.3061
#> reaction~~reaction 1.9061 0.2218 20  1.4690  1.4741  1.4970 2.1933 2.2069
#> pmi~~pmi           1.7423 0.1976 20  1.2500  1.2594  1.3012 2.0015 2.0926
#> indirect           0.1633 0.1029 20 -0.0593 -0.0549 -0.0352 0.3053 0.3439
#> direct             0.4634 0.2531 20 -0.0714 -0.0646 -0.0341 0.9006 0.9087
#> total              0.6267 0.2657 20 -0.0166 -0.0062  0.0401 0.9235 0.9895
#>                    99.95%
#> cp                 0.9105
#> b                  0.7579
#> a                  0.7625
#> cond~~cond         0.3071
#> reaction~~reaction 2.2100
#> pmi~~pmi           2.1131
#> indirect           0.3526
#> direct             0.9105
#> total              1.0043
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1393 0.0842 20 -0.0207 -0.0187 -0.0097 0.2946 0.2974
#> b                  0.4351 0.0805 20  0.2648  0.2682  0.2833 0.5378 0.5461
#> a                  0.1673 0.0944 20 -0.0653 -0.0602 -0.0375 0.3029 0.3290
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7710 0.0606 20  0.6998  0.7004  0.7031 0.8968 0.9006
#> pmi~~pmi           0.9720 0.0283 20  0.8878  0.8914  0.9071 1.0000 1.0000
#> indirect           0.0728 0.0339 20 -0.0199 -0.0184 -0.0119 0.1044 0.1157
#> direct             0.1393 0.0842 20 -0.0207 -0.0187 -0.0097 0.2946 0.2974
#> total              0.2121 0.0875 20 -0.0048 -0.0012  0.0145 0.3007 0.3171
#>                    99.95%
#> cp                 0.2980
#> b                  0.5480
#> a                  0.3349
#> cond~~cond         1.0000
#> reaction~~reaction 0.9014
#> pmi~~pmi           1.0000
#> indirect           0.1182
#> direct             0.2980
#> total              0.3207