Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.2419 0.3117 100 -0.8247 -0.6097 -0.3339 0.9243 1.0302
#> b                  0.4875 0.1060 100  0.2286  0.2299  0.2528 0.6584 0.7140
#> a                  0.4645 0.2777 100 -0.1081 -0.1061  0.0116 1.0752 1.1592
#> cond~~cond         0.2500 0.0358 100  0.1390  0.1423  0.1597 0.3158 0.3279
#> reaction~~reaction 1.9477 0.2621 100  1.2549  1.3996  1.5953 2.4768 2.5701
#> pmi~~pmi           1.6476 0.2112 100  1.1259  1.1756  1.2630 2.0563 2.0834
#> reaction~1         0.6029 0.5926 100 -0.6883 -0.5524 -0.3473 1.8870 2.0392
#> pmi~1              5.3685 0.1757 100  4.9541  4.9643  5.0021 5.6739 5.7424
#> cond~1             0.4892 0.0419 100  0.4000  0.4004  0.4203 0.5694 0.5936
#> indirect           0.2264 0.1466 100 -0.0518 -0.0448  0.0038 0.5540 0.7235
#> direct             0.2419 0.3117 100 -0.8247 -0.6097 -0.3339 0.9243 1.0302
#> total              0.4683 0.3292 100 -0.3552 -0.3023 -0.0447 1.1135 1.2678
#>                    99.95%
#> cp                 1.0509
#> b                  0.7144
#> a                  1.1600
#> cond~~cond         0.3380
#> reaction~~reaction 2.6065
#> pmi~~pmi           2.0865
#> reaction~1         2.1283
#> pmi~1              5.7484
#> cond~1             0.6024
#> indirect           0.8163
#> direct             1.0509
#> total              1.3342
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0782 0.0981 100 -0.2808 -0.2064 -0.1038 0.2836 0.3399
#> b                  0.4110 0.0864 100  0.1748  0.1898  0.2118 0.5621 0.5892
#> a                  0.1781 0.1030 100 -0.0387 -0.0368  0.0057 0.3906 0.4119
#> cond~~cond         1.0000 0.0000 100  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.8136 0.0668 100  0.6257  0.6372  0.6589 0.9353 0.9440
#> pmi~~pmi           0.9683 0.0399 100  0.8243  0.8303  0.8474 0.9996 0.9999
#> indirect           0.3897 0.0459 100 -0.0152 -0.0137  0.0015 0.1766 0.2261
#> direct             4.1155 0.0981 100 -0.2808 -0.2064 -0.1038 0.2836 0.3399
#> total              0.9783 0.1021 100 -0.1209 -0.1021 -0.0143 0.3555 0.3993
#>                    99.95%
#> cp                 0.3413
#> b                  0.6002
#> a                  0.4191
#> cond~~cond         1.0000
#> reaction~~reaction 0.9481
#> pmi~~pmi           1.0000
#> indirect           0.2499
#> direct             0.3413
#> total              0.4278

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.2864 0.2713 100 -0.3030 -0.2958 -0.2300 0.8131 0.9276
#> b                  0.4655 0.1162 100  0.2039  0.2131  0.2716 0.6688 0.6815
#> a                  0.4215 0.2769 100 -0.3543 -0.2754 -0.1142 0.8513 1.0728
#> cond~~cond         0.2495 0.0295 100  0.1874  0.1898  0.1942 0.3097 0.3128
#> reaction~~reaction 1.9553 0.2629 100  1.4039  1.4277  1.4605 2.5099 2.5802
#> pmi~~pmi           1.6323 0.2286 100  1.1392  1.1431  1.2036 2.1584 2.2198
#> indirect           0.1994 0.1498 100 -0.1806 -0.1337 -0.0399 0.4962 0.6690
#> direct             0.2864 0.2713 100 -0.3030 -0.2958 -0.2300 0.8131 0.9276
#> total              0.4858 0.2911 100 -0.1127 -0.1021 -0.0647 1.0255 1.0806
#>                    99.95%
#> cp                 0.9948
#> b                  0.6864
#> a                  1.0943
#> cond~~cond         0.3131
#> reaction~~reaction 2.6021
#> pmi~~pmi           2.2358
#> indirect           0.6989
#> direct             0.9948
#> total              1.0967
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0284 0.0897 100 -0.0951 -0.0949 -0.0753 0.2666 0.3096
#> b                  0.3968 0.0887 100  0.1767  0.1817  0.2297 0.5515 0.5722
#> a                  0.1885 0.1067 100 -0.1290 -0.0993 -0.0454 0.3346 0.3829
#> cond~~cond         1.0000 0.0000 100  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.8375 0.0668 100  0.6391  0.6474  0.6593 0.9224 0.9450
#> pmi~~pmi           0.9645 0.0339 100  0.8319  0.8524  0.8880 1.0000 1.0000
#> indirect           0.0748 0.0467 100 -0.0564 -0.0430 -0.0126 0.1632 0.1977
#> direct             0.0284 0.0897 100 -0.0951 -0.0949 -0.0753 0.2666 0.3096
#> total              0.1032 0.0953 100 -0.0382 -0.0344 -0.0218 0.3279 0.3613
#>                    99.95%
#> cp                 0.3213
#> b                  0.5805
#> a                  0.4098
#> cond~~cond         1.0000
#> reaction~~reaction 0.9453
#> pmi~~pmi           1.0000
#> indirect           0.2089
#> direct             0.3213
#> total              0.3688