Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.3549 0.2801 20 -0.1242 -0.1032 -0.0100 0.9675 0.9908
#> b                  0.5587 0.0745 20  0.3848  0.3870  0.3970 0.6332 0.6474
#> a                  0.6029 0.1617 20  0.0957  0.1028  0.1343 0.6570 0.6646
#> cond~~cond         0.2499 0.0320 20  0.2111  0.2115  0.2134 0.3167 0.3223
#> reaction~~reaction 1.7231 0.3395 20  1.1684  1.1794  1.2282 2.3491 2.4335
#> pmi~~pmi           1.7036 0.2607 20  1.1907  1.2177  1.3377 2.2729 2.3019
#> reaction~1         0.2148 0.3947 20 -0.1723 -0.1688 -0.1531 1.0522 1.0791
#> pmi~1              5.3160 0.1312 20  5.1454  5.1482  5.1606 5.6368 5.6503
#> cond~1             0.4864 0.0402 20  0.4095  0.4102  0.4129 0.5671 0.5759
#> indirect           0.3368 0.0932 20  0.0563  0.0573  0.0620 0.3826 0.3962
#> direct             0.3549 0.2801 20 -0.1242 -0.1032 -0.0100 0.9675 0.9908
#> total              0.6918 0.2818 20  0.2385  0.2492  0.2970 1.2803 1.2953
#>                    99.95%
#> cp                 0.9961
#> b                  0.6505
#> a                  0.6664
#> cond~~cond         0.3235
#> reaction~~reaction 2.4525
#> pmi~~pmi           2.3085
#> reaction~1         1.0851
#> pmi~1              5.6534
#> cond~1             0.5779
#> indirect           0.3992
#> direct             0.9961
#> total              1.2987
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%   2.5%  97.5%  99.5% 99.95%
#> cp                 0.1151 0.0957 20 -0.0332 -0.0271 0.0001 0.3365 0.3673 0.3742
#> b                  0.4857 0.0837 20  0.3238  0.3253 0.3323 0.5761 0.5791 0.5798
#> a                  0.2250 0.0639 20  0.0360  0.0390 0.0524 0.2541 0.2569 0.2576
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000 1.0000 1.0000 1.0000 1.0000
#> reaction~~reaction 0.7257 0.0818 20  0.5355  0.5439 0.5816 0.8328 0.8336 0.8338
#> pmi~~pmi           0.9494 0.0202 20  0.9337  0.9340 0.9354 0.9969 0.9984 0.9987
#> indirect           0.1394 0.0287 20  0.0204  0.0206 0.0216 0.1157 0.1169 0.1172
#> direct             3.9684 0.0957 20 -0.0332 -0.0271 0.0001 0.3365 0.3673 0.3742
#> total              0.9731 0.0984 20  0.0641  0.0691 0.0912 0.4446 0.4799 0.4879

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.4237 0.2818 20 -0.2376 -0.2206 -0.1452 0.7659 0.7854
#> b                  0.5717 0.1082 20  0.3435  0.3440  0.3465 0.7297 0.7552
#> a                  0.5971 0.2410 20  0.0132  0.0231  0.0670 0.8454 0.8616
#> cond~~cond         0.2496 0.0409 20  0.1679  0.1697  0.1778 0.3131 0.3208
#> reaction~~reaction 1.7374 0.2520 20  1.3063  1.3125  1.3398 2.2280 2.2360
#> pmi~~pmi           1.7521 0.1846 20  1.3184  1.3328  1.3969 1.9998 2.0078
#> indirect           0.3421 0.1609 20  0.0075  0.0135  0.0400 0.5915 0.6052
#> direct             0.4237 0.2818 20 -0.2376 -0.2206 -0.1452 0.7659 0.7854
#> total              0.7658 0.3096 20  0.0825  0.0895  0.1203 1.2391 1.2971
#>                    99.95%
#> cp                 0.7897
#> b                  0.7610
#> a                  0.8653
#> cond~~cond         0.3225
#> reaction~~reaction 2.2378
#> pmi~~pmi           2.0096
#> indirect           0.6083
#> direct             0.7897
#> total              1.3102
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1323 0.0916 20 -0.0827 -0.0767 -0.0500 0.2446 0.2601
#> b                  0.4722 0.0849 20  0.2825  0.2862  0.3026 0.5818 0.5914
#> a                  0.2105 0.0852 20  0.0055  0.0093  0.0262 0.3065 0.3098
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7332 0.0816 20  0.5584  0.5621  0.5786 0.8461 0.8573
#> pmi~~pmi           0.9557 0.0283 20  0.9035  0.9040  0.9061 0.9988 0.9997
#> indirect           0.0994 0.0471 20  0.0025  0.0044  0.0128 0.1737 0.1823
#> direct             0.1323 0.0916 20 -0.0827 -0.0767 -0.0500 0.2446 0.2601
#> total              0.2317 0.0902 20  0.0287  0.0312  0.0421 0.3614 0.3687
#>                    99.95%
#> cp                 0.2635
#> b                  0.5936
#> a                  0.3106
#> cond~~cond         1.0000
#> reaction~~reaction 0.8598
#> pmi~~pmi           1.0000
#> indirect           0.1843
#> direct             0.2635
#> total              0.3704