Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.4026 0.2650 100 -0.3611 -0.2244 -0.0412 0.9324 1.0715
#> b                  0.4646 0.0998 100  0.2254  0.2397  0.2852 0.6516 0.7007
#> a                  0.4710 0.2798 100 -0.3611 -0.2914 -0.1475 0.9890 1.0876
#> cond~~cond         0.2491 0.0313 100  0.1718  0.1786  0.1921 0.3081 0.3263
#> reaction~~reaction 1.8542 0.2274 100  1.3162  1.3551  1.4259 2.2664 2.4023
#> pmi~~pmi           1.7572 0.2206 100  1.2189  1.2518  1.4000 2.2122 2.2630
#> reaction~1         0.6705 0.5870 100 -0.6715 -0.6587 -0.4349 1.6998 2.0785
#> pmi~1              5.3741 0.1661 100  4.9488  4.9681  5.0689 5.7026 5.7256
#> cond~1             0.4729 0.0465 100  0.3574  0.3673  0.3874 0.5690 0.5934
#> indirect           0.2188 0.1417 100 -0.1254 -0.1186 -0.0680 0.4763 0.5953
#> direct             0.4026 0.2650 100 -0.3611 -0.2244 -0.0412 0.9324 1.0715
#> total              0.6214 0.2989 100 -0.2715 -0.1377  0.1363 1.1476 1.3253
#>                    99.95%
#> cp                 1.0799
#> b                  0.7040
#> a                  1.0977
#> cond~~cond         0.3324
#> reaction~~reaction 2.4175
#> pmi~~pmi           2.2970
#> reaction~1         2.0914
#> pmi~1              5.7307
#> cond~1             0.6123
#> indirect           0.6226
#> direct             1.0799
#> total              1.3321
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1316 0.0813 100 -0.1005 -0.0628 -0.0129 0.2777 0.3378
#> b                  0.4098 0.0742 100  0.2232  0.2345  0.2775 0.5413 0.5840
#> a                  0.1746 0.0990 100 -0.1265 -0.1047 -0.0516 0.3405 0.3773
#> cond~~cond         1.0000 0.0000 100  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7959 0.0685 100  0.6036  0.6199  0.6441 0.8995 0.9234
#> pmi~~pmi           0.9695 0.0327 100  0.8395  0.8570  0.8840 0.9990 0.9996
#> indirect           0.4393 0.0440 100 -0.0432 -0.0428 -0.0213 0.1507 0.1852
#> direct             3.9919 0.0813 100 -0.1005 -0.0628 -0.0129 0.2777 0.3378
#> total              0.9477 0.0897 100 -0.0755 -0.0379  0.0486 0.3698 0.3879
#>                    99.95%
#> cp                 0.3513
#> b                  0.6193
#> a                  0.4005
#> cond~~cond         1.0000
#> reaction~~reaction 0.9364
#> pmi~~pmi           0.9999
#> indirect           0.1940
#> direct             0.3513
#> total              0.3987

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.3261 0.2229 100 -0.3909 -0.2891 -0.0792 0.7640 0.8753
#> b                  0.4921 0.0905 100  0.2708  0.2727  0.3264 0.6564 0.6753
#> a                  0.4662 0.2611 100 -0.2559 -0.2300 -0.0897 0.9427 1.0422
#> cond~~cond         0.2489 0.0287 100  0.1820  0.1870  0.1963 0.3080 0.3115
#> reaction~~reaction 1.8761 0.2414 100  1.1635  1.1980  1.4572 2.2711 2.3845
#> pmi~~pmi           1.7548 0.2490 100  1.1744  1.2173  1.3210 2.2166 2.3847
#> indirect           0.2300 0.1397 100 -0.1281 -0.1136 -0.0353 0.5442 0.5754
#> direct             0.3261 0.2229 100 -0.3909 -0.2891 -0.0792 0.7640 0.8753
#> total              0.5561 0.2373 100 -0.0718 -0.0456  0.1269 1.0160 1.1052
#>                    99.95%
#> cp                 0.8925
#> b                  0.6801
#> a                  1.0865
#> cond~~cond         0.3130
#> reaction~~reaction 2.4600
#> pmi~~pmi           2.4235
#> indirect           0.5884
#> direct             0.8925
#> total              1.1596
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1083 0.0720 100 -0.1172 -0.0904 -0.0227 0.2520 0.2839
#> b                  0.4112 0.0737 100  0.2560  0.2602  0.2898 0.5750 0.5926
#> a                  0.1410 0.0908 100 -0.0980 -0.0873 -0.0308 0.3306 0.3570
#> cond~~cond         1.0000 0.0000 100  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.8066 0.0629 100  0.6257  0.6292  0.6507 0.8769 0.9064
#> pmi~~pmi           0.9801 0.0297 100  0.8686  0.8725  0.8907 0.9994 1.0000
#> indirect           0.0580 0.0436 100 -0.0401 -0.0367 -0.0115 0.1674 0.1980
#> direct             0.1083 0.0720 100 -0.1172 -0.0904 -0.0227 0.2520 0.2839
#> total              0.1663 0.0753 100 -0.0215 -0.0140  0.0443 0.3191 0.3502
#>                    99.95%
#> cp                 0.2877
#> b                  0.5998
#> a                  0.3625
#> cond~~cond         1.0000
#> reaction~~reaction 0.9214
#> pmi~~pmi           1.0000
#> indirect           0.2031
#> direct             0.2877
#> total              0.3729