Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1109 0.2784 100 -0.6750 -0.5293 -0.3320 0.6894 0.8164
#> b                  0.4889 0.0833 100  0.3145  0.3292  0.3483 0.6270 0.6586
#> a                  0.3688 0.2561 100 -0.4494 -0.3151 -0.1212 0.8467 1.1319
#> cond~~cond         0.2486 0.0369 100  0.1460  0.1538  0.1735 0.3105 0.3239
#> reaction~~reaction 2.0252 0.2974 100  1.2991  1.3471  1.5012 2.5129 2.8878
#> pmi~~pmi           1.6421 0.2396 100  1.1331  1.1478  1.3056 2.0713 2.3223
#> reaction~1         0.6881 0.4954 100 -0.3666 -0.3458 -0.2367 1.5529 1.7541
#> pmi~1              5.4491 0.1614 100  4.9960  5.0678  5.1619 5.7583 5.7995
#> cond~1             0.4643 0.0423 100  0.3576  0.3627  0.3767 0.5401 0.5697
#> indirect           0.1803 0.1324 100 -0.2887 -0.1855 -0.0551 0.4684 0.5368
#> direct             0.1109 0.2784 100 -0.6750 -0.5293 -0.3320 0.6894 0.8164
#> total              0.2912 0.3201 100 -0.6903 -0.4871 -0.2070 0.9233 1.1729
#>                    99.95%
#> cp                 0.8462
#> b                  0.6693
#> a                  1.3151
#> cond~~cond         0.3277
#> reaction~~reaction 2.9773
#> pmi~~pmi           2.4099
#> reaction~1         1.8159
#> pmi~1              5.8033
#> cond~1             0.5807
#> indirect           0.5467
#> direct             0.8462
#> total              1.3748
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0354 0.0893 100 -0.2162 -0.1673 -0.1066 0.2248 0.2429
#> b                  0.4053 0.0687 100  0.2507  0.2598  0.2967 0.5306 0.5585
#> a                  0.1420 0.0962 100 -0.1902 -0.1317 -0.0465 0.3295 0.4009
#> cond~~cond         1.0000 0.0000 100  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.8304 0.0582 100  0.6447  0.6600  0.6981 0.9016 0.9315
#> pmi~~pmi           0.9798 0.0306 100  0.8203  0.8386  0.8908 0.9997 0.9998
#> indirect           0.4407 0.0420 100 -0.0941 -0.0608 -0.0175 0.1524 0.1622
#> direct             4.2092 0.0893 100 -0.2162 -0.1673 -0.1066 0.2248 0.2429
#> total              0.9313 0.1018 100 -0.2212 -0.1545 -0.0620 0.2960 0.3584
#>                    99.95%
#> cp                 0.2436
#> b                  0.5683
#> a                  0.4237
#> cond~~cond         1.0000
#> reaction~~reaction 0.9393
#> pmi~~pmi           0.9998
#> indirect           0.1650
#> direct             0.2436
#> total              0.3943

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0329 0.2965 100 -0.6190 -0.5890 -0.4516 0.6781 0.6948
#> b                  0.4761 0.0982 100  0.2458  0.2461  0.2739 0.6580 0.7009
#> a                  0.4007 0.2584 100 -0.3692 -0.2853 -0.0683 0.8896 0.9817
#> cond~~cond         0.2485 0.0298 100  0.1700  0.1764  0.1887 0.3040 0.3178
#> reaction~~reaction 1.9951 0.2460 100  1.2834  1.3625  1.4738 2.3844 2.5315
#> pmi~~pmi           1.6964 0.2479 100  1.0884  1.1386  1.2413 2.1754 2.2193
#> indirect           0.1924 0.1355 100 -0.1473 -0.1148 -0.0257 0.4420 0.5938
#> direct             0.0329 0.2965 100 -0.6190 -0.5890 -0.4516 0.6781 0.6948
#> total              0.2253 0.3260 100 -0.5181 -0.4888 -0.2189 0.8981 1.0095
#>                    99.95%
#> cp                 0.6994
#> b                  0.7092
#> a                  1.0451
#> cond~~cond         0.3207
#> reaction~~reaction 2.5793
#> pmi~~pmi           2.2205
#> indirect           0.6246
#> direct             0.6994
#> total              1.0588
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0096 0.0956 100 -0.2158 -0.2033 -0.1564 0.2062 0.2392
#> b                  0.3909 0.0808 100  0.1914  0.2051  0.2492 0.5441 0.5830
#> a                  0.1204 0.0946 100 -0.1452 -0.1148 -0.0260 0.3308 0.3405
#> cond~~cond         1.0000 0.0000 100  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.8462 0.0616 100  0.6677  0.6686  0.6833 0.9236 0.9462
#> pmi~~pmi           0.9855 0.0289 100  0.8832  0.8841  0.8906 0.9994 0.9998
#> indirect           0.0471 0.0434 100 -0.0509 -0.0391 -0.0082 0.1490 0.1909
#> direct             0.0096 0.0956 100 -0.2158 -0.2033 -0.1564 0.2062 0.2392
#> total              0.0567 0.1025 100 -0.1801 -0.1630 -0.0852 0.2762 0.3242
#>                    99.95%
#> cp                 0.2493
#> b                  0.6057
#> a                  0.3418
#> cond~~cond         1.0000
#> reaction~~reaction 0.9494
#> pmi~~pmi           0.9999
#> indirect           0.2047
#> direct             0.2493
#> total              0.3486