Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.4591 0.2591 100 -0.1922 -0.1361 -0.0391 0.9082 1.0207
#> b                  0.5014 0.0990 100  0.2050  0.2619  0.3354 0.7616 0.7641
#> a                  0.4258 0.2462 100 -0.1853 -0.1729 -0.0732 0.8266 0.8889
#> cond~~cond         0.2505 0.0346 100  0.1591  0.1681  0.1852 0.3206 0.3356
#> reaction~~reaction 1.8586 0.2626 100  1.2563  1.2958  1.3645 2.3455 2.3758
#> pmi~~pmi           1.6791 0.2322 100  1.0940  1.1426  1.2601 2.1288 2.3124
#> reaction~1         0.4731 0.5728 100 -0.9432 -0.9297 -0.8372 1.3769 1.8000
#> pmi~1              5.3536 0.1725 100  4.7648  4.8285  5.0020 5.6627 5.7085
#> cond~1             0.5048 0.0486 100  0.3394  0.3613  0.4021 0.5900 0.6216
#> indirect           0.2135 0.1334 100 -0.0860 -0.0857 -0.0372 0.4740 0.5714
#> direct             0.4591 0.2591 100 -0.1922 -0.1361 -0.0391 0.9082 1.0207
#> total              0.6725 0.2677 100  0.0453  0.0604  0.1909 1.2347 1.3100
#>                    99.95%
#> cp                 1.0272
#> b                  0.7655
#> a                  0.9217
#> cond~~cond         0.3371
#> reaction~~reaction 2.3822
#> pmi~~pmi           2.4350
#> reaction~1         2.0533
#> pmi~1              5.7130
#> cond~1             0.6332
#> indirect           0.6373
#> direct             1.0272
#> total              1.3178
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1485 0.0835 100 -0.0584 -0.0425 -0.0120 0.3066 0.3595
#> b                  0.4255 0.0714 100  0.1914  0.2286  0.2865 0.5619 0.5747
#> a                  0.1623 0.0916 100 -0.0692 -0.0648 -0.0290 0.3214 0.3370
#> cond~~cond         1.0000 0.0000 100  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7764 0.0632 100  0.6096  0.6227  0.6407 0.8818 0.9013
#> pmi~~pmi           0.9737 0.0289 100  0.8802  0.8863  0.8967 0.9998 1.0000
#> indirect           0.3058 0.0413 100 -0.0272 -0.0270 -0.0122 0.1398 0.1741
#> direct             4.0768 0.0835 100 -0.0584 -0.0425 -0.0120 0.3066 0.3595
#> total              1.0087 0.0840 100  0.0152  0.0191  0.0553 0.3834 0.4613
#>                    99.95%
#> cp                 0.3869
#> b                  0.5803
#> a                  0.3461
#> cond~~cond         1.0000
#> reaction~~reaction 0.9016
#> pmi~~pmi           1.0000
#> indirect           0.1989
#> direct             0.3869
#> total              0.4964

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.4777 0.3074 100 -0.2251 -0.1907 -0.0702 1.1140 1.2714
#> b                  0.4959 0.0830 100  0.3173  0.3200  0.3364 0.6638 0.7065
#> a                  0.4237 0.2513 100 -0.3114 -0.2497 -0.0769 0.9339 1.0979
#> cond~~cond         0.2497 0.0341 100  0.1776  0.1814  0.1874 0.3119 0.3162
#> reaction~~reaction 1.8638 0.2301 100  1.1251  1.2676  1.4634 2.3022 2.3389
#> pmi~~pmi           1.6570 0.2223 100  1.1130  1.1560  1.2375 2.1369 2.2557
#> indirect           0.2109 0.1282 100 -0.1343 -0.1046 -0.0356 0.4543 0.6152
#> direct             0.4777 0.3074 100 -0.2251 -0.1907 -0.0702 1.1140 1.2714
#> total              0.6886 0.3351 100 -0.2192 -0.1202  0.0921 1.3251 1.5301
#>                    99.95%
#> cp                 1.3890
#> b                  0.7280
#> a                  1.1293
#> cond~~cond         0.3173
#> reaction~~reaction 2.3541
#> pmi~~pmi           2.2930
#> indirect           0.6682
#> direct             1.3890
#> total              1.6060
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1435 0.0949 100 -0.0773 -0.0655 -0.0244 0.3439 0.3822
#> b                  0.3969 0.0661 100  0.2666  0.2691  0.2827 0.5461 0.5682
#> a                  0.1660 0.0947 100 -0.1304 -0.1004 -0.0284 0.3438 0.3887
#> cond~~cond         1.0000 0.0000 100  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.8030 0.0580 100  0.6337  0.6353  0.6443 0.8641 0.8992
#> pmi~~pmi           0.9725 0.0321 100  0.8268  0.8479  0.8817 0.9996 1.0000
#> indirect           0.0659 0.0398 100 -0.0471 -0.0359 -0.0116 0.1436 0.1775
#> direct             0.1435 0.0949 100 -0.0773 -0.0655 -0.0244 0.3439 0.3822
#> total              0.2094 0.1018 100 -0.0771 -0.0422  0.0305 0.4116 0.4449
#>                    99.95%
#> cp                 0.3904
#> b                  0.5806
#> a                  0.4160
#> cond~~cond         1.0000
#> reaction~~reaction 0.9009
#> pmi~~pmi           1.0000
#> indirect           0.1783
#> direct             0.3904
#> total              0.4500