Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0329 0.2503 100 -0.5616 -0.5164 -0.4434 0.5192 0.5813
#> b                  0.5757 0.1169 100  0.1813  0.2557  0.3571 0.8317 0.8669
#> a                  0.5945 0.1937 100 -0.0331  0.0751  0.2315 0.9415 0.9897
#> cond~~cond         0.2477 0.0296 100  0.1797  0.1814  0.1893 0.2955 0.3030
#> reaction~~reaction 1.7962 0.2279 100  1.2502  1.2543  1.3099 2.2332 2.3657
#> pmi~~pmi           1.5681 0.2086 100  1.0404  1.0822  1.1378 1.9242 2.0738
#> reaction~1         0.1698 0.6547 100 -1.5196 -1.4393 -1.2817 1.3431 1.8896
#> pmi~1              5.3779 0.1368 100  5.1049  5.1205  5.1705 5.6538 5.7357
#> cond~1             0.4688 0.0541 100  0.3311  0.3362  0.3451 0.5701 0.6029
#> indirect           0.3422 0.1359 100 -0.0196  0.0257  0.1062 0.6218 0.6869
#> direct             0.0329 0.2503 100 -0.5616 -0.5164 -0.4434 0.5192 0.5813
#> total              0.3751 0.2555 100 -0.3167 -0.2509 -0.1403 0.8242 0.8944
#>                    99.95%
#> cp                 0.6209
#> b                  0.8706
#> a                  1.0086
#> cond~~cond         0.3043
#> reaction~~reaction 2.3681
#> pmi~~pmi           2.0986
#> reaction~1         2.2981
#> pmi~1              5.7430
#> cond~1             0.6246
#> indirect           0.7027
#> direct             0.6209
#> total              0.9171
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0107 0.0816 100 -0.1797 -0.1717 -0.1454 0.1594 0.1935
#> b                  0.4831 0.0874 100  0.1699  0.2069  0.3042 0.6405 0.6555
#> a                  0.2299 0.0741 100 -0.0159  0.0304  0.0849 0.3533 0.3822
#> cond~~cond         1.0000 0.0000 100  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7641 0.0771 100  0.5600  0.5724  0.5987 0.8768 0.9449
#> pmi~~pmi           0.9471 0.0326 100  0.8363  0.8533  0.8752 0.9928 0.9964
#> indirect           0.1108 0.0428 100 -0.0065  0.0099  0.0332 0.1991 0.2171
#> direct             4.1796 0.0816 100 -0.1797 -0.1717 -0.1454 0.1594 0.1935
#> total              0.9419 0.0811 100 -0.1010 -0.0805 -0.0461 0.2434 0.3019
#>                    99.95%
#> cp                 0.1976
#> b                  0.6629
#> a                  0.4044
#> cond~~cond         1.0000
#> reaction~~reaction 0.9499
#> pmi~~pmi           0.9992
#> indirect           0.2282
#> direct             0.1976
#> total              0.3106

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1192 0.2758 100 -0.6880 -0.5463 -0.3578 0.6389 0.7188
#> b                  0.5720 0.0974 100  0.3320  0.3426  0.3863 0.7541 0.7623
#> a                  0.5500 0.2345 100  0.1122  0.1220  0.1671 0.9979 1.1766
#> cond~~cond         0.2483 0.0318 100  0.1771  0.1807  0.1897 0.3104 0.3199
#> reaction~~reaction 1.7440 0.2197 100  1.1914  1.1925  1.3592 2.2333 2.2771
#> pmi~~pmi           1.6148 0.2194 100  1.1561  1.2050  1.2734 2.0942 2.2978
#> indirect           0.3155 0.1548 100  0.0565  0.0649  0.0940 0.6101 0.7650
#> direct             0.1192 0.2758 100 -0.6880 -0.5463 -0.3578 0.6389 0.7188
#> total              0.4346 0.2847 100 -0.2046 -0.1880 -0.1453 1.0210 1.1632
#>                    99.95%
#> cp                 0.7370
#> b                  0.7647
#> a                  1.2097
#> cond~~cond         0.3202
#> reaction~~reaction 2.2870
#> pmi~~pmi           2.3005
#> indirect           0.8744
#> direct             0.7370
#> total              1.2674
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                        est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 -0.0026 0.0881 100 -0.1981 -0.1626 -0.1097 0.2155 0.2412
#> b                   0.5035 0.0764 100  0.2697  0.2891  0.3446 0.6264 0.6468
#> a                   0.2433 0.0910 100  0.0447  0.0476  0.0608 0.3904 0.4494
#> cond~~cond          1.0000 0.0000 100  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction  0.7471 0.0716 100  0.5420  0.5560  0.5817 0.8541 0.8782
#> pmi~~pmi            0.9408 0.0427 100  0.7711  0.7970  0.8471 0.9963 0.9977
#> indirect            0.1225 0.0496 100  0.0204  0.0224  0.0288 0.2003 0.2350
#> direct             -0.0026 0.0881 100 -0.1981 -0.1626 -0.1097 0.2155 0.2412
#> total               0.1199 0.0909 100 -0.0590 -0.0569 -0.0455 0.3534 0.3715
#>                    99.95%
#> cp                 0.2453
#> b                  0.6594
#> a                  0.4782
#> cond~~cond         1.0000
#> reaction~~reaction 0.8783
#> pmi~~pmi           0.9980
#> indirect           0.2568
#> direct             0.2453
#> total              0.3736