Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.2699 0.2875 20 -0.1397 -0.1372 -0.1260 0.8185 0.8189
#> b                  0.5324 0.0830 20  0.3965  0.4001  0.4163 0.6940 0.7473
#> a                  0.5281 0.2208 20  0.1414  0.1499  0.1877 0.9425 0.9458
#> cond~~cond         0.2493 0.0279 20  0.1920  0.1925  0.1948 0.2873 0.2886
#> reaction~~reaction 1.8505 0.2502 20  1.3752  1.3904  1.4577 2.3128 2.3157
#> pmi~~pmi           1.7032 0.2816 20  1.4144  1.4208  1.4494 2.3041 2.3051
#> reaction~1         0.3620 0.4571 20 -0.8893 -0.8346 -0.5914 0.9304 0.9312
#> pmi~1              5.3922 0.1452 20  5.1297  5.1334  5.1497 5.5910 5.6193
#> cond~1             0.4879 0.0548 20  0.3854  0.3872  0.3955 0.5934 0.6162
#> indirect           0.2812 0.1558 20  0.0666  0.0691  0.0800 0.6472 0.7054
#> direct             0.2699 0.2875 20 -0.1397 -0.1372 -0.1260 0.8185 0.8189
#> total              0.5511 0.3196 20  0.1032  0.1062  0.1194 1.2284 1.3533
#>                    99.95%
#> cp                 0.8190
#> b                  0.7593
#> a                  0.9465
#> cond~~cond         0.2889
#> reaction~~reaction 2.3164
#> pmi~~pmi           2.3053
#> reaction~1         0.9314
#> pmi~1              5.6257
#> cond~1             0.6213
#> indirect           0.7185
#> direct             0.8190
#> total              1.3814
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0868 0.0927 20 -0.0437 -0.0427 -0.0385 0.2570 0.2686
#> b                  0.4567 0.0640 20  0.3603  0.3620  0.3695 0.5810 0.6010
#> a                  0.1980 0.0851 20  0.0452  0.0491  0.0662 0.3718 0.3755
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7681 0.0680 20  0.6027  0.6031  0.6050 0.8294 0.8409
#> pmi~~pmi           0.9608 0.0381 20  0.8584  0.8590  0.8617 0.9951 0.9974
#> indirect           0.2332 0.0482 20  0.0194  0.0207  0.0264 0.1960 0.2170
#> direct             4.0499 0.0927 20 -0.0437 -0.0427 -0.0385 0.2570 0.2686
#> total              0.9770 0.1019 20  0.0300  0.0309  0.0350 0.3935 0.4050
#>                    99.95%
#> cp                 0.2713
#> b                  0.6055
#> a                  0.3764
#> cond~~cond         1.0000
#> reaction~~reaction 0.8434
#> pmi~~pmi           0.9979
#> indirect           0.2217
#> direct             0.2713
#> total              0.4076

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1829 0.3023 20 -0.3266 -0.3264 -0.3255 0.6869 0.7653
#> b                  0.5695 0.0985 20  0.3935  0.3978  0.4172 0.7502 0.7587
#> a                  0.5129 0.2386 20  0.0412  0.0465  0.0700 0.9526 0.9731
#> cond~~cond         0.2493 0.0331 20  0.1878  0.1893  0.1959 0.3224 0.3389
#> reaction~~reaction 1.8301 0.2430 20  1.3228  1.3352  1.3902 2.1194 2.1282
#> pmi~~pmi           1.6600 0.2212 20  1.1125  1.1369  1.2451 2.0351 2.0767
#> indirect           0.2910 0.1399 20  0.0305  0.0346  0.0528 0.5280 0.5616
#> direct             0.1829 0.3023 20 -0.3266 -0.3264 -0.3255 0.6869 0.7653
#> total              0.4739 0.3542 20 -0.2451 -0.2353 -0.1918 1.0482 1.0641
#>                    99.95%
#> cp                 0.7830
#> b                  0.7606
#> a                  0.9777
#> cond~~cond         0.3426
#> reaction~~reaction 2.1302
#> pmi~~pmi           2.0860
#> indirect           0.5692
#> direct             0.7830
#> total              1.0677
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0789 0.1020 20 -0.0945 -0.0941 -0.0925 0.2523 0.2967
#> b                  0.4578 0.0776 20  0.3586  0.3620  0.3771 0.6348 0.6361
#> a                  0.2278 0.0846 20  0.0150  0.0175  0.0286 0.3437 0.3496
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7678 0.0755 20  0.5934  0.5935  0.5941 0.8263 0.8283
#> pmi~~pmi           0.9481 0.0325 20  0.8769  0.8778  0.8818 0.9990 0.9996
#> indirect           0.1043 0.0436 20  0.0095  0.0105  0.0152 0.1635 0.1749
#> direct             0.0789 0.1020 20 -0.0945 -0.0941 -0.0925 0.2523 0.2967
#> total              0.1831 0.1192 20 -0.0683 -0.0660 -0.0555 0.3492 0.3568
#>                    99.95%
#> cp                 0.3067
#> b                  0.6364
#> a                  0.3509
#> cond~~cond         1.0000
#> reaction~~reaction 0.8288
#> pmi~~pmi           0.9998
#> indirect           0.1775
#> direct             0.3067
#> total              0.3585