Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.2820 0.3025 20 -0.2513 -0.2506 -0.2474 0.6892 0.7175
#> b                  0.5313 0.1029 20  0.3043  0.3145  0.3597 0.6839 0.6950
#> a                  0.5952 0.2307 20 -0.0043  0.0072  0.0583 0.8144 0.8334
#> cond~~cond         0.2485 0.0332 20  0.1973  0.1981  0.2016 0.3104 0.3108
#> reaction~~reaction 1.8420 0.2590 20  1.2764  1.2784  1.2874 2.1328 2.1604
#> pmi~~pmi           1.7448 0.2134 20  1.3291  1.3354  1.3636 2.0543 2.0792
#> reaction~1         0.3076 0.6166 20 -0.9482 -0.8907 -0.6350 1.4177 1.6286
#> pmi~1              5.3694 0.1749 20  5.2091  5.2109  5.2187 5.7740 5.7766
#> cond~1             0.4695 0.0440 20  0.3810  0.3837  0.3958 0.5356 0.5411
#> indirect           0.3162 0.1302 20 -0.0025  0.0025  0.0249 0.4499 0.4831
#> direct             0.2820 0.3025 20 -0.2513 -0.2506 -0.2474 0.6892 0.7175
#> total              0.5982 0.3546 20 -0.2528 -0.2375 -0.1695 0.9843 1.0324
#>                    99.95%
#> cp                 0.7239
#> b                  0.6975
#> a                  0.8376
#> cond~~cond         0.3109
#> reaction~~reaction 2.1666
#> pmi~~pmi           2.0848
#> reaction~1         1.6761
#> pmi~1              5.7772
#> cond~1             0.5424
#> indirect           0.4906
#> direct             0.7239
#> total              1.0432
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0903 0.0970 20 -0.0927 -0.0918 -0.0879 0.2102 0.2161
#> b                  0.4620 0.0835 20  0.2958  0.3026  0.3331 0.6093 0.6145
#> a                  0.2192 0.0861 20 -0.0017  0.0026  0.0220 0.3085 0.3164
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7601 0.0731 20  0.6300  0.6308  0.6346 0.8585 0.8662
#> pmi~~pmi           0.9520 0.0287 20  0.8988  0.8999  0.9047 0.9989 0.9998
#> indirect           0.1976 0.0420 20 -0.0008  0.0009  0.0085 0.1503 0.1639
#> direct             3.9661 0.0970 20 -0.0927 -0.0918 -0.0879 0.2102 0.2161
#> total              0.9418 0.1112 20 -0.0830 -0.0781 -0.0561 0.3090 0.3126
#>                    99.95%
#> cp                 0.2174
#> b                  0.6157
#> a                  0.3182
#> cond~~cond         1.0000
#> reaction~~reaction 0.8679
#> pmi~~pmi           1.0000
#> indirect           0.1669
#> direct             0.2174
#> total              0.3134

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.2323 0.2404 20 -0.1147 -0.1121 -0.1008 0.7564 0.8490
#> b                  0.5349 0.0831 20  0.4071  0.4076  0.4098 0.6976 0.6988
#> a                  0.6073 0.3298 20 -0.1518 -0.1422 -0.0999 1.0358 1.0647
#> cond~~cond         0.2486 0.0306 20  0.1898  0.1906  0.1940 0.2980 0.2983
#> reaction~~reaction 1.8656 0.3100 20  1.2251  1.2481  1.3501 2.4118 2.5378
#> pmi~~pmi           1.7011 0.1400 20  1.3509  1.3708  1.4593 1.9626 1.9747
#> indirect           0.3246 0.1627 20 -0.0742 -0.0697 -0.0497 0.4773 0.4880
#> direct             0.2323 0.2404 20 -0.1147 -0.1121 -0.1008 0.7564 0.8490
#> total              0.5569 0.3234 20  0.0279  0.0308  0.0441 1.0221 1.0850
#>                    99.95%
#> cp                 0.8698
#> b                  0.6991
#> a                  1.0713
#> cond~~cond         0.2984
#> reaction~~reaction 2.5662
#> pmi~~pmi           1.9774
#> indirect           0.4904
#> direct             0.8698
#> total              1.0991
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0737 0.0782 20 -0.0351 -0.0342 -0.0304 0.2450 0.2839
#> b                  0.4384 0.0608 20  0.3803  0.3805  0.3813 0.5632 0.5659
#> a                  0.2602 0.1198 20 -0.0506 -0.0475 -0.0340 0.3800 0.3805
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7855 0.0560 20  0.6742  0.6746  0.6762 0.8444 0.8472
#> pmi~~pmi           0.9323 0.0424 20  0.8551  0.8552  0.8556 0.9996 0.9997
#> indirect           0.1141 0.0512 20 -0.0238 -0.0224 -0.0162 0.1483 0.1495
#> direct             0.0737 0.0782 20 -0.0351 -0.0342 -0.0304 0.2450 0.2839
#> total              0.1878 0.1040 20  0.0090  0.0098  0.0134 0.3435 0.3652
#>                    99.95%
#> cp                 0.2927
#> b                  0.5666
#> a                  0.3806
#> cond~~cond         1.0000
#> reaction~~reaction 0.8479
#> pmi~~pmi           0.9998
#> indirect           0.1498
#> direct             0.2927
#> total              0.3701