Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.4126 0.2083 20 -0.0097 -0.0083 -0.0023 0.6043 0.6156
#> b                  0.4567 0.0903 20  0.3411  0.3421  0.3463 0.6409 0.6502
#> a                  0.4006 0.2232 20 -0.0564 -0.0466 -0.0030 0.8520 0.8630
#> cond~~cond         0.2477 0.0303 20  0.1810  0.1838  0.1965 0.2986 0.3133
#> reaction~~reaction 1.7401 0.3095 20  1.1645  1.1749  1.2213 2.3547 2.3844
#> pmi~~pmi           1.6906 0.1787 20  1.3133  1.3136  1.3149 1.8517 1.8652
#> reaction~1         0.6704 0.5328 20 -0.4984 -0.4814 -0.4056 1.3420 1.4064
#> pmi~1              5.4341 0.1309 20  5.1427  5.1534  5.2013 5.6018 5.6055
#> cond~1             0.4602 0.0444 20  0.3633  0.3674  0.3859 0.5355 0.5379
#> indirect           0.1830 0.1020 20 -0.0294 -0.0249 -0.0049 0.3546 0.4026
#> direct             0.4126 0.2083 20 -0.0097 -0.0083 -0.0023 0.6043 0.6156
#> total              0.5956 0.1999 20  0.1064  0.1074  0.1117 0.7865 0.8511
#>                    99.95%
#> cp                 0.6181
#> b                  0.6522
#> a                  0.8654
#> cond~~cond         0.3166
#> reaction~~reaction 2.3911
#> pmi~~pmi           1.8683
#> reaction~1         1.4209
#> pmi~1              5.6063
#> cond~1             0.5384
#> indirect           0.4135
#> direct             0.6181
#> total              0.8656
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1391 0.0725 20 -0.0033 -0.0028 -0.0008 0.2143 0.2162
#> b                  0.4068 0.0763 20  0.2878  0.2892  0.2959 0.5423 0.5636
#> a                  0.1516 0.0815 20 -0.0216 -0.0176  0.0002 0.3133 0.3164
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7980 0.0700 20  0.6521  0.6553  0.6692 0.9019 0.9036
#> pmi~~pmi           0.9770 0.0278 20  0.8995  0.8999  0.9018 0.9995 0.9995
#> indirect           0.4540 0.0339 20 -0.0111 -0.0094 -0.0021 0.1184 0.1380
#> direct             4.1310 0.0725 20 -0.0033 -0.0028 -0.0008 0.2143 0.2162
#> total              0.9247 0.0683 20  0.0362  0.0373  0.0422 0.2654 0.2962
#>                    99.95%
#> cp                 0.2166
#> b                  0.5683
#> a                  0.3171
#> cond~~cond         1.0000
#> reaction~~reaction 0.9040
#> pmi~~pmi           0.9995
#> indirect           0.1424
#> direct             0.2166
#> total              0.3031

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.4475 0.3151 20 -0.4571 -0.4221 -0.2665 0.8459 0.9503
#> b                  0.4342 0.0684 20  0.2925  0.2939  0.2997 0.5463 0.5534
#> a                  0.3654 0.2080 20  0.0119  0.0239  0.0772 0.7512 0.8094
#> cond~~cond         0.2479 0.0348 20  0.1854  0.1864  0.1909 0.3019 0.3062
#> reaction~~reaction 1.7687 0.2345 20  1.3207  1.3241  1.3390 2.0789 2.1235
#> pmi~~pmi           1.6576 0.1846 20  1.3044  1.3159  1.3667 1.9659 2.0099
#> indirect           0.1591 0.0899 20  0.0048  0.0093  0.0293 0.3374 0.3610
#> direct             0.4475 0.3151 20 -0.4571 -0.4221 -0.2665 0.8459 0.9503
#> total              0.6066 0.3005 20 -0.1545 -0.1397 -0.0736 0.9258 1.0157
#>                    99.95%
#> cp                 0.9738
#> b                  0.5550
#> a                  0.8225
#> cond~~cond         0.3071
#> reaction~~reaction 2.1336
#> pmi~~pmi           2.0198
#> indirect           0.3663
#> direct             0.9738
#> total              1.0359
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1290 0.1086 20 -0.1597 -0.1476 -0.0938 0.2898 0.3029
#> b                  0.4054 0.0656 20  0.2656  0.2678  0.2774 0.4886 0.5094
#> a                  0.1345 0.0727 20  0.0046  0.0097  0.0325 0.2559 0.2607
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.8049 0.0557 20  0.6985  0.7016  0.7158 0.8968 0.9020
#> pmi~~pmi           0.9819 0.0217 20  0.9315  0.9320  0.9345 0.9980 0.9996
#> indirect           0.0545 0.0290 20  0.0016  0.0031  0.0099 0.1078 0.1087
#> direct             0.1290 0.1086 20 -0.1597 -0.1476 -0.0938 0.2898 0.3029
#> total              0.1835 0.1018 20 -0.0540 -0.0488 -0.0259 0.3183 0.3241
#>                    99.95%
#> cp                 0.3058
#> b                  0.5140
#> a                  0.2618
#> cond~~cond         1.0000
#> reaction~~reaction 0.9032
#> pmi~~pmi           0.9999
#> indirect           0.1089
#> direct             0.3058
#> total              0.3254