Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1564 0.2651 100 -0.5198 -0.4776 -0.2910 0.5665 0.7501
#> b                  0.4698 0.0982 100  0.2229  0.2445  0.2778 0.6621 0.6673
#> a                  0.4509 0.2665 100 -0.2672 -0.2573 -0.1629 0.9633 1.0107
#> cond~~cond         0.2476 0.0339 100  0.1700  0.1796  0.1967 0.3135 0.3269
#> reaction~~reaction 1.8656 0.2484 100  1.1019  1.1645  1.4028 2.3168 2.3616
#> pmi~~pmi           1.7641 0.2374 100  1.0859  1.1523  1.2919 2.1060 2.3225
#> reaction~1         0.7554 0.5601 100 -0.3918 -0.3575 -0.2652 1.8898 2.1008
#> pmi~1              5.4057 0.1606 100  5.0708  5.0713  5.0864 5.6619 5.7670
#> cond~1             0.4577 0.0500 100  0.3443  0.3468  0.3572 0.5370 0.5552
#> indirect           0.2118 0.1356 100 -0.1392 -0.1318 -0.0526 0.4500 0.6099
#> direct             0.1564 0.2651 100 -0.5198 -0.4776 -0.2910 0.5665 0.7501
#> total              0.3682 0.2927 100 -0.3094 -0.2973 -0.1306 0.8690 0.9900
#>                    99.95%
#> cp                 0.8504
#> b                  0.6712
#> a                  1.0232
#> cond~~cond         0.3324
#> reaction~~reaction 2.3714
#> pmi~~pmi           2.4208
#> reaction~1         2.2513
#> pmi~1              5.7894
#> cond~1             0.5618
#> indirect           0.6718
#> direct             0.8504
#> total              1.0337
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0514 0.0883 100 -0.1738 -0.1551 -0.0980 0.1930 0.2566
#> b                  0.4183 0.0794 100  0.2142  0.2176  0.2548 0.5653 0.5676
#> a                  0.1666 0.1001 100 -0.0990 -0.0971 -0.0646 0.3371 0.3897
#> cond~~cond         1.0000 0.0000 100  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.8152 0.0627 100  0.6521  0.6638  0.6812 0.9266 0.9534
#> pmi~~pmi           0.9723 0.0316 100  0.8366  0.8479  0.8862 0.9992 1.0000
#> indirect           0.4994 0.0441 100 -0.0450 -0.0446 -0.0185 0.1596 0.1873
#> direct             4.0131 0.0883 100 -0.1738 -0.1551 -0.0980 0.1930 0.2566
#> total              0.9197 0.0970 100 -0.1034 -0.0964 -0.0482 0.2862 0.3198
#>                    99.95%
#> cp                 0.3059
#> b                  0.5686
#> a                  0.4041
#> cond~~cond         1.0000
#> reaction~~reaction 0.9550
#> pmi~~pmi           1.0000
#> indirect           0.2091
#> direct             0.3059
#> total              0.3379

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.2010 0.2454 100 -0.3122 -0.2591 -0.1448 0.7625 0.8917
#> b                  0.4939 0.0952 100  0.2828  0.2843  0.3289 0.6601 0.7185
#> a                  0.4944 0.2662 100 -0.1999 -0.1679 -0.0296 0.9861 1.1722
#> cond~~cond         0.2493 0.0318 100  0.1754  0.1797  0.1876 0.3057 0.3167
#> reaction~~reaction 1.8216 0.2288 100  1.2464  1.2575  1.3899 2.2556 2.4102
#> pmi~~pmi           1.7110 0.2144 100  1.0954  1.1469  1.2337 2.0892 2.1426
#> indirect           0.2440 0.1478 100 -0.1192 -0.0953 -0.0133 0.5150 0.6887
#> direct             0.2010 0.2454 100 -0.3122 -0.2591 -0.1448 0.7625 0.8917
#> total              0.4450 0.2881 100 -0.1003 -0.0766 -0.0022 1.0479 1.1316
#>                    99.95%
#> cp                 0.9583
#> b                  0.7272
#> a                  1.1866
#> cond~~cond         0.3207
#> reaction~~reaction 2.4350
#> pmi~~pmi           2.1448
#> indirect           0.7358
#> direct             0.9583
#> total              1.2041
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0425 0.0792 100 -0.1076 -0.0878 -0.0512 0.2251 0.3031
#> b                  0.4917 0.0728 100  0.2543  0.2587  0.2863 0.5536 0.6202
#> a                  0.1381 0.0968 100 -0.0873 -0.0696 -0.0107 0.3643 0.3905
#> cond~~cond         1.0000 0.0000 100  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7507 0.0660 100  0.5570  0.5619  0.6519 0.8950 0.9329
#> pmi~~pmi           0.9809 0.0354 100  0.8464  0.8475  0.8672 0.9997 0.9999
#> indirect           0.0679 0.0452 100 -0.0421 -0.0331 -0.0040 0.1547 0.2012
#> direct             0.0425 0.0792 100 -0.1076 -0.0878 -0.0512 0.2251 0.3031
#> total              0.1104 0.0901 100 -0.0347 -0.0272 -0.0015 0.3304 0.3381
#>                    99.95%
#> cp                 0.3062
#> b                  0.6439
#> a                  0.3919
#> cond~~cond         1.0000
#> reaction~~reaction 0.9348
#> pmi~~pmi           1.0000
#> indirect           0.2056
#> direct             0.3062
#> total              0.3426