Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.2821 0.2872 100 -0.6856 -0.6097 -0.3113 0.7810 0.8053
#> b                  0.4754 0.0858 100  0.2117  0.2264  0.3038 0.6409 0.6710
#> a                  0.7004 0.2562 100  0.0605  0.1237  0.2334 1.2072 1.3613
#> cond~~cond         0.2487 0.0334 100  0.1771  0.1800  0.1887 0.3227 0.3442
#> reaction~~reaction 1.9058 0.2468 100  1.3796  1.3934  1.4587 2.3802 2.4381
#> pmi~~pmi           1.6753 0.2256 100  1.1172  1.1201  1.1846 2.1097 2.2277
#> reaction~1         0.6923 0.4770 100 -0.5018 -0.4625 -0.2683 1.5742 1.9162
#> pmi~1              5.2967 0.1583 100  4.9274  4.9797  5.0528 5.6275 5.7401
#> cond~1             0.4567 0.0468 100  0.3381  0.3418  0.3652 0.5448 0.5573
#> indirect           0.3330 0.1353 100  0.0270  0.0581  0.1040 0.6148 0.6663
#> direct             0.2821 0.2872 100 -0.6856 -0.6097 -0.3113 0.7810 0.8053
#> total              0.6151 0.2997 100 -0.1512 -0.1063  0.0170 1.0989 1.2219
#>                    99.95%
#> cp                 0.8071
#> b                  0.6744
#> a                  1.3940
#> cond~~cond         0.3473
#> reaction~~reaction 2.4488
#> pmi~~pmi           2.2327
#> reaction~1         2.1199
#> pmi~1              5.7490
#> cond~1             0.5669
#> indirect           0.6777
#> direct             0.8071
#> total              1.2792
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0912 0.0924 100 -0.2276 -0.1912 -0.1009 0.2454 0.2589
#> b                  0.4133 0.0700 100  0.1803  0.2029  0.2804 0.5361 0.5883
#> a                  0.2605 0.0964 100  0.0260  0.0532  0.0955 0.4409 0.4846
#> cond~~cond         1.0000 0.0000 100  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.8012 0.0585 100  0.6454  0.6604  0.6906 0.9008 0.9245
#> pmi~~pmi           0.9321 0.0538 100  0.7271  0.7634  0.8056 0.9909 0.9962
#> indirect           0.4489 0.0436 100  0.0107  0.0207  0.0356 0.1939 0.2211
#> direct             3.9509 0.0924 100 -0.2276 -0.1912 -0.1009 0.2454 0.2589
#> total              0.9157 0.0969 100 -0.0503 -0.0342  0.0061 0.3509 0.3985
#>                    99.95%
#> cp                 0.2662
#> b                  0.5985
#> a                  0.5221
#> cond~~cond         1.0000
#> reaction~~reaction 0.9441
#> pmi~~pmi           0.9991
#> indirect           0.2367
#> direct             0.2662
#> total              0.4245

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1901 0.2871 100 -0.6222 -0.5261 -0.3398 0.7077 0.8814
#> b                  0.5088 0.0974 100  0.3195  0.3254  0.3405 0.7063 0.7488
#> a                  0.7466 0.2429 100  0.0738  0.1535  0.3194 1.1690 1.2726
#> cond~~cond         0.2486 0.0305 100  0.1740  0.1758  0.1888 0.3076 0.3338
#> reaction~~reaction 1.9636 0.2803 100  1.4000  1.4147  1.4558 2.5069 2.5973
#> pmi~~pmi           1.6301 0.1966 100  1.1333  1.1942  1.2804 1.9738 2.1505
#> indirect           0.3800 0.1541 100  0.0294  0.0762  0.1363 0.7080 0.8454
#> direct             0.1901 0.2871 100 -0.6222 -0.5261 -0.3398 0.7077 0.8814
#> total              0.5701 0.3237 100 -0.1805 -0.1370 -0.0486 1.1798 1.3942
#>                    99.95%
#> cp                 1.0158
#> b                  0.7523
#> a                  1.3010
#> cond~~cond         0.3435
#> reaction~~reaction 2.6172
#> pmi~~pmi           2.2568
#> indirect           0.8669
#> direct             1.0158
#> total              1.4999
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0954 0.0874 100 -0.1798 -0.1575 -0.1042 0.2161 0.2633
#> b                  0.4413 0.0740 100  0.2708  0.2712  0.2915 0.5732 0.6171
#> a                  0.2212 0.0871 100  0.0280  0.0583  0.1334 0.4178 0.4531
#> cond~~cond         1.0000 0.0000 100  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7776 0.0654 100  0.6257  0.6284  0.6511 0.8959 0.9178
#> pmi~~pmi           0.9511 0.0476 100  0.7781  0.7943  0.8254 0.9822 0.9954
#> indirect           0.0976 0.0445 100  0.0094  0.0249  0.0438 0.2144 0.2538
#> direct             0.0954 0.0874 100 -0.1798 -0.1575 -0.1042 0.2161 0.2633
#> total              0.1930 0.0964 100 -0.0591 -0.0444 -0.0166 0.3405 0.3977
#>                    99.95%
#> cp                 0.2826
#> b                  0.6212
#> a                  0.4709
#> cond~~cond         1.0000
#> reaction~~reaction 0.9220
#> pmi~~pmi           0.9990
#> indirect           0.2635
#> direct             0.2826
#> total              0.4156