Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.2213 0.2791 20 -0.2949 -0.2889 -0.2622 0.6586 0.6637
#> b                  0.4936 0.1313 20  0.2730  0.2753  0.2856 0.7179 0.7446
#> a                  0.4357 0.1968 20  0.2322  0.2356  0.2509 0.7952 0.7975
#> cond~~cond         0.2489 0.0377 20  0.1830  0.1845  0.1912 0.3263 0.3543
#> reaction~~reaction 1.9210 0.2625 20  1.5518  1.5522  1.5543 2.3786 2.3951
#> pmi~~pmi           1.5201 0.2086 20  1.1008  1.1116  1.1599 1.8660 1.8691
#> reaction~1         0.6248 0.6553 20 -0.8142 -0.7807 -0.6319 1.5042 1.6993
#> pmi~1              5.4278 0.1166 20  5.2165  5.2184  5.2264 5.5945 5.6288
#> cond~1             0.4667 0.0496 20  0.3738  0.3748  0.3789 0.5538 0.5639
#> indirect           0.2151 0.1448 20  0.0951  0.0955  0.0973 0.5564 0.5909
#> direct             0.2213 0.2791 20 -0.2949 -0.2889 -0.2622 0.6586 0.6637
#> total              0.4364 0.3003 20 -0.1338 -0.1209 -0.0635 1.0191 1.0653
#>                    99.95%
#> cp                 0.6648
#> b                  0.7506
#> a                  0.7981
#> cond~~cond         0.3607
#> reaction~~reaction 2.3988
#> pmi~~pmi           1.8698
#> reaction~1         1.7432
#> pmi~1              5.6366
#> cond~1             0.5662
#> indirect           0.5987
#> direct             0.6648
#> total              1.0757
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0722 0.0885 20 -0.1028 -0.1018 -0.0973 0.2025 0.2088
#> b                  0.4041 0.0972 20  0.2410  0.2428  0.2508 0.5620 0.5700
#> a                  0.1736 0.0772 20  0.0941  0.0949  0.0982 0.3242 0.3369
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.8214 0.0761 20  0.6268  0.6298  0.6429 0.9117 0.9223
#> pmi~~pmi           0.9698 0.0327 20  0.8845  0.8864  0.8946 0.9903 0.9910
#> indirect           0.4086 0.0435 20  0.0338  0.0339  0.0343 0.1631 0.1715
#> direct             4.3356 0.0885 20 -0.1028 -0.1018 -0.0973 0.2025 0.2088
#> total              0.9355 0.0907 20 -0.0466 -0.0420 -0.0217 0.3085 0.3301
#>                    99.95%
#> cp                 0.2102
#> b                  0.5718
#> a                  0.3398
#> cond~~cond         1.0000
#> reaction~~reaction 0.9246
#> pmi~~pmi           0.9911
#> indirect           0.1734
#> direct             0.2102
#> total              0.3350

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.2945 0.3098 20 -0.3737 -0.3552 -0.2727 0.6895 0.7415
#> b                  0.4198 0.0921 20  0.2220  0.2258  0.2427 0.5263 0.5381
#> a                  0.3992 0.3006 20 -0.0955 -0.0764  0.0084 0.9429 1.0008
#> cond~~cond         0.2486 0.0376 20  0.1750  0.1766  0.1840 0.3118 0.3123
#> reaction~~reaction 2.0283 0.3384 20  1.4863  1.4887  1.4995 2.7124 2.7886
#> pmi~~pmi           1.4770 0.1685 20  1.1529  1.1534  1.1554 1.7443 1.7466
#> indirect           0.1684 0.1198 20 -0.0478 -0.0406 -0.0087 0.3657 0.3842
#> direct             0.2945 0.3098 20 -0.3737 -0.3552 -0.2727 0.6895 0.7415
#> total              0.4628 0.3167 20 -0.1793 -0.1725 -0.1421 0.9521 1.0099
#>                    99.95%
#> cp                 0.7532
#> b                  0.5407
#> a                  1.0138
#> cond~~cond         0.3124
#> reaction~~reaction 2.8058
#> pmi~~pmi           1.7471
#> indirect           0.3883
#> direct             0.7532
#> total              1.0229
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0241 0.1042 20 -0.1211 -0.1150 -0.0877 0.2416 0.2508
#> b                  0.3748 0.0685 20  0.2215  0.2219  0.2235 0.4260 0.4281
#> a                  0.1612 0.1189 20 -0.0441 -0.0359  0.0004 0.3712 0.3945
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.8560 0.0551 20  0.7437  0.7475  0.7645 0.9437 0.9437
#> pmi~~pmi           0.9740 0.0462 20  0.8402  0.8441  0.8613 0.9977 0.9979
#> indirect           0.0604 0.0400 20 -0.0182 -0.0156 -0.0043 0.1214 0.1319
#> direct             0.0241 0.1042 20 -0.1211 -0.1150 -0.0877 0.2416 0.2508
#> total              0.0846 0.1078 20 -0.0581 -0.0558 -0.0456 0.3237 0.3396
#>                    99.95%
#> cp                 0.2529
#> b                  0.4285
#> a                  0.3997
#> cond~~cond         1.0000
#> reaction~~reaction 0.9437
#> pmi~~pmi           0.9980
#> indirect           0.1343
#> direct             0.2529
#> total              0.3432