Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1613 0.2708 20 -0.3451 -0.3350 -0.2900 0.6205 0.6544
#> b                  0.4962 0.1129 20  0.2293  0.2374  0.2735 0.6502 0.6554
#> a                  0.5403 0.2705 20  0.0150  0.0222  0.0539 0.9764 1.0262
#> cond~~cond         0.2494 0.0316 20  0.2024  0.2045  0.2135 0.3094 0.3110
#> reaction~~reaction 1.9219 0.2337 20  1.4117  1.4264  1.4918 2.3353 2.4852
#> pmi~~pmi           1.7729 0.2170 20  1.3061  1.3245  1.4064 2.1033 2.1567
#> reaction~1         0.5716 0.6250 20 -0.5133 -0.5004 -0.4432 1.7819 2.0271
#> pmi~1              5.3252 0.1741 20  4.9559  4.9645  5.0027 5.5871 5.6099
#> cond~1             0.4830 0.0467 20  0.3820  0.3837  0.3911 0.5497 0.5502
#> indirect           0.2681 0.1359 20  0.0094  0.0125  0.0266 0.5115 0.5397
#> direct             0.1613 0.2708 20 -0.3451 -0.3350 -0.2900 0.6205 0.6544
#> total              0.4294 0.2955 20 -0.2979 -0.2773 -0.1855 0.8111 0.8341
#>                    99.95%
#> cp                 0.6620
#> b                  0.6565
#> a                  1.0374
#> cond~~cond         0.3114
#> reaction~~reaction 2.5189
#> pmi~~pmi           2.1687
#> reaction~1         2.0823
#> pmi~1              5.6150
#> cond~1             0.5503
#> indirect           0.5461
#> direct             0.6620
#> total              0.8393
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0520 0.0922 20 -0.1003 -0.0977 -0.0861 0.2281 0.2484
#> b                  0.4347 0.0810 20  0.2546  0.2563  0.2640 0.5483 0.5486
#> a                  0.1986 0.0960 20  0.0052  0.0076  0.0186 0.3548 0.3584
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7993 0.0629 20  0.6600  0.6610  0.6653 0.8929 0.8999
#> pmi~~pmi           0.9606 0.0389 20  0.8710  0.8715  0.8741 0.9994 0.9999
#> indirect           0.3686 0.0426 20  0.0026  0.0036  0.0077 0.1594 0.1687
#> direct             3.9197 0.0922 20 -0.1003 -0.0977 -0.0861 0.2281 0.2484
#> total              0.9673 0.0993 20 -0.0866 -0.0805 -0.0536 0.2810 0.2823
#>                    99.95%
#> cp                 0.2530
#> b                  0.5487
#> a                  0.3592
#> cond~~cond         1.0000
#> reaction~~reaction 0.9015
#> pmi~~pmi           1.0000
#> indirect           0.1707
#> direct             0.2530
#> total              0.2826

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1595 0.2414 20 -0.3193 -0.3109 -0.2736 0.5857 0.6455
#> b                  0.5233 0.1066 20  0.3707  0.3753  0.3958 0.7281 0.7294
#> a                  0.5081 0.2677 20 -0.1322 -0.1183 -0.0565 0.8718 0.9100
#> cond~~cond         0.2488 0.0331 20  0.2086  0.2092  0.2120 0.3269 0.3376
#> reaction~~reaction 1.9618 0.1927 20  1.5897  1.5920  1.6019 2.2617 2.2976
#> pmi~~pmi           1.7373 0.2625 20  1.3170  1.3221  1.3446 2.1772 2.2400
#> indirect           0.2647 0.1612 20 -0.0893 -0.0803 -0.0402 0.5282 0.5771
#> direct             0.1595 0.2414 20 -0.3193 -0.3109 -0.2736 0.5857 0.6455
#> total              0.4241 0.3351 20 -0.4084 -0.3884 -0.2996 0.9227 1.0249
#>                    99.95%
#> cp                 0.6589
#> b                  0.7297
#> a                  0.9186
#> cond~~cond         0.3400
#> reaction~~reaction 2.3057
#> pmi~~pmi           2.2542
#> indirect           0.5881
#> direct             0.6589
#> total              1.0479
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1030 0.0759 20 -0.0923 -0.0905 -0.0825 0.1821 0.2024
#> b                  0.4459 0.0816 20  0.3396  0.3398  0.3407 0.6065 0.6065
#> a                  0.2165 0.1026 20 -0.0467 -0.0418 -0.0199 0.3354 0.3536
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7707 0.0751 20  0.6112  0.6127  0.6196 0.8615 0.8813
#> pmi~~pmi           0.9531 0.0355 20  0.8720  0.8747  0.8869 0.9998 0.9999
#> indirect           0.0965 0.0487 20 -0.0258 -0.0231 -0.0114 0.1581 0.1651
#> direct             0.1030 0.0759 20 -0.0923 -0.0905 -0.0825 0.1821 0.2024
#> total              0.1995 0.1032 20 -0.1180 -0.1130 -0.0906 0.2783 0.3196
#>                    99.95%
#> cp                 0.2069
#> b                  0.6065
#> a                  0.3577
#> cond~~cond         1.0000
#> reaction~~reaction 0.8857
#> pmi~~pmi           0.9999
#> indirect           0.1667
#> direct             0.2069
#> total              0.3289