Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.2398 0.2931 100 -0.4774 -0.4728 -0.4201 0.8318 0.8752
#> b                  0.5544 0.1052 100  0.2742  0.2991  0.3521 0.7408 0.7656
#> a                  0.6400 0.2373 100  0.0823  0.0958  0.1395 1.0408 1.0916
#> cond~~cond         0.2480 0.0328 100  0.1702  0.1732  0.1819 0.3019 0.3310
#> reaction~~reaction 1.8432 0.2437 100  1.2080  1.2284  1.4161 2.3677 2.4346
#> pmi~~pmi           1.6817 0.2137 100  1.1807  1.2221  1.3255 2.1181 2.2493
#> reaction~1         0.2496 0.6000 100 -0.9222 -0.8914 -0.8061 1.4066 1.6048
#> pmi~1              5.3448 0.1643 100  4.8610  4.8970  5.0892 5.6655 5.7257
#> cond~1             0.4672 0.0466 100  0.3519  0.3567  0.3740 0.5606 0.5874
#> indirect           0.3548 0.1504 100  0.0436  0.0607  0.0829 0.6285 0.7349
#> direct             0.2398 0.2931 100 -0.4774 -0.4728 -0.4201 0.8318 0.8752
#> total              0.5946 0.3166 100 -0.2971 -0.2025 -0.0336 1.1763 1.3146
#>                    99.95%
#> cp                 0.9010
#> b                  0.7737
#> a                  1.1008
#> cond~~cond         0.3412
#> reaction~~reaction 2.4607
#> pmi~~pmi           2.2542
#> reaction~1         1.6433
#> pmi~1              5.7599
#> cond~1             0.6052
#> indirect           0.7478
#> direct             0.9010
#> total              1.3329
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0763 0.0932 100 -0.1677 -0.1595 -0.1427 0.2607 0.2927
#> b                  0.4732 0.0784 100  0.2606  0.2791  0.3174 0.6153 0.6514
#> a                  0.2387 0.0868 100  0.0307  0.0362  0.0498 0.3766 0.4132
#> cond~~cond         1.0000 0.0000 100  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7530 0.0796 100  0.5615  0.5640  0.5836 0.8887 0.9029
#> pmi~~pmi           0.9430 0.0391 100  0.8265  0.8293  0.8582 0.9975 0.9987
#> indirect           0.1596 0.0475 100  0.0132  0.0174  0.0248 0.2028 0.2381
#> direct             4.0024 0.0932 100 -0.1677 -0.1595 -0.1427 0.2607 0.2927
#> total              0.9382 0.0989 100 -0.1069 -0.0726 -0.0107 0.3667 0.4109
#>                    99.95%
#> cp                 0.3056
#> b                  0.6680
#> a                  0.4165
#> cond~~cond         1.0000
#> reaction~~reaction 0.9090
#> pmi~~pmi           0.9990
#> indirect           0.2592
#> direct             0.3056
#> total              0.4143

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.2815 0.2532 100 -0.4010 -0.3010 -0.1555 0.7334 0.8991
#> b                  0.5269 0.0953 100  0.2959  0.3239  0.3720 0.7006 0.7553
#> a                  0.6303 0.2413 100  0.0306  0.0407  0.1739 0.9868 1.0583
#> cond~~cond         0.2487 0.0347 100  0.1683  0.1691  0.1826 0.3126 0.3249
#> reaction~~reaction 1.8303 0.2329 100  1.2897  1.3139  1.3855 2.2990 2.3455
#> pmi~~pmi           1.6487 0.2239 100  1.1277  1.1278  1.2023 2.0121 2.1764
#> indirect           0.3318 0.1429 100  0.0188  0.0232  0.0831 0.5829 0.6270
#> direct             0.2815 0.2532 100 -0.4010 -0.3010 -0.1555 0.7334 0.8991
#> total              0.6133 0.2883 100 -0.1663 -0.1213  0.0616 1.1444 1.2822
#>                    99.95%
#> cp                 0.9335
#> b                  0.7639
#> a                  1.0680
#> cond~~cond         0.3297
#> reaction~~reaction 2.3617
#> pmi~~pmi           2.2454
#> indirect           0.6382
#> direct             0.9335
#> total              1.2980
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0333 0.0827 100 -0.1509 -0.1122 -0.0545 0.2481 0.2817
#> b                  0.4675 0.0694 100  0.2900  0.2968  0.3189 0.5659 0.5939
#> a                  0.1845 0.0884 100  0.0129  0.0166  0.0654 0.3749 0.3928
#> cond~~cond         1.0000 0.0000 100  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7745 0.0630 100  0.6081  0.6133  0.6538 0.8732 0.8781
#> pmi~~pmi           0.9660 0.0385 100  0.8454  0.8457  0.8595 0.9957 0.9997
#> indirect           0.0863 0.0433 100  0.0058  0.0074  0.0262 0.1777 0.1859
#> direct             0.0333 0.0827 100 -0.1509 -0.1122 -0.0545 0.2481 0.2817
#> total              0.1196 0.0915 100 -0.0626 -0.0451  0.0194 0.3695 0.3940
#>                    99.95%
#> cp                 0.2972
#> b                  0.5980
#> a                  0.3931
#> cond~~cond         1.0000
#> reaction~~reaction 0.8785
#> pmi~~pmi           0.9998
#> indirect           0.1897
#> direct             0.2972
#> total              0.4022