Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.2140 0.2738 20 -0.2987 -0.2877 -0.2390 0.7094 0.7548
#> b                  0.4822 0.0671 20  0.4079  0.4082  0.4096 0.6062 0.6089
#> a                  0.5170 0.2108 20  0.1609  0.1747  0.2361 0.9186 0.9197
#> cond~~cond         0.2484 0.0258 20  0.1801  0.1837  0.1996 0.2899 0.2983
#> reaction~~reaction 1.6993 0.2150 20  1.3762  1.3775  1.3834 2.1513 2.1971
#> pmi~~pmi           1.7037 0.1994 20  1.3412  1.3412  1.3412 1.9909 2.0558
#> reaction~1         0.6596 0.4559 20 -0.1733 -0.1582 -0.0913 1.3337 1.3700
#> pmi~1              5.3695 0.2131 20  4.9687  4.9729  4.9918 5.7115 5.7123
#> cond~1             0.4514 0.0472 20  0.3619  0.3629  0.3669 0.5073 0.5099
#> indirect           0.2493 0.1035 20  0.0813  0.0868  0.1112 0.4247 0.4267
#> direct             0.2140 0.2738 20 -0.2987 -0.2877 -0.2390 0.7094 0.7548
#> total              0.4632 0.3157 20 -0.0706 -0.0593 -0.0093 1.0386 1.0794
#>                    99.95%
#> cp                 0.7650
#> b                  0.6095
#> a                  0.9200
#> cond~~cond         0.3002
#> reaction~~reaction 2.2074
#> pmi~~pmi           2.0704
#> reaction~1         1.3782
#> pmi~1              5.7125
#> cond~1             0.5104
#> indirect           0.4271
#> direct             0.7650
#> total              1.0886
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0728 0.0922 20 -0.0974 -0.0942 -0.0797 0.2395 0.2639
#> b                  0.4376 0.0624 20  0.3394  0.3406  0.3457 0.5452 0.5520
#> a                  0.1937 0.0813 20  0.0596  0.0639  0.0831 0.3501 0.3643
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7909 0.0617 20  0.6636  0.6653  0.6731 0.8661 0.8733
#> pmi~~pmi           0.9625 0.0355 20  0.8649  0.8672  0.8771 0.9925 0.9957
#> indirect           0.4500 0.0348 20  0.0277  0.0296  0.0381 0.1411 0.1414
#> direct             4.0359 0.0922 20 -0.0974 -0.0942 -0.0797 0.2395 0.2639
#> total              0.9058 0.1060 20 -0.0230 -0.0193 -0.0026 0.3500 0.3772
#>                    99.95%
#> cp                 0.2694
#> b                  0.5535
#> a                  0.3675
#> cond~~cond         1.0000
#> reaction~~reaction 0.8749
#> pmi~~pmi           0.9964
#> indirect           0.1415
#> direct             0.2694
#> total              0.3833

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.2609 0.2864 20 -0.3017 -0.2938 -0.2587 0.7122 0.7335
#> b                  0.4500 0.0984 20  0.2371  0.2405  0.2553 0.5627 0.5715
#> a                  0.4288 0.2758 20 -0.1048 -0.1023 -0.0913 0.8243 0.9287
#> cond~~cond         0.2467 0.0269 20  0.1829  0.1858  0.1986 0.2842 0.2855
#> reaction~~reaction 1.7410 0.1769 20  1.5157  1.5166  1.5208 2.0863 2.1291
#> pmi~~pmi           1.7289 0.2529 20  1.2232  1.2246  1.2309 2.0967 2.1195
#> indirect           0.1923 0.1385 20 -0.0441 -0.0433 -0.0399 0.4305 0.4985
#> direct             0.2609 0.2864 20 -0.3017 -0.2938 -0.2587 0.7122 0.7335
#> total              0.4532 0.3000 20 -0.0844 -0.0751 -0.0339 1.0328 1.0833
#>                    99.95%
#> cp                 0.7383
#> b                  0.5735
#> a                  0.9521
#> cond~~cond         0.2857
#> reaction~~reaction 2.1387
#> pmi~~pmi           2.1247
#> indirect           0.5138
#> direct             0.7383
#> total              1.0947
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0908 0.0969 20 -0.1017 -0.0990 -0.0870 0.2488 0.2684
#> b                  0.4134 0.0837 20  0.2230  0.2234  0.2253 0.4932 0.5016
#> a                  0.2081 0.1137 20 -0.0382 -0.0374 -0.0334 0.3519 0.3956
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.8053 0.0603 20  0.7382  0.7392  0.7438 0.9234 0.9274
#> pmi~~pmi           0.9567 0.0402 20  0.8355  0.8424  0.8729 0.9998 1.0000
#> indirect           0.0860 0.0449 20 -0.0150 -0.0149 -0.0141 0.1379 0.1569
#> direct             0.0908 0.0969 20 -0.1017 -0.0990 -0.0870 0.2488 0.2684
#> total              0.1769 0.0976 20 -0.0283 -0.0252 -0.0113 0.3289 0.3406
#>                    99.95%
#> cp                 0.2729
#> b                  0.5035
#> a                  0.4054
#> cond~~cond         1.0000
#> reaction~~reaction 0.9283
#> pmi~~pmi           1.0000
#> indirect           0.1611
#> direct             0.2729
#> total              0.3433