Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.2997 0.2848 100 -0.4241 -0.4142 -0.3008 0.8293 0.9041
#> b                  0.5523 0.0931 100  0.3708  0.3763  0.3876 0.7233 0.7756
#> a                  0.3910 0.2379 100 -0.2010 -0.1776 -0.1331 0.7986 0.9669
#> cond~~cond         0.2480 0.0356 100  0.1641  0.1679  0.1800 0.3099 0.3212
#> reaction~~reaction 1.8918 0.2677 100  1.2288  1.2628  1.3408 2.3242 2.4310
#> pmi~~pmi           1.6927 0.2384 100  1.0102  1.1170  1.2767 2.1479 2.2305
#> reaction~1         0.2642 0.5438 100 -1.2521 -1.0975 -0.8606 1.1246 1.3170
#> pmi~1              5.3586 0.1641 100  4.9932  5.0017  5.0211 5.7230 5.7611
#> cond~1             0.4592 0.0460 100  0.3405  0.3542  0.3794 0.5448 0.5535
#> indirect           0.2159 0.1483 100 -0.1073 -0.0916 -0.0670 0.5281 0.6216
#> direct             0.2997 0.2848 100 -0.4241 -0.4142 -0.3008 0.8293 0.9041
#> total              0.5156 0.2948 100 -0.1273 -0.0862 -0.0319 1.0944 1.2374
#>                    99.95%
#> cp                 0.9378
#> b                  0.8180
#> a                  1.0201
#> cond~~cond         0.3279
#> reaction~~reaction 2.5113
#> pmi~~pmi           2.2428
#> reaction~1         1.4272
#> pmi~1              5.7650
#> cond~1             0.5555
#> indirect           0.6381
#> direct             0.9378
#> total              1.3141
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0949 0.0859 100 -0.1332 -0.1313 -0.0857 0.2446 0.3006
#> b                  0.4619 0.0672 100  0.3348  0.3403  0.3549 0.5907 0.6298
#> a                  0.1480 0.0881 100 -0.0667 -0.0586 -0.0470 0.3133 0.3570
#> cond~~cond         1.0000 0.0000 100  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7647 0.0682 100  0.5650  0.5828  0.6123 0.8674 0.8761
#> pmi~~pmi           0.9781 0.0272 100  0.8545  0.8718  0.9018 0.9992 0.9996
#> indirect           0.1680 0.0454 100 -0.0266 -0.0250 -0.0213 0.1649 0.1954
#> direct             4.0733 0.0859 100 -0.1332 -0.1313 -0.0857 0.2446 0.3006
#> total              0.9222 0.0874 100 -0.0354 -0.0250 -0.0095 0.3167 0.3898
#>                    99.95%
#> cp                 0.3123
#> b                  0.6574
#> a                  0.3812
#> cond~~cond         1.0000
#> reaction~~reaction 0.8781
#> pmi~~pmi           0.9997
#> indirect           0.2065
#> direct             0.3123
#> total              0.3923

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 100L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.3697 0.2911 100 -0.3575 -0.2659 -0.1441 0.9310 1.0441
#> b                  0.5368 0.1030 100  0.2516  0.2857  0.3770 0.7403 0.7833
#> a                  0.3280 0.2496 100 -0.2577 -0.1926 -0.0993 0.9296 1.0252
#> cond~~cond         0.2468 0.0365 100  0.1570  0.1600  0.1699 0.3057 0.3194
#> reaction~~reaction 2.0260 0.2917 100  1.2595  1.2913  1.4767 2.5261 2.7420
#> pmi~~pmi           1.6809 0.2246 100  1.0894  1.0899  1.1695 2.0737 2.1074
#> indirect           0.1757 0.1381 100 -0.0977 -0.0848 -0.0508 0.4713 0.6942
#> direct             0.3697 0.2911 100 -0.3575 -0.2659 -0.1441 0.9310 1.0441
#> total              0.5453 0.3057 100 -0.1754 -0.1542 -0.0409 1.1091 1.2291
#>                    99.95%
#> cp                 1.0723
#> b                  0.7865
#> a                  1.0555
#> cond~~cond         0.3226
#> reaction~~reaction 2.8992
#> pmi~~pmi           2.1235
#> indirect           0.7363
#> direct             1.0723
#> total              1.2944
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se   R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1294 0.0892 100 -0.1285 -0.0908 -0.0410 0.2802 0.3359
#> b                  0.4765 0.0801 100  0.2329  0.2525  0.2873 0.6044 0.6286
#> a                  0.1093 0.0959 100 -0.0992 -0.0740 -0.0354 0.3462 0.3897
#> cond~~cond         1.0000 0.0000 100  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7427 0.0691 100  0.5850  0.5937  0.6330 0.8875 0.9093
#> pmi~~pmi           0.9880 0.0329 100  0.8238  0.8470  0.8801 0.9997 1.0000
#> indirect           0.0521 0.0446 100 -0.0346 -0.0269 -0.0157 0.1477 0.2270
#> direct             0.1294 0.0892 100 -0.1285 -0.0908 -0.0410 0.2802 0.3359
#> total              0.1815 0.0944 100 -0.0538 -0.0474 -0.0124 0.3233 0.3796
#>                    99.95%
#> cp                 0.3738
#> b                  0.6408
#> a                  0.4195
#> cond~~cond         1.0000
#> reaction~~reaction 0.9198
#> pmi~~pmi           1.0000
#> indirect           0.2550
#> direct             0.3738
#> total              0.4034