Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0796 0.2512 20 -0.4169 -0.4154 -0.4090 0.3549 0.3744
#> b                  0.5441 0.0800 20  0.4666  0.4673  0.4701 0.7525 0.7590
#> a                  0.4569 0.2696 20  0.0349  0.0448  0.0889 0.9775 1.0113
#> cond~~cond         0.2496 0.0310 20  0.1910  0.1914  0.1934 0.2927 0.2946
#> reaction~~reaction 1.8509 0.2328 20  1.3004  1.3125  1.3662 2.2037 2.2518
#> pmi~~pmi           1.8038 0.1774 20  1.4870  1.4920  1.5145 2.1370 2.1489
#> reaction~1         0.3773 0.5170 20 -1.0421 -1.0006 -0.8165 0.9068 0.9460
#> pmi~1              5.3580 0.1713 20  4.9525  4.9613  5.0008 5.6417 5.6785
#> cond~1             0.4879 0.0533 20  0.3975  0.3980  0.4006 0.5602 0.5683
#> indirect           0.2486 0.1591 20  0.0264  0.0324  0.0592 0.5974 0.6051
#> direct             0.0796 0.2512 20 -0.4169 -0.4154 -0.4090 0.3549 0.3744
#> total              0.3282 0.2779 20 -0.1669 -0.1582 -0.1195 0.7625 0.8188
#>                    99.95%
#> cp                 0.3787
#> b                  0.7604
#> a                  1.0189
#> cond~~cond         0.2951
#> reaction~~reaction 2.2626
#> pmi~~pmi           2.1516
#> reaction~1         0.9548
#> pmi~1              5.6868
#> cond~1             0.5701
#> indirect           0.6068
#> direct             0.3787
#> total              0.8315
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0256 0.0818 20 -0.1501 -0.1497 -0.1480 0.1160 0.1185
#> b                  0.4773 0.0594 20  0.4198  0.4202  0.4220 0.6233 0.6377
#> a                  0.1675 0.0910 20  0.0138  0.0183  0.0380 0.3447 0.3500
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7674 0.0627 20  0.5885  0.5898  0.5954 0.8011 0.8085
#> pmi~~pmi           0.9719 0.0371 20  0.8767  0.8775  0.8811 0.9979 0.9994
#> indirect           0.2430 0.0494 20  0.0088  0.0109  0.0202 0.1849 0.1889
#> direct             3.9330 0.0818 20 -0.1501 -0.1497 -0.1480 0.1160 0.1185
#> total              0.9767 0.0860 20 -0.0583 -0.0551 -0.0412 0.2276 0.2364
#>                    99.95%
#> cp                 0.1191
#> b                  0.6409
#> a                  0.3512
#> cond~~cond         1.0000
#> reaction~~reaction 0.8102
#> pmi~~pmi           0.9998
#> indirect           0.1898
#> direct             0.1191
#> total              0.2384

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1798 0.3482 20 -0.4753 -0.4485 -0.3294 0.8598 0.9291
#> b                  0.5378 0.0913 20  0.3223  0.3240  0.3318 0.6758 0.6801
#> a                  0.4298 0.2955 20 -0.2525 -0.2231 -0.0923 0.9129 0.9705
#> cond~~cond         0.2499 0.0338 20  0.1867  0.1895  0.2020 0.3101 0.3172
#> reaction~~reaction 1.8774 0.2258 20  1.4215  1.4354  1.4970 2.2571 2.2592
#> pmi~~pmi           1.7633 0.1948 20  1.3547  1.3586  1.3759 2.0122 2.0162
#> indirect           0.2321 0.1562 20 -0.1175 -0.1049 -0.0489 0.5473 0.5560
#> direct             0.1798 0.3482 20 -0.4753 -0.4485 -0.3294 0.8598 0.9291
#> total              0.4119 0.4126 20 -0.2511 -0.2509 -0.2502 1.1853 1.2804
#>                    99.95%
#> cp                 0.9446
#> b                  0.6811
#> a                  0.9834
#> cond~~cond         0.3188
#> reaction~~reaction 2.2596
#> pmi~~pmi           2.0171
#> indirect           0.5580
#> direct             0.9446
#> total              1.3018
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0253 0.1181 20 -0.1600 -0.1502 -0.1067 0.3024 0.3227
#> b                  0.5144 0.0694 20  0.2737  0.2792  0.3037 0.5577 0.5747
#> a                  0.1640 0.1056 20 -0.0908 -0.0797 -0.0307 0.3233 0.3308
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7305 0.0614 20  0.6547  0.6566  0.6653 0.8712 0.8868
#> pmi~~pmi           0.9731 0.0329 20  0.8894  0.8905  0.8954 0.9981 0.9985
#> indirect           0.0843 0.0473 20 -0.0396 -0.0353 -0.0162 0.1565 0.1577
#> direct             0.0253 0.1181 20 -0.1600 -0.1502 -0.1067 0.3024 0.3227
#> total              0.1096 0.1390 20 -0.0846 -0.0845 -0.0842 0.4168 0.4447
#>                    99.95%
#> cp                 0.3272
#> b                  0.5786
#> a                  0.3325
#> cond~~cond         1.0000
#> reaction~~reaction 0.8903
#> pmi~~pmi           0.9986
#> indirect           0.1580
#> direct             0.3272
#> total              0.4509