Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.4474 0.1656 20  0.2460  0.2465  0.2490 0.7715 0.8144
#> b                  0.4830 0.1203 20  0.2155  0.2274  0.2805 0.6871 0.6925
#> a                  0.4632 0.2140 20  0.1384  0.1475  0.1880 0.9220 0.9254
#> cond~~cond         0.2479 0.0341 20  0.1752  0.1763  0.1809 0.2952 0.3106
#> reaction~~reaction 1.8551 0.1889 20  1.4214  1.4334  1.4870 2.1034 2.1136
#> pmi~~pmi           1.7248 0.2303 20  1.3392  1.3497  1.3963 2.2231 2.3281
#> reaction~1         0.5895 0.6361 20 -0.3742 -0.3692 -0.3467 1.7271 1.9782
#> pmi~1              5.3762 0.1680 20  4.9785  4.9816  4.9955 5.5608 5.5733
#> cond~1             0.4636 0.0523 20  0.3761  0.3786  0.3898 0.5685 0.5833
#> indirect           0.2237 0.1044 20  0.0504  0.0516  0.0568 0.4321 0.4424
#> direct             0.4474 0.1656 20  0.2460  0.2465  0.2490 0.7715 0.8144
#> total              0.6712 0.1824 20  0.3863  0.3892  0.4019 0.9780 1.0017
#>                    99.95%
#> cp                 0.8241
#> b                  0.6937
#> a                  0.9262
#> cond~~cond         0.3141
#> reaction~~reaction 2.1159
#> pmi~~pmi           2.3517
#> reaction~1         2.0347
#> pmi~1              5.5761
#> cond~1             0.5867
#> indirect           0.4447
#> direct             0.8241
#> total              1.0070
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R  0.05%   0.5%   2.5%  97.5%  99.5% 99.95%
#> cp                 0.1447 0.0601 20 0.0748 0.0757 0.0799 0.2635 0.2971 0.3047
#> b                  0.4184 0.0870 20 0.2049 0.2168 0.2696 0.5743 0.5874 0.5904
#> a                  0.1729 0.0778 20 0.0506 0.0550 0.0742 0.3347 0.3407 0.3421
#> cond~~cond         1.0000 0.0000 20 1.0000 1.0000 1.0000 1.0000 1.0000 1.0000
#> reaction~~reaction 0.7830 0.0741 20 0.6169 0.6216 0.6427 0.8979 0.9323 0.9401
#> pmi~~pmi           0.9701 0.0322 20 0.8830 0.8839 0.8879 0.9939 0.9968 0.9974
#> indirect           0.3830 0.0331 20 0.0186 0.0188 0.0195 0.1364 0.1424 0.1438
#> direct             4.0320 0.0601 20 0.0748 0.0757 0.0799 0.2635 0.2971 0.3047
#> total              0.9313 0.0646 20 0.1174 0.1188 0.1253 0.3174 0.3228 0.3240

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.3919 0.2815 20 -0.0334 -0.0297 -0.0131 0.9425 1.0727
#> b                  0.4745 0.0892 20  0.3621  0.3627  0.3651 0.6536 0.6550
#> a                  0.4334 0.2363 20 -0.0270 -0.0053  0.0911 0.9315 1.0207
#> cond~~cond         0.2485 0.0290 20  0.1946  0.1959  0.2018 0.3003 0.3068
#> reaction~~reaction 1.9056 0.2698 20  1.3517  1.3745  1.4759 2.4366 2.5653
#> pmi~~pmi           1.6885 0.2235 20  1.3298  1.3362  1.3647 2.1235 2.1894
#> indirect           0.2050 0.1468 20 -0.0130 -0.0046  0.0331 0.5542 0.6545
#> direct             0.3919 0.2815 20 -0.0334 -0.0297 -0.0131 0.9425 1.0727
#> total              0.5970 0.3417 20 -0.0455 -0.0249  0.0668 1.2219 1.3034
#>                    99.95%
#> cp                 1.1020
#> b                  0.6554
#> a                  1.0408
#> cond~~cond         0.3082
#> reaction~~reaction 2.5943
#> pmi~~pmi           2.2042
#> indirect           0.6770
#> direct             1.1020
#> total              1.3217
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1243 0.0928 20 -0.0111 -0.0099 -0.0045 0.3061 0.3494
#> b                  0.4534 0.0672 20  0.2785  0.2809  0.2917 0.5262 0.5440
#> a                  0.1524 0.0878 20 -0.0101 -0.0023  0.0325 0.3328 0.3727
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7618 0.0661 20  0.6379  0.6423  0.6615 0.8931 0.8976
#> pmi~~pmi           0.9768 0.0342 20  0.8543  0.8602  0.8865 0.9969 0.9993
#> indirect           0.0691 0.0455 20 -0.0043 -0.0015  0.0113 0.1761 0.2031
#> direct             0.1243 0.0928 20 -0.0111 -0.0099 -0.0045 0.3061 0.3494
#> total              0.1934 0.1121 20 -0.0152 -0.0092  0.0178 0.3901 0.4232
#>                    99.95%
#> cp                 0.3592
#> b                  0.5480
#> a                  0.3816
#> cond~~cond         1.0000
#> reaction~~reaction 0.8986
#> pmi~~pmi           0.9998
#> indirect           0.2092
#> direct             0.3592
#> total              0.4306