Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.2499 0.2877 20 -0.3387 -0.3305 -0.2939 0.6982 0.7547
#> b                  0.5416 0.1181 20  0.2900  0.2944  0.3138 0.7122 0.7333
#> a                  0.4454 0.2729 20 -0.1481 -0.1305 -0.0521 0.9463 0.9893
#> cond~~cond         0.2479 0.0292 20  0.2014  0.2016  0.2024 0.3025 0.3151
#> reaction~~reaction 1.9064 0.2100 20  1.3941  1.4123  1.4933 2.1576 2.1651
#> pmi~~pmi           1.7278 0.1821 20  1.3400  1.3429  1.3557 1.9773 2.0358
#> reaction~1         0.3433 0.6181 20 -0.5406 -0.5400 -0.5370 1.4993 1.5308
#> pmi~1              5.3935 0.2020 20  5.0636  5.0750  5.1253 5.8605 5.9804
#> cond~1             0.4587 0.0548 20  0.3394  0.3426  0.3568 0.5428 0.5543
#> indirect           0.2412 0.1513 20 -0.0735 -0.0645 -0.0242 0.5471 0.6200
#> direct             0.2499 0.2877 20 -0.3387 -0.3305 -0.2939 0.6982 0.7547
#> total              0.4911 0.3550 20 -0.2251 -0.2173 -0.1830 0.9319 1.0068
#>                    99.95%
#> cp                 0.7674
#> b                  0.7380
#> a                  0.9990
#> cond~~cond         0.3180
#> reaction~~reaction 2.1667
#> pmi~~pmi           2.0490
#> reaction~1         1.5379
#> pmi~1              6.0074
#> cond~1             0.5569
#> indirect           0.6364
#> direct             0.7674
#> total              1.0237
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0791 0.0932 20 -0.1028 -0.1006 -0.0909 0.2295 0.2448
#> b                  0.4591 0.0832 20  0.2624  0.2671  0.2878 0.5609 0.5617
#> a                  0.1664 0.0930 20 -0.0517 -0.0453 -0.0167 0.3195 0.3213
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7709 0.0638 20  0.6826  0.6828  0.6836 0.8798 0.8845
#> pmi~~pmi           0.9723 0.0303 20  0.8965  0.8968  0.8979 0.9986 0.9993
#> indirect           0.2183 0.0415 20 -0.0247 -0.0217 -0.0084 0.1441 0.1642
#> direct             4.0461 0.0932 20 -0.1028 -0.1006 -0.0909 0.2295 0.2448
#> total              0.9213 0.1115 20 -0.0683 -0.0664 -0.0576 0.3044 0.3262
#>                    99.95%
#> cp                 0.2482
#> b                  0.5619
#> a                  0.3217
#> cond~~cond         1.0000
#> reaction~~reaction 0.8855
#> pmi~~pmi           0.9995
#> indirect           0.1687
#> direct             0.2482
#> total              0.3311

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.2014 0.2417 20 -0.3439 -0.3283 -0.2589 0.5212 0.5541
#> b                  0.5312 0.0970 20  0.3740  0.3774  0.3926 0.6841 0.6906
#> a                  0.4656 0.2056 20  0.1370  0.1449  0.1799 0.8642 0.9424
#> cond~~cond         0.2487 0.0303 20  0.1566  0.1585  0.1673 0.2776 0.2842
#> reaction~~reaction 1.9028 0.2604 20  1.3740  1.3873  1.4461 2.3346 2.4231
#> pmi~~pmi           1.7393 0.3146 20  1.1971  1.2088  1.2609 2.3087 2.3986
#> indirect           0.2468 0.1116 20  0.0807  0.0828  0.0919 0.4446 0.4704
#> direct             0.2014 0.2417 20 -0.3439 -0.3283 -0.2589 0.5212 0.5541
#> total              0.4481 0.2134 20 -0.1010 -0.0914 -0.0487 0.7342 0.7729
#>                    99.95%
#> cp                 0.5615
#> b                  0.6920
#> a                  0.9599
#> cond~~cond         0.2857
#> reaction~~reaction 2.4430
#> pmi~~pmi           2.4188
#> indirect           0.4762
#> direct             0.5615
#> total              0.7816
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0797 0.0784 20 -0.1194 -0.1136 -0.0879 0.1667 0.1791
#> b                  0.4594 0.0842 20  0.2930  0.2953  0.3057 0.5810 0.5904
#> a                  0.1471 0.0814 20  0.0517  0.0546  0.0673 0.3294 0.3613
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7718 0.0706 20  0.6438  0.6460  0.6559 0.8995 0.9027
#> pmi~~pmi           0.9784 0.0334 20  0.8642  0.8689  0.8897 0.9952 0.9969
#> indirect           0.0676 0.0350 20  0.0233  0.0243  0.0288 0.1460 0.1519
#> direct             0.0797 0.0784 20 -0.1194 -0.1136 -0.0879 0.1667 0.1791
#> total              0.1473 0.0665 20 -0.0351 -0.0318 -0.0171 0.2340 0.2438
#>                    99.95%
#> cp                 0.1818
#> b                  0.5925
#> a                  0.3684
#> cond~~cond         1.0000
#> reaction~~reaction 0.9034
#> pmi~~pmi           0.9973
#> indirect           0.1532
#> direct             0.1818
#> total              0.2460