Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.2259 0.2887 20 -0.2763 -0.2649 -0.2146 0.7053 0.7607
#> b                  0.5403 0.0690 20  0.4266  0.4299  0.4442 0.6643 0.6654
#> a                  0.3373 0.2261 20 -0.0146 -0.0112  0.0041 0.6618 0.6629
#> cond~~cond         0.2482 0.0271 20  0.1979  0.1997  0.2076 0.3028 0.3139
#> reaction~~reaction 1.8673 0.2162 20  1.3264  1.3504  1.4571 2.1993 2.2076
#> pmi~~pmi           1.7354 0.2045 20  1.3439  1.3452  1.3507 2.0306 2.0845
#> reaction~1         0.3980 0.4876 20 -0.4731 -0.4532 -0.3648 1.1018 1.1475
#> pmi~1              5.4060 0.2047 20  5.0956  5.0963  5.0994 5.7332 5.7402
#> cond~1             0.4589 0.0474 20  0.4019  0.4021  0.4034 0.5457 0.5509
#> indirect           0.1822 0.1231 20 -0.0080 -0.0059  0.0036 0.3602 0.3642
#> direct             0.2259 0.2887 20 -0.2763 -0.2649 -0.2146 0.7053 0.7607
#> total              0.4081 0.3548 20 -0.2593 -0.2461 -0.1872 1.0655 1.1250
#>                    99.95%
#> cp                 0.7732
#> b                  0.6657
#> a                  0.6632
#> cond~~cond         0.3164
#> reaction~~reaction 2.2095
#> pmi~~pmi           2.0967
#> reaction~1         1.1577
#> pmi~1              5.7417
#> cond~1             0.5520
#> indirect           0.3651
#> direct             0.7732
#> total              1.1383
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0724 0.0883 20 -0.0820 -0.0790 -0.0656 0.2044 0.2079
#> b                  0.4617 0.0565 20  0.3492  0.3530  0.3698 0.5539 0.5598
#> a                  0.1265 0.0818 20 -0.0058 -0.0045  0.0011 0.2443 0.2478
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7731 0.0541 20  0.6791  0.6801  0.6845 0.8408 0.8438
#> pmi~~pmi           0.9840 0.0210 20  0.9382  0.9386  0.9403 0.9999 1.0000
#> indirect           0.2561 0.0373 20 -0.0026 -0.0019  0.0010 0.1114 0.1123
#> direct             4.0707 0.0883 20 -0.0820 -0.0790 -0.0656 0.2044 0.2079
#> total              0.9212 0.1065 20 -0.0770 -0.0730 -0.0557 0.3098 0.3116
#>                    99.95%
#> cp                 0.2087
#> b                  0.5611
#> a                  0.2486
#> cond~~cond         1.0000
#> reaction~~reaction 0.8445
#> pmi~~pmi           1.0000
#> indirect           0.1125
#> direct             0.2087
#> total              0.3120

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.2025 0.3419 20 -0.2425 -0.2387 -0.2217 0.9368 1.2057
#> b                  0.5427 0.0841 20  0.2974  0.3062  0.3454 0.6377 0.6449
#> a                  0.3092 0.1977 20  0.0046  0.0047  0.0050 0.6039 0.6324
#> cond~~cond         0.2478 0.0318 20  0.1723  0.1743  0.1832 0.2958 0.3022
#> reaction~~reaction 1.8905 0.2735 20  1.2693  1.2873  1.3672 2.3479 2.4055
#> pmi~~pmi           1.7089 0.2300 20  1.3652  1.3690  1.3863 2.1074 2.1258
#> indirect           0.1685 0.0928 20  0.0025  0.0025  0.0027 0.2786 0.2931
#> direct             0.2025 0.3419 20 -0.2425 -0.2387 -0.2217 0.9368 1.2057
#> total              0.3710 0.3586 20 -0.1953 -0.1903 -0.1682 1.1283 1.3486
#>                    99.95%
#> cp                 1.2662
#> b                  0.6465
#> a                  0.6388
#> cond~~cond         0.3037
#> reaction~~reaction 2.4184
#> pmi~~pmi           2.1299
#> indirect           0.2964
#> direct             1.2662
#> total              1.3982
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0697 0.1049 20 -0.0836 -0.0824 -0.0770 0.2825 0.3468
#> b                  0.4779 0.0841 20  0.2643  0.2666  0.2768 0.5894 0.6318
#> a                  0.1176 0.0764 20  0.0016  0.0017  0.0019 0.2334 0.2460
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7589 0.0743 20  0.5878  0.5946  0.6247 0.9021 0.9053
#> pmi~~pmi           0.9862 0.0181 20  0.9381  0.9394  0.9452 1.0000 1.0000
#> indirect           0.0562 0.0307 20  0.0009  0.0009  0.0009 0.0929 0.0934
#> direct             0.0697 0.1049 20 -0.0836 -0.0824 -0.0770 0.2825 0.3468
#> total              0.1259 0.1099 20 -0.0684 -0.0666 -0.0585 0.3442 0.3889
#>                    99.95%
#> cp                 0.3612
#> b                  0.6413
#> a                  0.2488
#> cond~~cond         1.0000
#> reaction~~reaction 0.9061
#> pmi~~pmi           1.0000
#> indirect           0.0935
#> direct             0.3612
#> total              0.3990