Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.3234 0.4224 20 -0.4607 -0.4424 -0.3611 0.9464 0.9927
#> b                  0.4592 0.0799 20  0.2891  0.2908  0.2987 0.5866 0.5880
#> a                  0.3410 0.2833 20 -0.5122 -0.4705 -0.2852 0.6782 0.7222
#> cond~~cond         0.2483 0.0248 20  0.1922  0.1941  0.2024 0.2865 0.2910
#> reaction~~reaction 2.0168 0.2633 20  1.4380  1.4417  1.4585 2.3213 2.3558
#> pmi~~pmi           1.5962 0.1651 20  1.2113  1.2226  1.2728 1.8260 1.8317
#> reaction~1         0.7511 0.4817 20 -0.0069 -0.0016  0.0219 1.5044 1.5690
#> pmi~1              5.4928 0.1746 20  5.2270  5.2293  5.2397 5.8200 5.8583
#> cond~1             0.4632 0.0494 20  0.3679  0.3711  0.3854 0.5529 0.5673
#> indirect           0.1566 0.1261 20 -0.1586 -0.1459 -0.0894 0.3553 0.3897
#> direct             0.3234 0.4224 20 -0.4607 -0.4424 -0.3611 0.9464 0.9927
#> total              0.4800 0.4634 20 -0.3145 -0.3054 -0.2649 1.2524 1.3725
#>                    99.95%
#> cp                 1.0031
#> b                  0.5884
#> a                  0.7321
#> cond~~cond         0.2920
#> reaction~~reaction 2.3635
#> pmi~~pmi           1.8330
#> reaction~1         1.5836
#> pmi~1              5.8669
#> cond~1             0.5705
#> indirect           0.3974
#> direct             1.0031
#> total              1.3995
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1038 0.1356 20 -0.1662 -0.1586 -0.1247 0.3076 0.3280
#> b                  0.3770 0.0671 20  0.2469  0.2485  0.2555 0.4927 0.4945
#> a                  0.1333 0.1084 20 -0.1894 -0.1741 -0.1062 0.2671 0.2868
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.8366 0.0723 20  0.5973  0.6072  0.6512 0.8929 0.9023
#> pmi~~pmi           0.9822 0.0225 20  0.9151  0.9175  0.9280 0.9999 0.9999
#> indirect           0.4837 0.0406 20 -0.0502 -0.0462 -0.0285 0.1141 0.1285
#> direct             4.3088 0.1356 20 -0.1662 -0.1586 -0.1247 0.3076 0.3280
#> total              0.9296 0.1469 20 -0.1134 -0.1090 -0.0893 0.3854 0.4492
#>                    99.95%
#> cp                 0.3326
#> b                  0.4949
#> a                  0.2913
#> cond~~cond         1.0000
#> reaction~~reaction 0.9044
#> pmi~~pmi           0.9999
#> indirect           0.1317
#> direct             0.3326
#> total              0.4636

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.2803 0.2605 20 -0.2352 -0.2217 -0.1614 0.6833 0.7365
#> b                  0.4349 0.1046 20  0.2820  0.2823  0.2840 0.6420 0.6702
#> a                  0.2690 0.2147 20 -0.2300 -0.2213 -0.1824 0.5677 0.5693
#> cond~~cond         0.2483 0.0315 20  0.1922  0.1940  0.2022 0.3133 0.3222
#> reaction~~reaction 2.0630 0.3238 20  1.1212  1.1654  1.3618 2.5183 2.6068
#> pmi~~pmi           1.6010 0.2451 20  0.9990  1.0178  1.1012 1.9518 1.9818
#> indirect           0.1170 0.1012 20 -0.0761 -0.0736 -0.0627 0.2825 0.2855
#> direct             0.2803 0.2605 20 -0.2352 -0.2217 -0.1614 0.6833 0.7365
#> total              0.3973 0.2861 20 -0.0707 -0.0668 -0.0497 0.8188 0.8373
#>                    99.95%
#> cp                 0.7484
#> b                  0.6765
#> a                  0.5696
#> cond~~cond         0.3242
#> reaction~~reaction 2.6267
#> pmi~~pmi           1.9885
#> indirect           0.2861
#> direct             0.7484
#> total              0.8415
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0657 0.0892 20 -0.0747 -0.0704 -0.0513 0.2414 0.2883
#> b                  0.3820 0.0815 20  0.2282  0.2290  0.2325 0.4996 0.5074
#> a                  0.1184 0.0902 20 -0.1129 -0.1076 -0.0838 0.2315 0.2316
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.8438 0.0704 20  0.6957  0.6964  0.6995 0.9465 0.9482
#> pmi~~pmi           0.9860 0.0166 20  0.9464  0.9464  0.9464 0.9999 0.9999
#> indirect           0.0452 0.0321 20 -0.0258 -0.0247 -0.0200 0.0874 0.0882
#> direct             0.0657 0.0892 20 -0.0747 -0.0704 -0.0513 0.2414 0.2883
#> total              0.1110 0.0969 20 -0.0202 -0.0192 -0.0148 0.2959 0.3287
#>                    99.95%
#> cp                 0.2989
#> b                  0.5092
#> a                  0.2316
#> cond~~cond         1.0000
#> reaction~~reaction 0.9485
#> pmi~~pmi           0.9999
#> indirect           0.0884
#> direct             0.2989
#> total              0.3361