Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.3474 0.2819 20 -0.3041 -0.2742 -0.1416 0.8928 0.9164
#> b                  0.5721 0.1037 20  0.3561  0.3606  0.3804 0.7138 0.7325
#> a                  0.4197 0.2227 20 -0.1509 -0.1238 -0.0031 0.7685 0.7710
#> cond~~cond         0.2488 0.0302 20  0.1992  0.1992  0.1992 0.3076 0.3112
#> reaction~~reaction 1.8574 0.2407 20  1.5026  1.5027  1.5036 2.2898 2.3753
#> pmi~~pmi           1.7844 0.2537 20  1.2431  1.2574  1.3208 2.1978 2.3112
#> reaction~1         0.1774 0.5013 20 -0.7742 -0.7625 -0.7106 1.0006 1.1232
#> pmi~1              5.3594 0.1619 20  5.0273  5.0401  5.0973 5.6667 5.7363
#> cond~1             0.4688 0.0563 20  0.3672  0.3704  0.3851 0.5864 0.5868
#> indirect           0.2401 0.1213 20 -0.0825 -0.0667  0.0038 0.4316 0.4550
#> direct             0.3474 0.2819 20 -0.3041 -0.2742 -0.1416 0.8928 0.9164
#> total              0.5875 0.2995 20 -0.0745 -0.0527  0.0444 1.1014 1.1577
#>                    99.95%
#> cp                 0.9217
#> b                  0.7367
#> a                  0.7715
#> cond~~cond         0.3120
#> reaction~~reaction 2.3946
#> pmi~~pmi           2.3367
#> reaction~1         1.1508
#> pmi~1              5.7520
#> cond~1             0.5869
#> indirect           0.4603
#> direct             0.9217
#> total              1.1704
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1090 0.0877 20 -0.0919 -0.0830 -0.0432 0.2722 0.2808
#> b                  0.4866 0.0794 20  0.3404  0.3404  0.3406 0.5627 0.5643
#> a                  0.1548 0.0810 20 -0.0516 -0.0420  0.0007 0.2794 0.2823
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7349 0.0667 20  0.6232  0.6240  0.6273 0.8440 0.8444
#> pmi~~pmi           0.9760 0.0238 20  0.9200  0.9203  0.9219 0.9969 0.9972
#> indirect           0.1116 0.0392 20 -0.0269 -0.0219  0.0003 0.1357 0.1522
#> direct             3.9638 0.0877 20 -0.0919 -0.0830 -0.0432 0.2722 0.2808
#> total              0.9399 0.0931 20 -0.0225 -0.0160  0.0128 0.3353 0.3507
#>                    99.95%
#> cp                 0.2827
#> b                  0.5647
#> a                  0.2829
#> cond~~cond         1.0000
#> reaction~~reaction 0.8445
#> pmi~~pmi           0.9972
#> indirect           0.1559
#> direct             0.2827
#> total              0.3542

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.2895 0.3177 20 -0.2969 -0.2744 -0.1744 0.7928 0.8359
#> b                  0.5451 0.0889 20  0.4280  0.4283  0.4293 0.7255 0.7338
#> a                  0.3555 0.2599 20 -0.1766 -0.1534 -0.0503 0.8332 0.8760
#> cond~~cond         0.2485 0.0272 20  0.1792  0.1810  0.1890 0.2792 0.2818
#> reaction~~reaction 1.9214 0.2598 20  1.2522  1.2708  1.3539 2.2551 2.2789
#> pmi~~pmi           1.7336 0.1964 20  1.1984  1.2177  1.3038 2.0181 2.0651
#> indirect           0.1996 0.1607 20 -0.1154 -0.1020 -0.0426 0.4894 0.4929
#> direct             0.2895 0.3177 20 -0.2969 -0.2744 -0.1744 0.7928 0.8359
#> total              0.4891 0.3301 20 -0.0539 -0.0451 -0.0061 0.9714 1.0332
#>                    99.95%
#> cp                 0.8456
#> b                  0.7356
#> a                  0.8857
#> cond~~cond         0.2824
#> reaction~~reaction 2.2842
#> pmi~~pmi           2.0756
#> indirect           0.4937
#> direct             0.8456
#> total              1.0471
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1294 0.0948 20 -0.0804 -0.0744 -0.0480 0.2540 0.2715
#> b                  0.4643 0.0769 20  0.3117  0.3152  0.3308 0.6076 0.6196
#> a                  0.1493 0.0960 20 -0.0650 -0.0565 -0.0189 0.3121 0.3312
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7498 0.0683 20  0.5988  0.5999  0.6048 0.8274 0.8282
#> pmi~~pmi           0.9777 0.0291 20  0.8875  0.8901  0.9019 0.9988 0.9989
#> indirect           0.0693 0.0491 20 -0.0338 -0.0298 -0.0119 0.1520 0.1521
#> direct             0.1294 0.0948 20 -0.0804 -0.0744 -0.0480 0.2540 0.2715
#> total              0.1987 0.0999 20 -0.0158 -0.0133 -0.0023 0.2927 0.2955
#>                    99.95%
#> cp                 0.2755
#> b                  0.6223
#> a                  0.3354
#> cond~~cond         1.0000
#> reaction~~reaction 0.8284
#> pmi~~pmi           0.9989
#> indirect           0.1521
#> direct             0.2755
#> total              0.2961