Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.3519 0.2075 20 -0.0602 -0.0551 -0.0324 0.6491 0.7271
#> b                  0.5702 0.0762 20  0.4542  0.4544  0.4552 0.6791 0.6839
#> a                  0.5476 0.1908 20  0.1985  0.2001  0.2072 0.7808 0.8065
#> cond~~cond         0.2476 0.0305 20  0.1799  0.1799  0.1799 0.2848 0.2963
#> reaction~~reaction 1.7349 0.2927 20  1.2394  1.2400  1.2429 2.1834 2.1939
#> pmi~~pmi           1.7008 0.2114 20  1.3260  1.3273  1.3331 1.9848 2.0215
#> reaction~1         0.1421 0.4458 20 -0.6275 -0.6172 -0.5716 0.9582 1.0546
#> pmi~1              5.3476 0.1526 20  5.1266  5.1275  5.1316 5.6186 5.6850
#> cond~1             0.4642 0.0425 20  0.3823  0.3829  0.3854 0.5210 0.5267
#> indirect           0.3122 0.1081 20  0.0999  0.1007  0.1040 0.4268 0.4526
#> direct             0.3519 0.2075 20 -0.0602 -0.0551 -0.0324 0.6491 0.7271
#> total              0.6642 0.2365 20  0.1687  0.1737  0.1959 1.0057 1.0972
#>                    99.95%
#> cp                 0.7447
#> b                  0.6850
#> a                  0.8123
#> cond~~cond         0.2989
#> reaction~~reaction 2.1963
#> pmi~~pmi           2.0297
#> reaction~1         1.0762
#> pmi~1              5.7000
#> cond~1             0.5280
#> indirect           0.4584
#> direct             0.7447
#> total              1.1178
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1131 0.0612 20 -0.0224 -0.0205 -0.0120 0.1880 0.1956
#> b                  0.4907 0.0680 20  0.3490  0.3516  0.3633 0.5898 0.5997
#> a                  0.2045 0.0678 20  0.0760  0.0769  0.0808 0.2694 0.2759
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7238 0.0612 20  0.6423  0.6424  0.6428 0.8447 0.8685
#> pmi~~pmi           0.9582 0.0241 20  0.9231  0.9239  0.9274 0.9934 0.9941
#> indirect           0.0918 0.0349 20  0.0301  0.0305  0.0325 0.1456 0.1481
#> direct             4.0138 0.0612 20 -0.0224 -0.0205 -0.0120 0.1880 0.1956
#> total              0.9329 0.0650 20  0.0538  0.0552  0.0612 0.2794 0.2930
#>                    99.95%
#> cp                 0.1974
#> b                  0.6019
#> a                  0.2774
#> cond~~cond         1.0000
#> reaction~~reaction 0.8738
#> pmi~~pmi           0.9942
#> indirect           0.1486
#> direct             0.1974
#> total              0.2961

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.3293 0.1921 20 -0.1324 -0.1266 -0.1007 0.5203 0.5238
#> b                  0.5596 0.0757 20  0.4380  0.4423  0.4614 0.6972 0.6981
#> a                  0.5430 0.3176 20  0.0342  0.0506  0.1237 1.1769 1.2666
#> cond~~cond         0.2474 0.0317 20  0.1987  0.2005  0.2085 0.3091 0.3123
#> reaction~~reaction 1.7312 0.2086 20  1.3442  1.3513  1.3827 2.0946 2.2119
#> pmi~~pmi           1.7388 0.2173 20  1.3214  1.3408  1.4267 2.1352 2.1388
#> indirect           0.3042 0.1759 20  0.0207  0.0308  0.0758 0.6616 0.6676
#> direct             0.3293 0.1921 20 -0.1324 -0.1266 -0.1007 0.5203 0.5238
#> total              0.6335 0.2598 20  0.0989  0.1112  0.1659 1.0901 1.1657
#>                    99.95%
#> cp                 0.5246
#> b                  0.6983
#> a                  1.2868
#> cond~~cond         0.3131
#> reaction~~reaction 2.2383
#> pmi~~pmi           2.1396
#> indirect           0.6689
#> direct             0.5246
#> total              1.1826
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.1082 0.0605 20 -0.0458 -0.0438 -0.0350 0.1607 0.1632
#> b                  0.4864 0.0588 20  0.3938  0.3972  0.4124 0.6147 0.6210
#> a                  0.1824 0.1064 20  0.0126  0.0183  0.0439 0.3794 0.3804
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7325 0.0554 20  0.6049  0.6060  0.6112 0.7969 0.8041
#> pmi~~pmi           0.9667 0.0456 20  0.8551  0.8553  0.8561 0.9970 0.9993
#> indirect           0.0887 0.0537 20  0.0068  0.0101  0.0245 0.1857 0.1895
#> direct             0.1082 0.0605 20 -0.0458 -0.0438 -0.0350 0.1607 0.1632
#> total              0.1969 0.0742 20  0.0324  0.0364  0.0541 0.3174 0.3332
#>                    99.95%
#> cp                 0.1637
#> b                  0.6224
#> a                  0.3807
#> cond~~cond         1.0000
#> reaction~~reaction 0.8058
#> pmi~~pmi           0.9998
#> indirect           0.1903
#> direct             0.1637
#> total              0.3367