Print Method for Object of Class semmcci

Print Method for Object of Class semmcci

Usage

# S3 method for semmcci
print(x, alpha = NULL, digits = 4, ...)

Arguments

x

an object of class semmcci.

alpha

Numeric vector. Significance level \(\alpha\). If alpha = NULL, use the argument alpha used in x.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Monte Carlo --------------------------------------------------------------
## Fit Model in lavaan -----------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  cond ~~ cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"
fit <- sem(data = df, model = model, missing = "fiml")

## MC() --------------------------------------------------------------------
unstd <- MC(
  fit,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0853 0.3519 20 -0.5044 -0.4884 -0.4177 0.6967 0.7056
#> b                  0.5021 0.0935 20  0.3307  0.3371  0.3655 0.6789 0.7054
#> a                  0.4950 0.2997 20 -0.0943 -0.0664  0.0577 1.1592 1.2044
#> cond~~cond         0.2496 0.0360 20  0.1462  0.1496  0.1647 0.2872 0.2878
#> reaction~~reaction 1.9195 0.2499 20  1.5445  1.5459  1.5521 2.3848 2.4927
#> pmi~~pmi           1.7321 0.1971 20  1.3731  1.3870  1.4486 2.0472 2.0480
#> reaction~1         0.6564 0.5188 20 -0.3500 -0.3401 -0.2959 1.2758 1.2773
#> pmi~1              5.3609 0.1737 20  5.0394  5.0404  5.0450 5.6016 5.6175
#> cond~1             0.4903 0.0475 20  0.3932  0.3968  0.4127 0.5917 0.6115
#> indirect           0.2485 0.1620 20 -0.0407 -0.0255  0.0420 0.6190 0.7306
#> direct             0.0853 0.3519 20 -0.5044 -0.4884 -0.4177 0.6967 0.7056
#> total              0.3338 0.3295 20 -0.2860 -0.2684 -0.1901 0.8855 0.9114
#>                    99.95%
#> cp                 0.7077
#> b                  0.7113
#> a                  1.2146
#> cond~~cond         0.2879
#> reaction~~reaction 2.5169
#> pmi~~pmi           2.0482
#> reaction~1         1.2777
#> pmi~1              5.6211
#> cond~1             0.6159
#> indirect           0.7557
#> direct             0.7077
#> total              0.9172
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0276 0.1121 20 -0.1882 -0.1806 -0.1469 0.2145 0.2230
#> b                  0.4355 0.0661 20  0.2952  0.3014  0.3292 0.5464 0.5487
#> a                  0.1847 0.1076 20 -0.0355 -0.0252  0.0206 0.4050 0.4352
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.8052 0.0496 20  0.6999  0.7005  0.7034 0.8637 0.8685
#> pmi~~pmi           0.9659 0.0480 20  0.8046  0.8101  0.8344 0.9960 0.9981
#> indirect           0.4251 0.0503 20 -0.0134 -0.0088  0.0120 0.1914 0.2307
#> direct             4.0032 0.1121 20 -0.1882 -0.1806 -0.1469 0.2145 0.2230
#> total              0.9814 0.1043 20 -0.1068 -0.0992 -0.0658 0.2657 0.2818
#>                    99.95%
#> cp                 0.2249
#> b                  0.5492
#> a                  0.4419
#> cond~~cond         1.0000
#> reaction~~reaction 0.8696
#> pmi~~pmi           0.9986
#> indirect           0.2396
#> direct             0.2249
#> total              0.2854

# Monte Carlo (Multiple Imputation) ----------------------------------------
## Multiple Imputation -----------------------------------------------------
mi <- mice::mice(
  data = df,
  print = FALSE,
  m = 5L, # use a large value e.g., 100L for actual research,
  seed = 42
)

## Fit Model in lavaan -----------------------------------------------------
fit <- sem(data = df, model = model) # use default listwise deletion

## MCMI() ------------------------------------------------------------------
unstd <- MCMI(
  fit,
  mi = mi,
  R = 20L # use a large value e.g., 20000L for actual research
)

## Standardized Monte Carlo ------------------------------------------------
std <- MCStd(unstd)
print(unstd)
#> Monte Carlo Confidence Intervals (Multiple Imputation Estimates)
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0421 0.2543 20 -0.3490 -0.3453 -0.3288 0.5234 0.5909
#> b                  0.5208 0.0819 20  0.3663  0.3710  0.3920 0.6855 0.7075
#> a                  0.5319 0.2178 20  0.2572  0.2613  0.2794 1.0695 1.1247
#> cond~~cond         0.2499 0.0349 20  0.1837  0.1852  0.1920 0.3084 0.3134
#> reaction~~reaction 1.9405 0.2391 20  1.4987  1.5082  1.5504 2.2863 2.3039
#> pmi~~pmi           1.6607 0.1664 20  1.3296  1.3302  1.3325 1.8829 1.8829
#> indirect           0.2777 0.1423 20  0.1465  0.1480  0.1546 0.6407 0.6831
#> direct             0.0421 0.2543 20 -0.3490 -0.3453 -0.3288 0.5234 0.5909
#> total              0.3198 0.2931 20  0.0132  0.0140  0.0178 0.9962 1.1511
#>                    99.95%
#> cp                 0.6061
#> b                  0.7124
#> a                  1.1372
#> cond~~cond         0.3146
#> reaction~~reaction 2.3078
#> pmi~~pmi           1.8829
#> indirect           0.6927
#> direct             0.6061
#> total              1.1860
print(std)
#> Standardized Monte Carlo Confidence Intervals
#>                       est     se  R   0.05%    0.5%    2.5%  97.5%  99.5%
#> cp                 0.0299 0.0798 20 -0.1051 -0.1040 -0.0991 0.1619 0.1782
#> b                  0.4446 0.0705 20  0.2936  0.2998  0.3273 0.5789 0.6044
#> a                  0.1819 0.0796 20  0.1015  0.1034  0.1119 0.3877 0.4011
#> cond~~cond         1.0000 0.0000 20  1.0000  1.0000  1.0000 1.0000 1.0000
#> reaction~~reaction 0.7966 0.0678 20  0.6365  0.6403  0.6568 0.8823 0.9000
#> pmi~~pmi           0.9669 0.0402 20  0.8366  0.8390  0.8494 0.9874 0.9893
#> indirect           0.0809 0.0440 20  0.0487  0.0490  0.0502 0.1990 0.2157
#> direct             0.0299 0.0798 20 -0.1051 -0.1040 -0.0991 0.1619 0.1782
#> total              0.1108 0.0920 20  0.0040  0.0043  0.0055 0.3004 0.3455
#>                    99.95%
#> cp                 0.1818
#> b                  0.6102
#> a                  0.4042
#> cond~~cond         1.0000
#> reaction~~reaction 0.9040
#> pmi~~pmi           0.9897
#> indirect           0.2194
#> direct             0.1818
#> total              0.3556